### 2023-09-26 #138 @raphaelquast
EOmaps
Submitting Author: Raphael Quast (@raphaelquast)
All current maintainers: @raphaelquast
Package Name: EOmaps
One-Line Description of Package: EOmaps is a python package to visualize, analyze and compare geographical datasets.
Repository Link:  https://github.com/raphaelquast/EOmaps
Version submitted:   7.3
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10822829.svg)](https://doi.org/10.5281/zenodo.10822829)
Version accepted: 8.0.2

---

EOmaps is a python package to visualize, analyze and compare geographical datasets.

It is intended to simplify the process of geospatial data visualization and to provide a straight forward way to turn the maps into interactive widgets for data analysis.

EOmaps is based on `matplotlib` and `cartopy` and extends cartopy's capabilities with the following features

- Multi-layer capabilities (compare/combine/overlay multiple layers)
- North-arrows, scalebars, gridlines on arbitrary projections, ...
- A PyQt5 GUI widget that can be used to quickly fetch webmaps, switch layers, check data-values etc.
- Capabilities to visualize datasets as ellipses, rectangles, geodesic circles, voronoi diagrams, contour plots ...
- Integration with datashader to visualize extremely large datasets
- A flexible API to quickly turn the maps into fully customizable interactive data-analysis widgets
- ...and many more features...

It is extensively [documented](https://eomaps.readthedocs.io/en/latest/), unit-tested, [citable](https://zenodo.org/record/8354476) via a zenodo and installable via [conda](https://anaconda.org/conda-forge/eomaps) or [pip](https://pypi.org/project/EOmaps/).
(However using pip is discouraged since dependencies like GDAL and PYPROJ can be difficult to install, especially on windows)

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

    The target audience for EOmaps are scientists and researchers working with geospatial datasets.
    EOmaps can be used to quilkly visualize structured (or unstructured) geo-spatial datasets, compare multiple datasets with each other or compare maps to an extensive list of open access webmap services. Since EOmaps is based on matplotlib, the created maps can also be connected to ordinary matplotlib axes to analyze multi-dimensional (e.g. timeseries) data.

    In addition to the interactive capabilities, maps created with EOmaps can be exported as high-resolution images or vector-graphics to create publication-ready plots.

    Finally, I believe that EOmaps also has great potential to be used in education to teach about projections, distortions, spatial resolution, rasterization of data, the subtleties of big-data visualization etc.

    EOmaps is based on cartopy. While cartopy provides similar functionalities in terms of data
    visualization, EOmaps greatly extends these capabilities (especially for large datasets),
    adds multi-layer support, a basic GUI, easy-access to webmaps, and many more features.

    There exist other packages that focus on interactive geo-data visualization, but to my knowledge
    none that focusses on local use in pure python.

     - folium (wrapper for javascript-library leaflet)
     - geemap (wrapper for google-earth-engine)
     - geoviews (provides an api for matplotlib, but it is a much more high-level plotting library than EOmaps)

    #137
