+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,23 @@
+---
+name: Bug report
+about: Report a problem or an incorrect behavior.
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**General information:**
+Please paste the output of the following code:
+
+    import sys, eomaps
+    print(f"OS = {sys.platform}\npython = {sys.version}\nEOmaps = {eomaps.__version__}")
