### 2023-09-27 5733568
fix docsctrings for get/set_extent
