### 2023-09-27 39f733b
check cartopy and initial crs before reprojecting geo-dataframes
