### 2023-09-28 ea242b8
avoid re-evaluating scalebar scale if the extent did not change
