### 2023-09-29 9e273ef
avoid activating matplotlibs interactive mode by default (fixes #184)
