### 2023-10-02 863d743
add zoom capabilities to speed up large raster data visualization
