### 2023-10-02 ce3ead8
add fast data aggregation-capabilities for raster-data
