### 2023-10-03 1ee5553
updates for raster-data aggregation methods
