### 2023-10-03 d3285b6
add basic unittests for raster aggregation
