### 2023-10-03 0a88af6
add basic unittests for raster aggregation
