### 2023-10-03 616a0e6
update raster aggregation unittests
