### 2023-10-03 b1f65dc
use more stable estimation for blocksize
