+++ eomaps/_data_manager.py
@@ -573,7 +573,7 @@ class DataManager:
         if self.full_data_in_extent((x0, x1, y0, y1)):
             self.last_extent = self.current_extent
             self._current_data = {**self._all_data}
-            return None, None, None
+            return True, True, True
 
         # get mask
         if self.x0_1D is not None:
@@ -597,16 +597,11 @@ class DataManager:
 
         return q, qx, qy
 
-    def _select_vals(self, val):
-        # select data based on currently visible region
-        q, qx, qy = self._last_qs
-        if all(i is None for i in (q, qx, qy)):
-            ret = val
-
-        val = np.asanyarray(val)
-
-        if len(val.shape) == 2 and qx is not None and qy is not None:
-            # TODO this requires sorted coordinate vectors to work!
+    def _estimate_slice_blocksize(self, qx, qy):
+        if qx is True and qy is True:
+            x0, x1 = 0, len(self.x0)
+            y0, y1 = 0, len(self.y0)
+        else:
             qx = qx.squeeze()
             qy = qy.squeeze()
 
@@ -622,52 +617,87 @@ class DataManager:
             if y1 == -1:
                 y1 = len(qy)
 
-            ret = val[y0:y1, x0:x1]
-            # ret = val[qy.squeeze()][:, qx.squeeze()]
-        else:
-            if q is not None:
-                ret = val.ravel()[q.squeeze()]
-            else:
-                ret = val
+        maxsize = getattr(self.m.shape, "_maxsize", None)
+        if maxsize is not None:
+            # estimate a suitable blocksize based on the max. data size
+            # in x- or y- direction
+            d = max((x1 - x0), (y1 - y0))
+            bs = int(d / np.sqrt(maxsize))
 
-        if self.m.shape.name not in ["raster", "shade_raster", "contour"]:
-            ret = ret.ravel()
+            if bs > 0:  # to avoid divided by zero errors in %
+                x0 = x0 - x0 % bs
+                y0 = y0 - y0 % bs
+                if x0 < 0:
+                    x1 = x1 - x0
+                    x0 = 0
+                if y0 < 0:
+                    y1 = y1 - y0
+                    y0 = 0
+
+                x1 = x1 - x1 % bs + bs
+                y1 = y1 - y1 % bs + bs
+        else:
+            bs = None
+
+        return (x0, x1, y0, y1), bs
+
+    def _select_vals(self, val, qs, slices=None):
+        # select data based on currently visible region
+        q, qx, qy = qs
+        if all(i is True for i in (q, qx, qy)):
+            ret = val
+        elif all(i is None for i in (q, qx, qy)):
+            ret = None
+        else:
+            val = np.asanyarray(val)
+
+            if len(val.shape) == 2 and qx is not None and qy is not None:
+                (x0, x1, y0, y1) = slices
+                ret = val[y0:y1, x0:x1]
+            else:
+                if q is not None:
+                    ret = val.ravel()[q.squeeze()]
+                else:
+                    ret = val
+
+            if self.m.shape.name not in ["raster", "shade_raster", "contour"]:
+                ret = ret.ravel()
 
         return ret
 
-    def _select_ids(self):
-        # identify ids based on currently visible region
-        q, qx, qy = self._last_qs
+    # def _select_ids(self, qs):
+    #     # identify ids based on currently visible region
+    #     q, qx, qy = qs
 
-        if all(i is None for i in (q, qx, qy)):
-            return self.ids
+    #     if all(i is None for i in (q, qx, qy)):
+    #         return self.ids
 
-        if q is None or len(q.shape) == 2:
-            # select columns that contain at least one value
[…]
