### 2023-10-03 9378634
increase default aggregation limit to 1e7
