### 2023-10-03 029ad7d
update docstrings
