### 2023-10-03 ee80798
use "raster" as default shape for all 2d datasets
