### 2023-10-04 f16b7f9
update docs for shape selection
