+++ docs/api_data_visualization.rst
@@ -21,9 +21,6 @@ then select how you want to visualize the data and finally call :py:meth:`Maps.p
 3. (optional) :ref:`Classify the data <classify_the_data>` via  :py:class:`Maps.set_classify`
 4. :ref:`Plot the data <plot_the_data>` by calling :py:meth:`Maps.plot_map`
 
-A :py:class:`Maps` object can only manage a single dataset.
-To plot multiple datasets on the same map, use :py:meth:`Maps.new_layer` to get a unique :py:class:`Maps` object for each dataset!
-
 .. code-block:: python
     :name: test_data_visualization_quick
 
@@ -43,6 +40,16 @@ To plot multiple datasets on the same map, use :py:meth:`Maps.new_layer` to get 
     m_data.set_classify.EqualInterval(k=3)         # Classify into 3 equal intervals
     m_data.plot_map(vmin=0, ec="k", cmap="magma")  # Plot the data
     m_data.add_colorbar(hist_bins="bins")          # Add a colorbar
+
+
+.. note::
+
+    A :py:class:`Maps` object can only manage a single dataset!
+
+    To plot multiple datasets on the same map, use :py:meth:`Maps.new_layer` to get a unique :py:class:`Maps` object for each dataset!
+    Use ``m_2 = m.new_layer(inherit_data=True, inherit_classification=True, inherit_shape=True)`` to quickly create a new layer
+    that uses the same dataset, cassification and shape as the parent!
+
 
 
 .. _assign_the_data:
@@ -138,13 +145,14 @@ The following data-types are accepted as input:
 |                                                                     |     m.plot_map()                                                                   |
 +---------------------------------------------------------------------+------------------------------------------------------------------------------------+
 
+
 .. _set_the_shape:
 
-
-2) Set the shape
-~~~~~~~~~~~~~~~~
+2) Plot shapes
+~~~~~~~~~~~~~~
 
 To specify how a dataset is visualized on the map, you have to set the *"plot-shape"* via :py:meth:`Maps.set_shape`.
+
 
 .. currentmodule:: eomaps.eomaps
 
@@ -153,88 +161,49 @@ To specify how a dataset is visualized on the map, you have to set the *"plot-sh
 
     Maps.set_shape
 
-.. admonition:: A note on speed and performance
-
-    Some ways to visualize the data require more computational effort than others!
-    Make sure to select an appropriate shape based on the size of the dataset you want to plot!
-
-    EOmaps dynamically pre-selects the data with respect to the current plot-extent before the actual plot is created!
-    If you do not need to see the whole extent of the data, make sure to **set the desired plot-extent**
-    via :py:meth:`Maps.set_extent` or :py:meth:`Maps.set_shape_to_extent` **BEFORE** calling :py:meth:`Maps.plot_map` to get a (possibly huge) speedup!
-
-    The number of datapoints mentioned in the following always refer to the number of datapoints that are
-    visible in the desired plot-extent.
-
-    For very large datasets, make sure to have a look at the :py:class:`raster`, :py:class:`shade_raster`, and :py:class:`shade_points` shapes!
-
-Possible shapes that work nicely for datasets with up to ~500 000 data-points:
+Available shapes (see bleow for details on each plot-shape!):
 
 .. currentmodule:: eomaps.shapes.Shapes
 
 .. autosummary::
     :nosignatures:
 
-    geod_circles
     ellipses
     rectangles
+    geod_circles
     voronoi_diagram
     delaunay_triangulation
-
-
-Possible shapes that work nicely for up to a few million data-points:
-
-.. autosummary::
-    :nosignatures:
-
     contour
-
-
-For **extremely large datasets** (> 5 million datapoints) there is little use in plotting all datapoints directly
[…]
