### 2023-10-05 00b5579
add "mode" as possible aggregation method
