### 2023-10-05 c22fc56
fix issues with misplaced grid-labels on figure export fix dataset for some shapes not properly exported with "agg" backend
