### 2023-10-06 #138 @raphaelquast
Hey @arianesasso. Just checking in to see if there are any updates on the EOmaps submission yet?

In the meantime I've pushed to [EOmaps v7.2](https://github.com/raphaelquast/EOmaps/releases/tag/v7.2)  to address some reported issues and since the formal review process did not yet start, I've also updated the "submitted version" above to v7.2 (please let me know in case this is not OK). Thanks!
