### 2023-10-06 87a6560
catch exceptions on default keypress callbacks to avoid crashing (e.g. when using backends that cannot initialize the widget)
