### 2023-10-06 4459b0a
make sure annotations use a black fontcolor by default (to avoid issues until proper treatment of rcparams is introduced)
