### 2023-10-06 344bebd
remove duplicate restore_region call in _get_background
