### 2023-10-06 4e6ccf5
fix issues due to plotted datasets outside the current extent
