### 2023-10-07 934c666
avoid triggering companion-widget keypress callbacks in webagg (otherwise pyqt issues might cause a crash)
