### 2023-10-08 abf9732
add update triggers for default move-callbacks
