### 2023-10-09 fdefd26
allow overriding override callback behavior during toolbar actions
