+++ eomaps/cb_container.py
@@ -131,6 +131,7 @@ class _CallbackContainer(object):
         self._event = None
 
         self._execute_on_all_layers = False
+        self._execute_while_toolbar_active = False
 
     def _getobj(self, m):
         """Get the equivalent callback container on another maps object."""
@@ -327,6 +328,35 @@ class _CallbackContainer(object):
                 f"EOmaps: 'execute_on_all_layers' is inherited for {self._method}!"
             )
         self._execute_on_all_layers = q
+
+    def _check_toolbar_mode(self):
+        if self._execute_while_toolbar_active:
+            return False
+
+        # returns True if a toolbar mode is active and False otherwise
+        if (
+            self._m.f.canvas.toolbar is not None
+        ) and self._m.f.canvas.toolbar.mode != "":
+            return True
+        else:
+            return False
+
+    def set_execute_during_toolbar_action(self, q):
+        """
+        Set if callbacks should be executed during a toolbar action (e.g. pan/zoom).
+
+        By default, callbacks are not executed during toolbar actions to make sure
+        pan/zoom is smooth. (e.g. to avoid things like constant re-fetching of webmaps
+        if a peek-layer callback is active during pan/zoom)
+
+        Parameters
+        ----------
+        q : bool
+            If True, callbacks will be triggered independent of the toolbar state.
+            if False, callbacks will only trigger if no toolbar action is active.
+
+        """
+        self._execute_while_toolbar_active = q
 
 
 class _ClickContainer(_CallbackContainer):
@@ -993,9 +1023,7 @@ class ClickContainer(_ClickContainer):
                 self._event = event
 
                 # don't execute callbacks if a toolbar-action is active
-                if (
-                    self._m.f.canvas.toolbar is not None
-                ) and self._m.f.canvas.toolbar.mode != "":
+                if self._check_toolbar_mode():
                     return
 
                 # execute onclick on the maps object that belongs to the clicked axis
@@ -1023,9 +1051,7 @@ class ClickContainer(_ClickContainer):
                 self._event = event
 
                 # don't execute callbacks if a toolbar-action is active
-                if (
-                    self._m.f.canvas.toolbar is not None
-                ) and self._m.f.canvas.toolbar.mode != "":
+                if self._check_toolbar_mode():
                     return
 
                 # execute onclick on the maps object that belongs to the clicked axis
@@ -1171,9 +1197,7 @@ class MoveContainer(ClickContainer):
                         return
 
                 # don't execute callbacks if a toolbar-action is active
-                if (
-                    self._m.f.canvas.toolbar is not None
-                ) and self._m.f.canvas.toolbar.mode != "":
+                if self._check_toolbar_mode():
                     return
 
                 # execute onclick on the maps object that belongs to the clicked axis
@@ -1473,9 +1497,7 @@ class PickContainer(_ClickContainer):
             return
 
         # don't execute callbacks if a toolbar-action is active
-        if (
-            self._m.f.canvas.toolbar is not None
-        ) and self._m.f.canvas.toolbar.mode != "":
+        if self._check_toolbar_mode():
             return
 
         # make sure temporary artists are cleared before executing new callbacks
@@ -1553,9 +1575,7 @@ class PickContainer(_ClickContainer):
                     return
 
                 # don't execute callbacks if a toolbar-action is active
-                if (
-                    self._m.f.canvas.toolbar is not None
-                ) and self._m.f.canvas.toolbar.mode != "":
+                if self._check_toolbar_mode():
                     return
 
                 if not self._artist_picked(event):
