### 2023-10-09 ba42feb
add option to add gridlines as dynamic artists
