### 2023-10-09 fb4ba84
add proper remove method for grid-labels
