### 2023-10-11 #138 @NickleDave
Hi @raphaelquast thank you for your continued patience.
We are still looking for an editor with spatial expertise for this review.

In the meantime, here are the initial editor checks. As expected, EOmaps passes with flying colors.

  - More difficult to install with pip because of native dependencies (e.g. GDAL) but conda works fine
    - yes, at https://eomaps.readthedocs.io/en/latest/index.html
    - yes, at https://eomaps.readthedocs.io/en/latest/EOmaps_examples.html

---
---

None except nice work :slig
