### 2023-10-12 #138 @raphaelquast
@NickleDave
Thanks! Great to hear that initial checks turned out fine!

Two quick points from my side:
- There is actually a code-of-conduct implemented: [CODE_OF_CONDUCT.md](https://github.com/raphaelquast/EOmaps/blob/master/CODE_OF_CONDUCT.md)
- Since there's no review activity yet, I'd like to push the submitted version up to [v7.2.1](https://github.com/raphaelquast/EOmaps/releases/tag/7.2.1) to incorporate latest fixes.
