### 2023-10-13 e2873c5
add "permanent" kwarg to m.add_gdf
