### 2023-10-13 c4b82c2
don't attempt to estimate vmin/vmax if no data-values are provided
