### 2023-10-13 a4b0511
fix axis background patch color is only visible on export
