### 2023-10-13 7ac588a
fix zooming datasets that use an individual size for each datapoint - affected shapes: ellipses, rectangles, geod_circles, scatter_points
