### 2023-10-13 97e61d5
allow overriding radius-selection (required for add_marker)
