### 2023-10-13 38e8984
parse coordinates as numpy-arrays (to properly support pandas.Series)
