### 2023-10-13 056cf84
avoid prasing geod-circle radius as 1D array
