### 2023-10-13 1348182
fix identification of layers that only exist on inset-maps
