### 2023-10-14 1e755af
fix annotation editor issues if annotation is not attached to the figure
