### 2023-10-14 a371867
fix selected annotation text is cleared on companion-widget initialization
