### 2023-10-14 f540109
annotation editor updates - clear all temporary artists on annotation-editor init - disable callbacks while annotation-editor is active - add _always_active click callback container
