### 2023-10-14 2554eaa
add "update" kwarg to add_annotation/marker to override update triggers
