### 2023-10-14 3facbcb
add "use_ids" option to override using IDs or xy for annotation-code export
