### 2023-10-15 9e26a5d
fix figure/axes patch zorder on __BG__ layer
