### 2023-10-15 cafeccb
fix companion widgets for child-maps remain open on figure close
