### 2023-10-15 91980ce
fix `m.get_extent()` should not return coordinates outside the crs bounds
