### 2023-10-15 fdfee82
add new convenience method `m.update_spine_props`
