### 2023-10-15 fad93fa
set logo interpolation_stage to rgba
