### 2023-10-15 fe89ace
call plt.show on m.show_layer for non-interactive backends
