### 2023-10-16 f0833a9
- avoid activating the maps-object's layer on m.show() - fix clearing snapshot cells with m.show()
