### 2023-10-16 bf095c8
inserts fig1 in inset map chapter of docs
