### 2023-10-17 82c8c56
includes figure2 for second code block
