### 2023-10-18 8b0bdf2
fix active toolbar mode must be checked for each Maps object individually (checks in onclick, onmove etc. only target state of parent Maps-object)
