### 2023-10-18 ef68d8c
fix layer-selector util should remain responsive if toolbar mode is active
