### 2023-10-18 ff5549f
add unittests for Maps.config
