### 2023-10-18 da08c4c
add unittest for gdf clipping methods
