### 2023-10-18 082dd0d
properly handle snapshot exceptions (if IPython is not available)
