### 2023-10-18 d0fdbae
avoid testing gdal clipping for now
