### 2023-10-19 #138 @NickleDave
Ah, whoops, sorry I failed to mark the code-of-conduct box, I saw it but missed the checkbox. Fixed now.

> Since there's no review activity yet, I'd like to push the submitted version up to [v7.2.1](https://github.com/raphaelquast/EOmaps/releases/tag/7.2.1) to incorporate latest fixes.

No worries. We will make sure that's clear here. You can feel free to change the metadata in your initial comment above, or I can.

We are getting close to finding an editor. Thanks for your continued patience, and expect us to reply back early next week.
