### 2023-10-20 cd36965
minor doc update
