### 2023-10-20 a396d2e
add info how to build the docs to the contribution guide
