+++ docs/contribute.rst
@@ -49,36 +49,8 @@ You can install it directly into your environment via:
     mamba install -c conda-forge spyder
 
 
-
-Getting started
----------------
-
-The source code of EOmaps is managed on `GitHub <https://github.com/raphaelquast/EOmaps>`_.
-
-To get started, create a new **fork** of the `EOmaps repository <https://github.com/raphaelquast/EOmaps/fork>`_
-to get your own copy of the source code.
-
-Then, open a terminal, navigate to the folder you want to work in and clone the forked repository via:
-
-.. code-block:: console
-
-    git clone < url to fork of EOmaps repository >
-
-For development, make sure that you first checkout the ``dev`` branch which contains all pending changes for the next release.
-Then, create a new feature or bug-fix branch and start coding!
-
-.. code-block:: console
-
-    git checkout dev
-    git checkout -b "awesome_new_feature"
-
-
-Once you're done or in case you want/need some feedback, open a `pull request <https://github.com/raphaelquast/EOmaps/pulls>`_ on GitHub!
-
-.. _pre_commit:
-
 Pre-commit hooks
-----------------
+~~~~~~~~~~~~~~~~
 
 To ensure uniform code formatting, EOmaps uses `pre-commit hooks <https://pre-commit.com/>`_ to automatically check (and fix) code-style issues such as:
 
@@ -86,7 +58,7 @@ To ensure uniform code formatting, EOmaps uses `pre-commit hooks <https://pre-co
 - Compliance to the used `black <https://github.com/psf/black>`_ code formatting standards
 
 To initialize pre-commit hooks in your current environment, navigate to the directory where you cloned the EOmaps repository and run the following command:
-(e.g. the parent directory containing the file `.pre-commit-config.yaml`)
+(e.g. the parent directory containing the file ``.pre-commit-config.yaml``)
 
 
 .. code-block:: console
@@ -105,3 +77,77 @@ This will install the required pre-commit hooks in your current environment so t
 
 To run the pre-commit hooks manually on selected files, simply **add the files you want to commit** with ``git add < filename >`` and then run ``pre-commit``.
 (If you want to run the hooks on all files, use ``pre-commit run --all-files``)
+
+
+
+
+
+Getting started
+---------------
+
+The source code of EOmaps is managed on `GitHub <https://github.com/raphaelquast/EOmaps>`_.
+
+To get started, create a new `fork <https://docs.github.com/en/get-started/quickstart/fork-a-repo>`_ of the `EOmaps repository <https://github.com/raphaelquast/EOmaps/fork>`_
+to get your own copy of the source code.
+
+Then, open a terminal, navigate to the folder you want to work in and clone the forked repository via:
+
+.. code-block:: console
+
+    git clone < url to  of EOmaps repository >
+
+For development, make sure that you first checkout the ``dev`` branch which contains all pending changes for the next release.
+Then, create a new feature or bug-fix branch and start coding!
+
+.. code-block:: console
+
+    git checkout dev
+    git checkout -b "awesome_new_feature"
+
+
+Once you're done or in case you want/need some feedback, open a `pull request <https://github.com/raphaelquast/EOmaps/pulls>`_ on GitHub!
+
+.. _pre_commit:
+
+
+Building the docs
+-----------------
+
+The documentation of EOmaps is written with `Sphinx <https://www.sphinx-doc.org/en/master/>`_ using the markup language `ReStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_.
+
+
+To build the documentation locally, simply navigate to the folder ``eomaps/docs`` and then run
+
+.. code-block::
+
+    make html
+
+
[…]
