+++ docs/contribute.rst
@@ -22,12 +22,41 @@ Any contributions are welcome!
 - ...
 
 
+🛞Getting started
+------------------
+
+The source code of EOmaps is managed on `GitHub <https://github.com/raphaelquast/EOmaps>`_.
+
+To get started, create a new `fork <https://docs.github.com/en/get-started/quickstart/fork-a-repo>`_ of the `EOmaps repository <https://github.com/raphaelquast/EOmaps/fork>`_
+to get your own copy of the source code.
+
+Then, open a terminal, navigate to the folder you want to work in and clone the forked repository via:
+
+.. code-block:: console
+
+    git clone < link to your fork of the EOmaps repository >
+
+For development, make sure that you first checkout the ``dev`` branch which contains all pending changes for the next release.
+Then, create a new feature or bug-fix branch and start coding!
+
+.. code-block:: console
+
+    git checkout dev
+    git checkout -b "awesome_new_feature"
+
+
+Once you're done or in case you want/need some feedback, open a `pull request <https://github.com/raphaelquast/EOmaps/pulls>`_ on GitHub!
+
+.. _pre_commit:
+
+
+
 .. _setup_a_dev_env:
 
 ⚙ How to set up a development environment
------------------------------------------
+------------------------------------------
 
-To get started you need a working ``python`` installation.
+To contribute to ``EOmaps``, you need a working ``python`` installation.
 I recommend using `miniconda <https://docs.conda.io/en/latest/miniconda.html>`_ to set up ``python`` with the following steps:
 
 - Download the latest `miniconda <https://docs.conda.io/en/latest/miniconda.html>`_ installer and install
@@ -75,43 +104,15 @@ This will install the required pre-commit hooks in your current environment so t
   only commits that comply with all pre-commit hooks can be pushed to GitHub.
 
 
-To run the pre-commit hooks manually on selected files, simply **add the files you want to commit** with ``git add < filename >`` and then run ``pre-commit``.
-(If you want to run the hooks on all files, use ``pre-commit run --all-files``)
+- To run the pre-commit hooks manually on selected files, simply **add the files you want to commit** with ``git add < filename >`` and then run ``pre-commit``
+- If you want to run the hooks on all files, use ``pre-commit run --all-files``
+
+Note: after running the pre-commit hooks, some files might have new changes that must be staged for commit again!
 
 
 
-
-
-Getting started
----------------
-
-The source code of EOmaps is managed on `GitHub <https://github.com/raphaelquast/EOmaps>`_.
-
-To get started, create a new `fork <https://docs.github.com/en/get-started/quickstart/fork-a-repo>`_ of the `EOmaps repository <https://github.com/raphaelquast/EOmaps/fork>`_
-to get your own copy of the source code.
-
-Then, open a terminal, navigate to the folder you want to work in and clone the forked repository via:
-
-.. code-block:: console
-
-    git clone < url to  of EOmaps repository >
-
-For development, make sure that you first checkout the ``dev`` branch which contains all pending changes for the next release.
-Then, create a new feature or bug-fix branch and start coding!
-
-.. code-block:: console
-
-    git checkout dev
-    git checkout -b "awesome_new_feature"
-
-
-Once you're done or in case you want/need some feedback, open a `pull request <https://github.com/raphaelquast/EOmaps/pulls>`_ on GitHub!
-
-.. _pre_commit:
-
-
-Building the docs
------------------
+📖 Building the docs
+---------------------
 
 The documentation of EOmaps is written with `Sphinx <https://www.sphinx-doc.org/en/master/>`_ using the markup language `ReStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_.
 
[…]
