+++ eomaps/_version.py
@@ -1,1 +1,1 @@
-__version__ = "7.2.1"
+__version__ = "7.3"
