### 2023-10-20 53dacaa
allow setting background color with `m.set_frame`
