### 2023-10-20 5c1a2aa
fix using mpl gui save button (or 'control+c') to save a eomaps figure
