### 2023-10-20 d751f2e
fix _data_manager must be initialized before figure is created
