### 2023-10-20 50fb346
fix restoring background after savefig
