### 2023-10-22 6fe37c6
make sure backgrounds are properly re-fetched with `m.set_frame`
