### 2023-10-22 a4ff054
fix re-drawing of backgrounds on add/remove grid (and/or grid-labels)
