+++ docs/index.rst
@@ -1,37 +1,168 @@
+.. include:: substitutions.rst
+
 .. image:: _static/logo.png
   :width: 40%
   :align: center
   :target: https://github.com/raphaelquast/EOmaps
 
-|
+.. raw:: html
 
-Welcome to the documentation for **EOmaps**!
-
-EOmaps is a Python package to visualize and analyze geographical datasets.
-It is built on top of ``matplotlib`` and ``cartopy`` and aims to provide an
-intuitive and easy-to-use interface to handle the following tasks:
-
-| ▶ Speed up and simplify the creation and comparison of maps
-| ▶ Visualize small datasets as well as millions of datapoints
-| ▶ Handle 1D and 2D datasets and create plots from NetCDF, GeoTIFF or CSV files
-| ▶ Take care of re-projecting the data
-| ▶ Compare or overlay different plot-layers and WebMap services
-| ▶ Use the maps as interactive data-analysis widgets (e.g. execute functions if you click on the map)
-| ▶ Provide a versatile set of tools to customize the maps
-| ▶ Arrange multiple maps in one figure
-| ▶ Get a nice colorbar with a histogram on top
-| ▶ Export high resolution images
-|
-| A detailed overview on how to use EOmaps is given in the :doc:`api`.
-| Make sure to check out the :doc:`EOmaps_examples` for an overview of the capabilities (incl. source code)!
-
-----------
+    <br>
+    <font size=5>Welcome to the documentation for <b>EOmaps</b>!</font>
+    <p>
 
 
-Contents
+Here you can find detailed explanations on all the functionalities of EOmaps.
+
+.. admonition:: Interested in contributing to EOmaps?
+
+    | Found a bug or got an idea for an interesting feature?
+    | Open an issue on `GitHub <https://github.com/raphaelquast/EOmaps>`_ or head over to the :doc:`contribute` to see how to setup EOmaps for development!
+
+
+
+Getting Started
+---------------
+
+To get started, have a look at the :doc:`api_basics` section to get to know
+the basic concepts of EOmaps.
+
+In addition, EOmaps also provides a GUI widget (the :doc:`api_companion_widget`) that can be used to interactively edit/compare/overlay maps and explore the features and functionalities.
+
+Data Visualization
+------------------
+
+.. toctree::
+    :hidden:
+    :maxdepth: 3
+
+    api_basics
+    api_data_visualization
+
+
+
+Want to visualize some data? Have a look at the :doc:`api_data_visualization` section to learn how to create beautiful maps with your datasets!
+
+EOmaps provides a variety of plot-shapes so you can select a shape that best suit the structure and spatial representativeness of your data:
+
+
+.. raw:: html
+
+    <table>
+    <tr>
+    <th>Ellipses<br/><a href=api_data_visualization.html#ellipses><img src="_static/shape_imgs/ellipses.png"/></a></th>
+    <th>Rectangles<br/><a href=api_data_visualization.html#rectangles><img src="_static/shape_imgs/rectangles.png"/></a></th>
+    <th>Geodesic Circles<br/><a href=api_data_visualization.html#geodesic-circles><img src="_static/shape_imgs/geod_circles.png"/></a></th>
+    <th>Raster<br/><a href=api_data_visualization.html#raster><img src="_static/shape_imgs/raster.png"/></a></th>
[…]
