+++ docs/substitutions.rst
@@ -7,3 +7,127 @@
 
    </details>
    <br/>
+
+.. |shp_ellipses| image:: _static/shape_imgs/ellipses.png
+    :width: 100%
+    :align: top
+    :target: api_data_visualization.html#ellipses
+
+.. |shp_rectangles| image:: _static/shape_imgs/rectangles.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#rectangles
+
+.. |shp_geod_circles| image:: _static/shape_imgs/geod_circles.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#geodesic-circles
+
+.. |shp_voronoi| image:: _static/shape_imgs/voronoi_diagram.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#voronoi-diagram
+
+.. |shp_delaunay| image:: _static/shape_imgs/delaunay_triangulation.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#delaunay-triangulation
+
+.. |shp_contour| image:: _static/shape_imgs/contour.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#contour
+
+.. |shp_scatter_points| image:: _static/shape_imgs/scatter_points.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#scatter-points
+
+.. |shp_raster| image:: _static/shape_imgs/raster.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#raster
+
+.. |shp_shade_raster| image:: _static/shape_imgs/shade_raster.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#shade-raster
+
+.. |shp_shade_points| image:: _static/shape_imgs/shade_points.png
+    :width: 100%
+    :align: middle
+    :target: api_data_visualization.html#shade-points
+
+
+
+.. |eximg01| image:: _static/fig1.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-1
+
+.. |eximg02| image:: _static/fig2.png
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-2
+
+.. |eximg03| image:: _static/fig3.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-3
+
+.. |eximg04| image:: _static/fig4.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-4
+
+.. |eximg05| image:: _static/fig5.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-5
+
+.. |eximg06| image:: _static/fig6.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-6
+
+.. |eximg07| image:: _static/fig7.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-7
+
+.. |eximg08| image:: _static/fig8.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-8
+
+.. |eximg09| image:: _static/fig9.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-9
+
+.. |eximg10| image:: _static/example_row_col_selector.gif
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-10
+
+.. |eximg11| image:: _static/example_inset_maps.png
+   :width: 100%
+   :align: middle
+   :target: EOmaps_examples.html#ex-11
+
[…]
