### 2023-10-28 be996d0
docs-update: fix image resizing in shape table
