### 2023-11-16 #138 @banesullivan
:wave: Hi @yeelauren and @jhkennedy! Thank you for volunteering to review for pyOpenSci!

2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

Please get in touch with any questions or concerns! Your review is due ~6th of December 2023
---

Reviewers: @yeelauren @jhkennedy
Due date:  2023-12-06
