### 2023-11-19 3fb6e35
connect InsetMap indicator line to intersection with inset boundary
