### 2023-11-19 a426f86
fix handling of multiple inset-indicator lines
