### 2023-11-19 1feedc8
fix passing InsetMap boundary colors as RGBA tuples
