### 2023-11-23 eb056a6
fix UserDefined classification issue: max-bin replaced by data-max
