### 2023-11-23 295b64b
fix UserDefined classification issue: max-bin replaced by data-max
