### 2023-11-23 d9fe396
allow using kwargs with m.add.feature.preset(...) for multiple features
