### 2023-11-23 178adf6
Add background image capabilities to the LayoutEditor
