### 2023-11-23 7ca2566
Add background image capabilities to the LayoutEditor
