+++ eomaps/helpers.py
@@ -582,7 +582,8 @@ class LayoutEditor:
                 "0 - 9:  Snap-grid spacing\n"
                 "SHIFT:  Multi-select\n"
                 "P:      Print to console\n"
-                "ESCAPE (or ALT + L): Exit\n"
+                "ESCAPE: Exit\n"
+                "R:      Refetch images\n"
                 "\n"
                 "ARROW-KEYS:   Move\n"
                 "SCROLL (+/-): Resize\n"
@@ -825,7 +826,6 @@ class LayoutEditor:
                 self._info_text_hidden = not vis
 
         eventax = event.inaxes
-
         if eventax not in self.axes:
             # if no axes is clicked "unpick" previously picked axes
             if len(self._ax_picked) + len(self._cb_picked) == 0:
@@ -845,6 +845,9 @@ class LayoutEditor:
             self.fetch_current_background()
             self.blit_artists()
             return
+
+        if eventax.name == "eomaps_ax_image":
+            eventax = eventax.eomaps_parent_ax
 
         if self._shift_pressed:
             if eventax in self.maxes:
@@ -910,6 +913,9 @@ class LayoutEditor:
         with ExitStack() as stack:
             for ax in self._ax_picked:
                 stack.enter_context(ax._cm_set(visible=False))
+                ax_image = getattr(ax, "eomaps_ax_image", None)
+                if ax_image:
+                    stack.enter_context(ax_image._cm_set(visible=False))
 
             for cb in self._cb_picked:
                 stack.enter_context(cb.ax_cb._cm_set(visible=False))
@@ -936,6 +942,8 @@ class LayoutEditor:
         for ax in self._ax_picked:
             bbox = self._get_move_with_key_bbox(ax, event.key)
             ax.set_position(bbox)
+            if hasattr(ax, "eomaps_ax_image"):
+                ax.eomaps_ax_image.set_position(bbox)
 
         for cb in self._cb_picked:
             bbox = self._get_move_with_key_bbox(cb._ax, event.key)
@@ -961,6 +969,8 @@ class LayoutEditor:
 
             bbox = self._get_move_bbox(ax, event.x, event.y)
             ax.set_position(bbox)
+            if hasattr(ax, "eomaps_ax_image"):
+                ax.eomaps_ax_image.set_position(bbox)
 
         for cb in self._cb_picked:
             if cb is None:
@@ -982,6 +992,8 @@ class LayoutEditor:
             return
 
         artists = [*self._ax_picked]
+        artists += [ax for ax in self.f.axes if ax.name == "eomaps_ax_image"]
+
         for cb in self._cb_picked:
             artists.append(cb.ax_cb)
             artists.append(cb.ax_cb_plot)
@@ -1006,6 +1018,8 @@ class LayoutEditor:
                 resize_bbox = self._get_resize_bbox(ax, event.step)
                 if resize_bbox is not None:
                     ax.set_position(resize_bbox)
+                    if hasattr(ax, "eomaps_ax_image"):
+                        ax.eomaps_ax_image.set_position(resize_bbox)
 
         for cb in self._cb_picked:
             if cb is None:
@@ -1025,6 +1039,12 @@ class LayoutEditor:
         # self._color_axes()
         self.blit_artists()
 
+    def print_layout(self):
+        s = "\nlayout = {\n    "
+        s += "\n    ".join(f'"{key}": {val},' for key, val in self.get_layout().items())
+        s += "\n}\n"
+        print(s)
+
     def cb_key_press(self, event):
         # release shift key on every keypress
         self._shift_pressed = False
@@ -1038,12 +1058,7 @@ class LayoutEditor:
             self._undo_draggable()
             return
[…]
