### 2023-11-23 347a85d
fix get_layout extents
