### 2023-11-23 371c47a
fix get_layout extents
