### 2023-11-30 8ccb41b
fix issues with RGBA snapshots for inline plots in jupyter notebooks
