### 2023-11-30 08273a5
hide ipympl header label in jupyter notebooks by default
