### 2023-12-01 d26f641
remember info-text visibility in layout-editor
