### 2023-12-01 57ecf27
improve debug logging on draw
