### 2023-12-01 9a3a06c
minor debug logging update
