### 2023-12-01 57cd8d1
implement better syntax for disabling draw or update in the BlitManager
