### 2023-12-01 cc1a2b2
implement better check for issues with "tight_bbox" redraws
