### 2023-12-01 36c0fa0
fix issues with jupyter inline backend and m.show()
