### 2023-12-01 8a58f39
don't turn off interactive backend with jupyter inline backend (it's no longer necessary)
