### 2023-12-01 5e64b30
fix re-fetching spines and backgrounds on `m.set_frame`
