+++ eomaps/eomaps.py
@@ -5638,11 +5638,8 @@ class Maps(metaclass=_MapsMeta):
         >>>     path_effects=[pe.withStroke(linewidth=7, foreground="m")])
 
         """
-        self.redraw("__SPINES__")
 
         for key in ("fc", "facecolor"):
-            self.redraw("__BG__")
-
             if key in kwargs:
                 self.ax.patch.set_facecolor(kwargs.pop(key))
 
@@ -5705,3 +5702,6 @@ class Maps(metaclass=_MapsMeta):
 
                 self.BM._before_fetch_bg_actions.append(cb)
                 self.ax._EOmaps_rounded_spine_attached = True
+
+            self.redraw("__SPINES__")
+        self.redraw("__BG__")
