### 2023-12-01 5261c9a
fix transparent snapshot issues in jupyter notebooks with `pillow 10.0.0`
