### 2023-12-01 da74de1
make sure backgrounds are properly fetched (requires activating on_draw)
