### 2023-12-01 677e245
allow passing kwargs to WebMap services with m.add_wms.get_service()
