### 2023-12-01 72dd042
use logger for layout-editor warning
