### 2023-12-02 eaeecf9
fix issues with RGB/RGBA composites
