### 2023-12-04 d2f3e79
improve handling of (transparent) snapshots in IPython/Jupyter Notebooks
