### 2023-12-04 #138 @yeelauren
Hi :wave: @raphaelquast ! Great work this far and happy to e-meet you!

Estimated hours spent reviewing:  ~2:45 onboarding, 6 hr reviewing

---

I really like the number of examples, blog articles and documentation provided. The docstrings within the package are very well done as well!  I'll definitely be excited to use this package for future work :)

I have a few questions and some suggestions below, happy to provide them as separate issues to discuss as well:

#### Questions

While I can definitely see the use-case for eomaps, in terms of this review, I think the README and documentation could use a line or two about what the package is and why you might want to use it stated more explicitly.

Since this package is building on cartopy, what were the motivations for creating a separate package versus adding a PR/additional functionality to cartopy directly?

#### Suggestions/Troubleshooting

- The inclusion of specific instructions in the setup for various IDES makes me wonder if I missed some steps for my initial set-up using VSCode. I'd argue that a lot of Developers and Data Scientists are users of VSCode so instructions should be included.
	-  [ ] Add instructions for VSCode
-  Provide instructions on notebook setup with examples? e.g. use jupyter, quarto, etc. to implement some of the notebook cell tricks in https://eomaps.readthedocs.io/en/latest/FAQ.html#from-0-to-eomaps-a-quickstart-guide
- When trying to run tests on my machine (ubuntu) I received this:
	`qt.qpa.plugin: Could not find the Qt platform plugin "wayland" in ""`
	- Is this an error I can ignore? It seems to be related to matplotlib and how ubuntu is rendering the graphics?
- Some errors and deprecation warnings come up while running the test(s) that may need to be addressed in a future version
- Happy to approve package once I can confirm or troubleshoot the error above
