### 2023-12-07 #138 @raphaelquast
Hey @yeelauren,
Thank you very much for your review!
Below you can find a detailed response to your comments (associated adjustments in the package will follow as soon as possible).

---

> I really like the number of examples, blog articles and documentation provided. The docstrings within the package are very well done as well! I'll definitely be excited to use this package for future work :)

Thanks for the kind words! I put a lot of effort in extensively documenting the package since I think it is extremely important (especially for beginners) to get as much information and examples as possible.

---

> While I can definitely see the use-case for eomaps, in terms of this review, I think the README and documentation could use a line or two about what the package is and why you might want to use it stated more explicitly.

Glad to hear that you feel that the package is useful! I agree that the README can be improved... I'll try to come up with some improvements and report back here as soon as they are implemented!

---

> Since this package is building on cartopy, what were the motivations for creating a separate package versus adding a PR/additional functionality to cartopy directly?

Cartopy is closely tied to the matplotlib syntax. Since my intention was to provide a package that goes beyond static maps and provides functionalities specifically tuned for interactive geo-data analysis, supercharging Figure or Axes objects with additional methods didn't seem like a good approach.
To be a bit more specific:
- multi-layer capabilities require extensive adaptions on the way how draw-events are handled in matplotlib
  I am unsure if such modifications are in scope for cartopy
- to simplify interacting with datasets, EOmaps keeps track of the datasets that were used to create the map and provides means to quickly identify nearest-neighbors etc.. (e.g. "pick callbacks" `m.cb.pick...`). For pure cartopy plots there is no need for this additional complexity.
- With EOmaps I wanted to also provide a more hands-on approach by providing a PyQt5 GUI widget... again I'm not sure something like this would be in scope for cartopy
- ...

In general, I try my best to port relevant fixes to cartopy, and it might even be good to one day have a unified package that does it all (or simply a more clear split between what goes into cartopy and what goes into EOmaps), but this would require a lot of time and effort. Right now, with 0 financial support (and ~0.01 contributors) this is out of scope for me.
I am open to participating in discussions on the future of the packages. At the moment my personal focus is to find maintainers and contributors (and ideally financial support) for EOmaps, to clean-up the codebase and to make the API stable and easy-to-use.

---

> The inclusion of specific instructions in the setup for various IDES makes me wonder if I missed some steps for my initial set-up using VSCode. I'd argue that a lot of Developers and Data Scientists are users of VSCode so instructions should be included.

Yes, I agree. I will get myself a copy of VSCode and see what steps are required to get EOmaps running properly and include it in the docs!

---

> Provide instructions on notebook setup with examples? e.g. use jupyter, quarto, etc. to implement some of the notebook cell tricks in [https://eomaps.readthedocs.io/en/latest/FAQ.html#from-0-to-eomaps-a-quickstart-guide](https://eomaps.readthedocs.io/en/latest/FAQ.html#from-0-to-eomaps-a-quickstart-guide)

The main reason why Jupyter Notebooks are a bit under-represented in the docs at the moment is that the interactivity is affected by a bug in the interactive jupyter-matplotlib backend `ipympl`. I tried my best to help leveraging a solution by investigating problems of the python-side of `ipympl` (see [https://github.com/raphaelquast/EOmaps/issues/62](https://github.com/raphaelquast/EOmaps/issues/62)) and the associated `webagg` backend (see https://github.com/matplotlib/matplotlib/pull/27160) , but so far there is little progress concerning Jupyter notebooks since the  problem requires adaptions in the javascript code (and my experience with js is limited).

Nonetheless, static maps now work just fine (and even interactivity is OK if update triggers are not too fast) so I agree that it might be time to better highlight the usability of EOmaps with Jupyter Notebooks in the docs.

My current idea is to actually completely refactor the Examples section. I gained some experience using [myst_nb](https://myst-nb.readthedocs.io/en/latest/) recently and I was thinking about using it to re-write all examples directly as Jupyter Notebooks and render them to the docs.

This will also help introducing a better structure for the examples (e.g. simple - advanced / static - interactive / ...).
I will try to start this process as soon as possible, but it might take some time until I created a decent number of examples.

---

> When trying to run tests on my machine (ubuntu) I received this: qt.qpa.plugin: Could not find the Qt platform plugin "wayland" in ""

So far I have never encountered this error and unfortunately I cannot reproduce it locally (tested on Linux Mint 21.2 Cinnamon) and also GitHub actions currently run on Ubuntu 22.04.3 and do not seem to be affected by this.

After doing some research, it seems to be a more general issue with Qt that affects multiple other packages as well. I'm not sure what I can do to solve this except for pointing you to some associated discussions I found that might provide some (OS-specific) workarounds:

- [https://stackoverflow.com/questions/69994530/qt-qpa-plugin-could-not-find-the-qt-platform-plugin-wayland](https://stackoverflow.com/questions/69994530/qt-qpa-plugin-could-not-find-the-qt-platform-plugin-wayland)
- [https://github.com/NixOS/nixpkgs/issues/158806](https://github.com/NixOS/nixpkgs/issues/158806)
- [https://bugs.archlinux.org/task/72972](https://bugs.archlinux.org/task/72972)
- [https://askubuntu.com/questions/1354665/matplotlib-complains-about-not-finding-qt-platform-plugin-wayland](https://askubuntu.com/questions/1354665/matplotlib-complains-about-not-finding-qt-platform-plugin-wayland)

If you feel that there is something that I can do in EOmaps to resolve this, please open an Issue and I'll do my best to help in finding a solution!

---

> Some errors and deprecation warnings come up while running the test(s) that may need to be addressed in a future version

I had a close look at the warnings reported in the latest test-run (from [here](https://github.com/raphaelquast/EOmaps/actions/runs/7084706375/job/19279496985)):

- cartopy `DownloadWarning` can savely be ignored. They just indicate that the associated NaturalEarth dataset has never been accessed before and needs to be downloaded (and cached for future use).
- geopandas `UserWarning: Geometry is in a geographic CRS. ...` can savely be ignored. it just indicates that `geometry.buffer(..)` was used in a geographic crs and therefore the size of the buffer might vary depending on the location of the data.
- [ ] numpy `RuntimeWarning: numpy.ndarray size changed, may indicate binary incompatibility.`:  This indicates that the uploaded NetCDF file was created with an earlier numpy version... I'll update the associated test-dataset!
- [ ] numpy `UserWarning: Warning: 'partition' will ignore the 'mask' of the MaskedArray.`: I have to investigate this.
- [ ] geopandas `FutureWarning: In a future version, "df.iloc[:, i] = newvals" will attempt to set the values inplace instead of always setting a new array. To retain the old behavior, use either "df[df.columns[i]] = newvals" or, if columns are non-unique, "df.isetitem(i, newvals)"`: I have to investigate this

- [ ] `DeprecationWarning: invalid escape sequence '\*'` This is from a docstring.
	It originates from trying to document what happens with additional keyword-arguments, e.g.  `**kwargs`
	The problem here is that in re-structured text and markdown, `**` is also used for bold-text which is why I had to escape `\*`. It seems that this is now Deprecated and I have to find a better solution.
	I'm happy about any suggestions for a clean way how to do this!

	To clarify, I'd need a "re-structured-text and markdown save" version of a docstring like this that can be parsed with `sphinx-autodoc`:
	```
	def func(**kwargs):
		"""
		A nice function.

		Parameters
		----------
		**kwargs:
			Keyword-arguments passed to print.
		"""
		print(kwargs)
	```
