### 2023-12-07 c05d6c3
add new "overlay_layer" keypress callback and improve "show_layer" callback
