### 2023-12-07 de7c545
add new "overlay_layer" keypress callback and improve "show_layer" callback
