### 2023-12-07 5062bdc
fix removesuffix requires python >= 3.9
