### 2023-12-08 57d2c20
fix spines are not properly re-drawn with m.set_frame if rounded=0
