+++ docs/contribute_env.yml
@@ -38,6 +38,4 @@ dependencies:
   - sphinx-copybutton
   - sphinx
   - docutils
-  - pip
-  - pip:
-    - sphinx_rtd_theme
+  - sphinx_rtd_theme
