### 2023-12-08 a339d0c
update readthedocs to install from source
