### 2023-12-08 9fc25ec
update readthedocs to install from source
