+++ docs/conf.py
@@ -26,6 +26,8 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx_copybutton",
     "sphinx_rtd_theme",
+    "myst_nb",
+    "sphinx_design",
 ]
 
 
@@ -61,10 +63,29 @@ napoleon_preprocess_types = False
 napoleon_type_aliases = None
 napoleon_attr_annotations = True
 
+myst_update_mathjax = False  # to use single $x^2$ for equations
+myst_render_markdown_format = "myst"  # to parse markdown output with MyST parser
+myst_enable_extensions = ["dollarmath", "colon_fence"]
+myst_title_to_header = True
+
+nb_execution_mode = "cache"
+nb_execution_timeout = 120
+
 # handle compiler warnings for duplicate labels due to documents
 # included via the  ..include:: directive
-exclude_patterns = ["introduction.rst"]
+exclude_patterns = [
+    "introduction.rst",
+    "build",
+    "jupyter_execute/*",
+    ".jupyter_cache/*",
+    ".virtual_documents/*",
+]
 
+source_suffix = {
+    ".rst": "restructuredtext",
+    ".ipynb": "myst-nb",
+    ".myst": "myst-nb",
+}
 
 # a global substitution used to fix issues with images in tables
 # in the mobile-theme (without a span they get resized to 0. This forces a size
