### 2023-12-08 c8f79ef
re-write feature section as notebook
