+++ .readthedocs.yaml
@@ -6,7 +6,7 @@ build:
     python: "mambaforge-4.10"
 
 conda:
-  environment: tests/docs_env.yml
+  environment: docs/docs_env.yml
 
 python:
   install:
