### 2023-12-08 c2a47c0
update mambaforge version for docs
