+++ docs/docs_env.yml
@@ -31,6 +31,6 @@ dependencies:
   - docutils<0.19
   - sphinx=7.2.6
   - sphinx-copybutton=0.5.2
-  - myst_nb=1.0.0
+  - myst-nb=1.0.0
   - sphinx-design=0.5.0
   - sphinx_rtd_theme=1.3
