### 2023-12-08 c88a6bf
avoid activating snapshot_on_update by default
