### 2023-12-08 e6f4c8e
fix creating a snapshot on m.show() in jupyter inline backend
