### 2023-12-09 228ae9c
fix setting snapshot_on_update
