### 2023-12-09 e4be13e
fix setting snapshot_on_update
