### 2023-12-09 2be67ce
add unittest to test all Jupyter Notebook code-cells in the docs
