+++ eomaps/qtcompanion/signal_container.py
@@ -1,27 +1,27 @@
 """A container class for signals sent to the CompanionWidget"""
 
 
-from PyQt5.QtCore import QObject, pyqtSignal
+from qtpy.QtCore import QObject, Signal
 
 
 class _SignalContainer(QObject):
-    cmapsChanged = pyqtSignal()
+    cmapsChanged = Signal()
 
-    clipboardKwargsChanged = pyqtSignal()
+    clipboardKwargsChanged = Signal()
 
-    dataPlotted = pyqtSignal()
+    dataPlotted = Signal()
 
     # -------- shape drawer
-    drawFinished = pyqtSignal()
-    drawAborted = pyqtSignal()
-    drawStarted = pyqtSignal(str)
+    drawFinished = Signal()
+    drawAborted = Signal()
+    drawStarted = Signal(str)
 
     # -------- annotation editor
-    annotationEditorActivated = pyqtSignal()
-    annotationEditorDeactivated = pyqtSignal()
-    annotationSelected = pyqtSignal()
-    annotationEdited = pyqtSignal()
+    annotationEditorActivated = Signal()
+    annotationEditorDeactivated = Signal()
+    annotationSelected = Signal()
+    annotationEdited = Signal()
 
     # -------- layout editor
-    layoutEditorActivated = pyqtSignal()
-    layoutEditorDeactivated = pyqtSignal()
+    layoutEditorActivated = Signal()
+    layoutEditorDeactivated = Signal()
