### 2023-12-09 9e7bb89
add nbformat to test-env
