### 2023-12-09 759d9b2
add basic info on how to use VSCode/VSCodium
