+++ docs/FAQ.rst
@@ -5,153 +5,175 @@
   :local:
   :depth: 1
 
-.. include:: introduction.rst
+
+Need some help to setup python for EOmaps?
+******************************************
+
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+
+   quickstart_quide
+
+
+Have a look at the :ref:`quickstart_guide` to learn how to setup a python environment that can be
+used to create EOmaps maps!
+
 
 .. _configuring_the_editor:
 
 Configuring the editor (IDE)
 ****************************
 
+EOmaps can be used with whatever editor you like... however, some editors have special settings that can be
+adjusted to improve your mapping experience with **EOmaps**:
+
 .. _config_spyder:
 
-Spyder IDE
-----------
+.. dropdown:: Spyder
 
-To use the whole potential of EOmaps with the awesome `Spyder IDE <https://www.spyder-ide.org>`_  ,
-the plot-settings must be adjusted to ensure that ``matplotlib`` plots remain interactive.
+    .. currentmodule:: eomaps.eomaps
 
-- By default, plots are rendered as static images into the "plots-pane"... to avoid this and create
-  interactive ``matplotlib`` widgets instead, go to the preferences and set the "Graphics backend" to "Automatic" :
+    To use the whole potential of EOmaps with the awesome `Spyder IDE <https://www.spyder-ide.org>`_  ,
+    the plot-settings must be adjusted to ensure that ``matplotlib`` plots are created as interactive Qt widgets.
 
-.. image:: _static/spyder_preferences.png
+    - By default, figures are rendered as static images into the `plots-pane <http://docs.spyder-ide.org/current/panes/plots.html>`_. To avoid this and **create interactive (popup) figures** instead, go to the preferences and set the "Graphics Backend" to "Automatic" :
+
+    .. image:: _static/spyder_preferences.png
+
+    .. raw:: html
+
+       <p/>
+
+    .. note::
+
+      If the gaphics-backend is set to *"Automatic"*, you can still plot static snapshots of a figure to the "plots-pane" with :py:meth:`Maps.snapshot`!
+
 
 .. _config_vscode:
 
-VSCode / VSCodium
------------------
 
-.. currentmodule:: eomaps.eomaps
+.. dropdown:: VSCode / VSCodium
 
-In general, EOmaps works "out of the box" with `VSCode <https://code.visualstudio.com/>`_ or the open-source variant `VSCodium <https://vscodium.com/>`_ (together with the standard `Python <https://marketplace.visualstudio.com/items?itemName=ms-python.python>`_ extension).
+    .. currentmodule:: eomaps.eomaps
 
-However, there are some tipps that might help with your mapping workflow:
+    In general, EOmaps works "out of the box" with `VSCode <https://code.visualstudio.com/>`_ or the open-source variant `VSCodium <https://vscodium.com/>`_ (together with the standard `Python <https://marketplace.visualstudio.com/items?itemName=ms-python.python>`_ extension).
 
-- In a normal python-terminal, the default matplotlib backend will be ``QtAgg`` in a **non-interactive mode**.
-  This means that you must call :py:meth:`Maps.show` at the end of the script to actually show the figure.
-  Once the figure is shown, **the terminal is blocked until the figure is closed**.
+    However, there are some tipps that might help with your mapping workflow:
 
-- To **avoid blocking the terminal** while a figure is running, you can **activate matplotlib's interactive-mode** using
+    - In a normal python-terminal, the default matplotlib backend will be ``QtAgg`` in a **non-interactive mode**.
[…]
