+++ docs/FAQ.rst
@@ -25,10 +25,9 @@ used to create EOmaps maps!
 Configuring the editor (IDE)
 ****************************
 
-EOmaps can be used with whatever editor you like... however, some editors have special settings that can be
-adjusted to improve your mapping experience with **EOmaps**:
+EOmaps can be used with whatever editor you like!
 
-.. _config_spyder:
+However, for some editors there are special settings that can be adjusted to improve your mapping experience with **EOmaps**:
 
 .. dropdown:: Spyder
 
@@ -49,15 +48,11 @@ adjusted to improve your mapping experience with **EOmaps**:
 
       If the gaphics-backend is set to *"Automatic"*, you can still plot static snapshots of a figure to the "plots-pane" with :py:meth:`Maps.snapshot`!
 
-
-.. _config_vscode:
-
-
 .. dropdown:: VSCode / VSCodium
 
     .. currentmodule:: eomaps.eomaps
 
-    In general, EOmaps works "out of the box" with `VSCode <https://code.visualstudio.com/>`_ or the open-source variant `VSCodium <https://vscodium.com/>`_ (together with the standard `Python <https://marketplace.visualstudio.com/items?itemName=ms-python.python>`_ extension).
+    In general, EOmaps should work "out of the box" with `VSCode <https://code.visualstudio.com/>`_ or the open-source variant `VSCodium <https://vscodium.com/>`_ (together with the standard `Python <https://marketplace.visualstudio.com/items?itemName=ms-python.python>`_ extension).
 
     However, there are some tipps that might help with your mapping workflow:
 
@@ -78,8 +73,9 @@ adjusted to improve your mapping experience with **EOmaps**:
 
        If you run a whole script using the **interactive mode**, the script will run until the end and then usually terminate the associated kernel... and in turn also closing the figure! If you want to keep the figure open, either make sure that the terminal is kept alive by entering debug-mode, or avoid activating the interactive mode and block the terminal with `m.show()`.
 
-    Interactive IPython (e.g. Jupyter)
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    .. raw:: html
+
+       <font size=+1><b>Interactive IPython (e.g. Jupyter)</b></font><p>
 
     If you enjoy interactive coding in a Jupyter-Notebook style, make sure to have a look at the nice `Jupyter <https://marketplace.visualstudio.com/items?itemName=ms-toolsai.jupyter>`_ extension!
 
@@ -94,9 +90,6 @@ adjusted to improve your mapping experience with **EOmaps**:
       - For interactive (inline) figures, you'll need to install `ipympl <https://github.com/matplotlib/ipympl>`_ and then activate the ``widget`` with ``%matplotlib widget``.
 
       - For more details, see the associated section for :ref:`config_jupyter_notebook`
-
-
-.. _config_pycharm:
 
 .. dropdown:: PyCharm
 
@@ -125,19 +118,23 @@ adjusted to improve your mapping experience with **EOmaps**:
 
     .. image:: _static/pycharm_preferences_2.png
 
-
-.. _config_jupyter_notebook:
-
 .. dropdown:: Jupyter Notebooks
 
-    To get the most out of EOmaps in Jupyter Notebooks, use `jupyter lab <https://jupyterlab.readthedocs.io/en/stable/>`_  together with the `ipympl <https://github.com/matplotlib/ipympl>`_ (``widget``) backend.
+    .. currentmodule:: eomaps.eomaps
 
-    - To install, use ``conda install -c conda-forge ipympl``
+    When working with Jupyter Notebooks, we recommend to use `Jupyter Lab <https://jupyterlab.readthedocs.io/en/stable/>`_.
 
[…]
