### 2023-12-11 96312d7
make sure the colorbar always uses the associated figure
