### 2023-12-11 da5f260
add `.show` and `.snapshot` to `MapsGrid` objects
