### 2023-12-11 #138 @raphaelquast
@yeelauren Basic info on how to use VSCode/VSCodium is now added to the docs!
(currently still in dev-branch for next release since there are also changes in the way how the docs are built).
You can have a look at the updated content [here](https://eomaps.readthedocs.io/en/dev/FAQ.html#configuring-the-editor-ide).
