### 2023-12-11 1215d4f
make sure m.show_layer adds a snapshot with jupyter-inline
