### 2023-12-11 39113b7
make sure layers are re-fetched in case colorbars change axes sizes
