### 2023-12-11 42dfc2e
allow overriding the layer at which a colorbar is plotted
