### 2023-12-11 b3f926e
fix forcing colorbar position inheritance
