### 2023-12-13 7a20e49
fix treatment of cb-histogram weights with `out_of_range_vals="masked"`
