+++ eomaps/colorbar.py
@@ -892,7 +892,14 @@ class ColorBar:
         self._set_extend(z_data)
 
         if self._out_of_range_vals == "mask":
-            z_data = z_data[(z_data >= self._vmin) & (z_data <= self._vmax)]
+            data_range_mask = (z_data >= self._vmin) & (z_data <= self._vmax)
+            z_data = z_data[data_range_mask]
+
+            # make sure that histogram weights are masked accordingly if provided
+            if "weights" in self._hist_kwargs:
+                self._hist_kwargs["weights"] = self._hist_kwargs["weights"][
+                    data_range_mask
+                ]
 
         # make sure the norm clips with respect to vmin/vmax
         # (only clip if either vmin or vmax is not None)
