### 2023-12-13 3533ef4
fix treatment of histogram style kwargs for colorbars (lw, ec, fc, ...)
