### 2023-12-13 4d6b313
address numpy binary incompatibility warning when reading NetCDFs
