### 2023-12-13 #138 @raphaelquast
@yeelauren I had a closer look at the numpy RuntimeWarning that appears in the tests...
> numpy RuntimeWarning: numpy.ndarray size changed, may indicate binary incompatibility.: This indicates that the uploaded NetCDF file was created with an earlier numpy version... I'll update the associated test-dataset!

This is apparently caused by some unresolved issue regarding the import-order of `numpy` and `netcdf4`.
Here is the associated (open) issue from `xarray`: https://github.com/pydata/xarray/issues/7259

My solution for now is to try to import `netCDF4` on init to ensure it is imported prior to `numpy` to address this warning. (see https://github.com/raphaelquast/EOmaps/pull/205/commits/da5de8d50ce193bb466b49bdf48e9e1129808a1e and the associated [test-run](https://github.com/raphaelquast/EOmaps/actions/runs/7201632536/job/19618178043?pr=205))
