### 2023-12-15 da22b42
fix linewidths of colorbar patches after layout editor exit
