### 2023-12-15 7d18dbc
fix identifying colorbar position from another colorbar instance
