### 2023-12-15 #138 @jhkennedy
Overall *fantastic* work @raphaelquast!

*NOTE:* If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that a badge for pyOpenSci peer-review will be provided upon acceptance.)*

Estimated hours spent reviewing: 2 on-boarding; 4-6 playing/using.

---

#### Overall
Overall, I'm really excited about this package, and it significantly reduces some of the boilerplate for the CartoPy maps I've produced in the past. Coming to the GitHub, I was impressed with the README, releases, and release notes immediately, and I could see many of the "community standards" (e.g., code of conduct, contributing, license, etc.) right up front.

The documentation is *fantastic*, as well as the examples, and it was overall pretty easy to jump in as a user.

I find myself wishing you were submitting to JOSS so you could expand on the *why* -- why create the package, why you designed it the way you did, etc., which @yeelauren also hit on. I think JOSS is likely a better place for that than in the README or docs.

#### Detailed Comments

Note: I'm reviewing specifically v7.3.2, as that was latest v7.3 release when I started in seriously.

##### A) First impressions
1. Using a copy-left license like GPL means, for me, I will happily use and contribute to EOMaps, but I will be unable to build upon it as I'm obligated to use a permissive use license in my work. This is fine for EOMaps for the most part, I think, as it's more of an end-user package than a library.
2. On the repository main page on GitHub, in the "About" (upper-left) there's a documentation link pointing to `github.io` but the README and rest of the docs links point to `readthedocs.io` -- I'd change the about to also point to ReadTheDocs
3. It's time to move from `setup.py` to `pyproject.toml`
4. It's not clear what python versions EOMaps supports:
   1. There's no badge on the README, or in the PyPI metadata about supported versions
   2. [`python_requires`](https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#python-requirement) is not set in `setup.py`
   3. `classifiers` in the `setup.py` only list Python 3.7, but [I see](https://github.com/raphaelquast/EOmaps/blob/master/.github/workflows/testMaps.yml#L16) that EOMaps is only tested on Python 3.8-3.10
   4. The [conda-forge recipe](https://github.com/conda-forge/eomaps-feedstock/blob/main/recipe/meta.yaml#L21) supports Python >=3.6, so users can install into Python 3.11 and 3.12, but it's not being tested on those versions

   I'd really like to see this cleaned up, with the supported python versions clearly listed, and all supported versions tested.

   *Note: I also have some additional comments related to this when setting up a development environment.

##### B) Setting up a development environment
1. This proved a bit more challenging as there was no top-level `environment.yml` or `requirements.txt` that'd set up a development environment. So I had to go: README -> installation docs -> scroll to bottom to find a small link to where the dev environment was described in the contributing docs. I didn't naturally jump to the contributing guide initially, and the section in the contributing guide also didn't work because:
   1. it instructs you to create a conda environment from `eomaps.yml`, which does not exist at the top level and the guide doesn't tell you where it exists in the repository. I ended up running `find . -iname '*.yml'` which then informed me that it doesn't exist *at all* and I likely wanted `docs/contribute_env.yml`.
   2. `docs/contribute_env.yml` at least matches the content of `eomaps.yml` as listed in the contributing guide (in the docs) and had I looked at `contributing.rst` I would have seen that *is* the environment file being used in the docs:
   ```rst
   Content of ``eomaps.yml``:

   .. literalinclude:: ../docs/contribute_env.yml
   ```

2. At this point, I was unsure if the Docs were in-sync with the latest released version, but the docs don't list a version number explicitly anywhere. I'd like to at least see that on the front page of the docs and explicitly as a version listed on ReadTheDocs (even if you don't keep old versions around).

3. When reading through the contributing guide:
   1. Overall I found it very well done! :+1:
   2. Consider switching to `git switch` over `get checkout` which is more confusing and does both switching branches and restoring specific revisions.
   3. You point people to the miniconda installer but then have users install mamba. I'd just point people to [miniforge](https://github.com/conda-forge/miniforge) which includes `conda` and `mamba` from the get go.
   4. There's *nothing* on how to run tests. I want to set up my environment and immediately run tests before starting in on changes. I had to go look at `.github/workflows/testMaps.yml` to see how tests should be run.

##### C) Running tests
1. WHOAH soo many figure windows get created and deleted! I was not prepared for that. The tests took just long enough I switched away and started on something else when my desktop started exploding with figure windows being created/deleted. Do you need this to happen? If so, you should definitely document how to run tests and *warn* about this.
2. I had 1 test failure from checkout of version 7.3.2 (I'll post the full traceback in a subsequent comment)
   ```
   FAILED tests/test_plot_shapes.py::TestPlotShapes::test_contour - AttributeError: 'NoneType' object has no attribute 'dpi'
   ```
3. `tests/test_doc_codeblocks.py` hangs between the third and fourth codeblock until you close a figure window, but I couldn't easily tell which codeblock this corresponds to easily, and it looks like there's no logging of which code block is being run that I could make visible with the `-s` pytest option.

##### D) Conda environment files
I am not really a fan how the conda environment files are set up here. Recommendations:
1. Provide a single `environment.yml` at the repository root, and:
   1. use this for **both testing and development**. This can be accomplished by setting the python requirement to `python>=3.6` or whatever your minimum supported version is. If you want a *preferred* development version that's not whatever's the latest conda/mamba resolve, add setting the version restriction to the dev docs.
   2. `rtree` is in `test_env.yml` in the runtime dependencies but not in `contribute_env.yml`. It should be moved to the appropriate section or removed if not needed.
   3. I would *really* prefer git *not* be installed in the environment -- it's much more common to use the system install of git and this could lead to very unexpected things with miss-match system and environment versions.
   4. Related to :point_up: I would encourage pre-commit, but not install it into the environment, especially since you also have a pre-commit workflow: `.github/workflows/pre-commit.yml`
2. `docs_env.yml` is fine where it is, but I'd just include sphinx-autobuild since you [highlight it in the docs](https://eomaps.readthedocs.io/en/latest/contribute.html#building-the-docs)

##### E) Reading the FAQs
I hopped over to the FAQs after messing around with the tests to see if anything was addressed there.
1. In the "From 0 to EOmaps - a quickstart guide" section
   1. After installing miniconda, you ask users to open a Anaconda Prompt terminal, but that is windows specific.
   2. `matplotlib` is unnecessarily included in the ~~2nd~~ 3rd TODO
   3. Really, the TODO blocks are subsections -- It'd be nice to at least number them.
2. In the "Configuing the editor (IDE)" setcion:
   1. I did not have any hanging with interactive plots like mentioned in the PyCharm section so setting `Gt5Agg` may be unnecessary now (fine to still recommend though)
   2. It's Pycharm Professional not commercial.
   3. The latest Pycharm Version don't have "Python Scientific" anymore; disabling plots into the tool window is located at "Setting -> Tools -> Python Plots -> [] show plots in tool window"
3. I really like the [Important Changes between major versions](https://eomaps.readthedocs.io/en/latest/FAQ.html#important-changes-between-major-versions) but
   1. It'd be good to include that in an `UPGRADING.rst` at the top level of the repo or at list mention it right up front in the README
   2. It'd also be good to reverse the order so the changes between the latest version and the last major version are at the top

##### F) Finally taking it for a spin with examples and data
1. in [the Basics](https://eomaps.readthedocs.io/en/latest/api_basics.html) the layer management figure needs to be bigger or zoomable. If that doesn't happen as an overlay, I'd set it top open in a new tab/window (`target="_blank"`)
2. In the [data visualization](https://eomaps.readthedocs.io/en/latest/api_data_visualization.html) section:
   1. in your note on data-reprojection about 1D coordinate vectors, you should also mention that these will be broadcast using matrix indexing, not cartesian indexing, as this could lead to the data being transposed in the plot.
   2. The speed and performance notes might also be worth linking to or putting in the FAQ
   3. Optional dependencies should be discussed on the installation page
3. In the [examples](https://eomaps.readthedocs.io/en/latest/EOmaps_examples.html) section:
   1. It would be nice to number the example's content to match the example figure numbers and code files
   2. They also appear out of order in places, e.g., ? -> 3 -> 2.
   3. The actual examples code being located in the tests directory was unexpected -- I initially looked for it in the docs directory
4. The [API reference](https://eomaps.readthedocs.io/en/latest/generated/eomaps.html) is hard to navigate
   1. it's not in alphabetical order, or is at least missing headings, and the `add_feature` and `add_wms` methods were hard to find
   2. `add_feature.preset` isn't documented and seems to be a NaturalEarth_features class. In fact, the `add_feature` reference doesn't seem to match at all
   3. IDE tab completion also doesn't discover NE categories; hard to know what I'm able to access
   4. I ended up just opening `NE_features.json`, reformatting it so it's not on a single line, and looking at what's available there.
