### 2023-12-17 raphaelquast/EOmaps#209 @banesullivan
License may not be properly implemented?
Please reference the instructions from the license itself on how to apply it to your work. My impression is that you are not correctly following this at this time and technically it is unclear who the copyright is held by.

https://github.com/raphaelquast/EOmaps/blob/71361494f5af1fb6674c9862339d0e40946f5e89/LICENSE#L623-L674

---

I also want to highlight a note from @jhkennedy in https://github.com/pyOpenSci/software-submission/issues/138#issuecomment-1858583110 that I second **_greatly_**:

> Using a copy-left license like GPL means, for me, I will happily use and contribute to EOMaps, but I will be unable to build upon it as I'm obligated to use a permissive use license in my work. This is fine for EOMaps for the most part, I think, as it's more of an end-user package than a library.

Major +1 from me.

-----

I want to ensure that you have deeply thought about your choice of license here and feel confident in using and applying that license to your work as well as understanding the tradeoffs that license may provide:

-  If **no to the above**, then I'll do my best to provide you with support and resources to make that decision with confidence. A good place to start is PyOpenSci's licensing guide here: https://www.pyopensci.org/python-package-guide/documentation/repository-files/license-files.html
- If **yes to the above**, then I'll do my best to help ensure you are properly applying that license to your work
