### 2023-12-17 raphaelquast/EOmaps#210 @banesullivan
Migrate to pyproject.toml
Tracking a comment from @jhkennedy in https://github.com/pyOpenSci/software-submission/issues/138#issuecomment-1858583110 here:

> It's time to move from setup.py to pyproject.toml

While `setup.py` is not technically deprecated, ref https://packaging.python.org/en/latest/discussions/setup-py-deprecated/, `pyproject.toml` is the new standardized format to describe project metadata declaratively, introduced with [PEP 621](https://peps.python.org/pep-0621/). It's easier to work with and allows for shared configuration between different tools.
