### 2023-12-17 raphaelquast/EOmaps#211 @banesullivan
Sunset Python 3.7 support and clarify supported versions
According to [PEP 537](https://peps.python.org/pep-0537/), Python 3.7 has stopped maintenance in June 2023. NumPy's [NEP 29](https://numpy.org/neps/nep-0029-deprecation_policy.html) dropped Python 3.7 in 2021. So it's time to drop it here too, and replace it with Python 3.11 where used on CI:

https://github.com/raphaelquast/EOmaps/blob/71361494f5af1fb6674c9862339d0e40946f5e89/.github/workflows/Publish_to_PyPi.yml#L19

Please also reference a request from @jhkennedy in https://github.com/pyOpenSci/software-submission/issues/138#issuecomment-1858583110:

> It's not clear what python versions EOMaps supports:
>
> - There's no badge on the README, or in the PyPI metadata about supported versions
> - [python_requires](https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#python-requirement) is not set in setup.py
> - classifiers in the setup.py only list Python 3.7, but [I see](https://github.com/raphaelquast/EOmaps/blob/master/.github/workflows/testMaps.yml#L16) that EOMaps is only tested on Python 3.8-3.10
> - The [conda-forge recipe](https://github.com/conda-forge/eomaps-feedstock/blob/main/recipe/meta.yaml#L21) supports Python >=3.6, so users can install into Python 3.11 and 3.12, but it's not being tested on those versions
>
> I'd really like to see this cleaned up, with the supported python versions clearly listed, and all supported versions tested.
>
> *Note: I also have some additional comments related to this when setting up a development environment.
