### 2023-12-17 e2905e7
Draft improvements to contributing guide
