### 2023-12-18 c373e04
Add support info to README
