### 2023-12-18 #138 @raphaelquast
> Overall, great work on EOmaps, @raphaelquast!! I'm excited to enter the next stage of this submission given the fantastic reviews by @yeelauren and @jhkennedy (thank you both very, very much for making the time!!!).

@banesullivan Thank you very much! I'm indeed very exited that EOmaps is ready to continue to the next stage and I'll start implementing the required changes and suggestions right away!

> It looks like there are a few things that need to be addressed but overall EOmaps appears in good health. I've further opened a few issues in the EOmaps repository directly and linked back here -- some of these issues are just restating what our reviewers pointed out above, but I felt needed a dedicated issue to track the task.

Thanks a lot for taking the time to open the issues! I agree that this will make it easier to track progress on the remaining TODOs.

> @raphaelquast, I'd appreciate if you would likewise track the reviewer tasks here as new issues in EOmaps where you feel appropriate (no need to open an issue for every single comment). Similarly, I'd appreciate if you would link this issue for any Pull Requests to EOmaps resulting from this feedback.

OK, I will do that!
