### 2023-12-19 1bd02f3
fix restoring visible layer after calling savefig with `transparent=False` (this addresses problems with inset-maps reported in #214)
