### 2023-12-19 raphaelquast/EOmaps#209 @raphaelquast
Thank you very much for bringing this up!
I agree that licensing has to be improved and that GPL might not be the best choice for EOmaps due to the "same license" clause.
I will check the implications of switching to MPL or maybe LGPL and report back once I have a good idea of how to continue!
