### 2023-12-19 53e97f4
switch from setup.py to pyproject.toml for install
- single-source version number to pyproject.toml
- implement optional dependency groups
