+++ eomaps/_version.py
@@ -1,1 +1,0 @@
-__version__ = "7.3.3"
