+++ pyproject.toml
@@ -1,0 +1,69 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+include = ["eomaps"]
+
+[project]
+name = "eomaps"
+version = "7.4"
+description = "A library to create interactive maps of geographical datasets."
+requires-python = ">=3.8"
+
+authors = [
+  { name="Raphael Quast", email="raphael.quast@geo.tuwien.ac.at" },
+]
+
+keywords = ["Visualization", "Plotting", "Maps", "Geographical Data"]
+
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
+    "Operating System :: OS Independent",
+]
+
+dependencies = [
+    "numpy",
+    "scipy",
+    "matplotlib>=3.4",
+    "cartopy>=0.20.0",
+    "descartes",
+    "pyproj",
+    "packaging",
+    "click"
+]
+
+[project.optional-dependencies]
+
+all = ["eomaps[io, classify, wms, shade, gui]"]
+
+all_nogui = ["eomaps[io, classify, wms, shade]"]
+
+io = [
+    "pandas",
+    "geopandas",
+    "xarray",
+    "netcdf4",
+    "rioxarray"
+]
+
+classify = ["mapclassify"]
+
+wms = [
+    "owslib",
+    "requests",
+    "cairosvg",
+]
+
+shade = ["datashader"]
+
+gui = [
+    "PyQt5",
+    "qtpy"
+    ]
+
+
+[project.urls]
+Documentation = "https://eomaps.readthedocs.io/"
+Repository = "https://github.com/raphaelquast/eomaps"
