### 2023-12-19 f5ca61f
make sure package is installed in unittests
