### 2023-12-20 raphaelquast/EOmaps#209 @raphaelquast
@banesullivan I spent some time investigating details of licenses and their implications and there are some remaining questions where I was not able to find a definitive answer:

- To what extent am I responsible to choose a license that is compliant with the licenses of my dependencies?
- Considering the fact that one of the current core dependencies (e.g. `cartopy`) is licensed as `LGPL` would you recommend to just switch to `LGPL` or do you think it is still preferable (and legally OK) to still go with a more permissive license such as `MIT` or `BSD`? (what would be the benefit of this if the package still contains an `LGPL` licensed dependency ...especially since (as noted by @jhkennedy), EOmaps is not really intended as a "library" but rather as a end-user package...)
   - ... or in other words... would switching to a permissive license somewhat imply a *"long-term goal"* to get rid of any dependencies with more restrictive licenses?

Any insights from your side on this would be highly appreciated!

---

To get an idea of what my dependencies are using, I checked with the nice  [pip-licenses](https://github.com/raimon49/pip-licenses) tool:

Here are some results (using dependency-groups from #216).
(output of `pip-licenses --order=license` in a fresh environment with only `eomaps` installed)

<summary>minimal dependencies (e.g. <code>pip install eomaps</code> ) </summary>

| Name            | Version    | License                                                 |
|-----------------|------------|---------------------------------------------------------|
| packaging       | 23.2       | Apache Software License; BSD License                    |
| python-dateutil | 2.8.2      | Apache Software License; BSD License                    |
| click           | 8.1.7      | BSD License                                             |
| colorama        | 0.4.6      | BSD License                                             |
| contourpy       | 1.2.0      | BSD License                                             |
| cycler          | 0.12.1     | BSD License                                             |
| descartes       | 1.1.0      | BSD License                                             |
| kiwisolver      | 1.4.5      | BSD License                                             |
| numpy           | 1.26.2     | BSD License                                             |
| scipy           | 1.11.4     | BSD License                                             |
| shapely         | 2.0.2      | BSD License                                             |
| Cartopy         | 0.22.0     | GNU Lesser General Public License v3 or later (LGPLv3+) |
| Pillow          | 10.1.0     | Historical Permission Notice and Disclaimer (HPND)      |
| pyshp           | 2.3.1      | MIT                                                     |
| fonttools       | 4.47.0     | MIT License                                             |
| pyparsing       | 3.1.1      | MIT License                                             |
| pyproj          | 3.6.1      | MIT License                                             |
| six             | 1.16.0     | MIT License                                             |
| certifi         | 2023.11.17 | Mozilla Public License 2.0 (MPL 2.0)                    |
| matplotlib      | 3.8.2      | Python Software Foundation License                      |

<summary>full set of dependencies (e.g. <code>pip install eomaps[all]</code> ) </summary>

| Name               | Version      | License                                                 |
|--------------------|--------------|---------------------------------------------------------|
| importlib-metadata | 7.0.0        | Apache Software License                                 |
| requests           | 2.31.0       | Apache Software License                                 |
| rioxarray          | 0.15.0       | Apache Software License                                 |
| tzdata             | 2023.3       | Apache Software License                                 |
| xarray             | 2023.12.0    | Apache Software License                                 |
| packaging          | 23.2         | Apache Software License; BSD License                    |
| python-dateutil    | 2.8.2        | Apache Software License; BSD License                    |
| cligj              | 0.7.2        | BSD                                                     |
| llvmlite           | 0.41.1       | BSD                                                     |
| multipledispatch   | 1.0.0        | BSD                                                     |
| partd              | 1.4.1        | BSD                                                     |
| OWSLib             | 0.29.3       | BSD License                                             |
| affine             | 2.4.0        | BSD License                                             |
| cairocffi          | 1.6.1        | BSD License                                             |
| click              | 8.1.7        | BSD License                                             |
| click-plugins      | 1.1.1        | BSD License                                             |
| cloudpickle        | 3.0.0        | BSD License                                             |
| colorama           | 0.4.6        | BSD License                                             |
| contourpy          | 1.2.0        | BSD License                                             |
| cssselect2         | 0.7.0        | BSD License                                             |
| cycler             | 0.12.1       | BSD License                                             |
| dask               | 2023.12.1    | BSD License                                             |
| datashader         | 0.16.0       | BSD License                                             |
| descartes          | 1.1.0        | BSD License                                             |
| fiona              | 1.9.5        | BSD License                                             |
| fsspec             | 2023.12.2    | BSD License                                             |
| geopandas          | 0.14.1       | BSD License                                             |
| idna               | 3.6          | BSD License                                             |
| joblib             | 1.3.2        | BSD License                                             |
| kiwisolver         | 1.4.5        | BSD License                                             |
| locket             | 1.0.0        | BSD License                                             |
| lxml               | 4.9.4        | BSD License                                             |
| mapclassify        | 2.6.1        | BSD License                                             |
| networkx           | 3.2.1        | BSD License                                             |
| numba              | 0.58.1       | BSD License                                             |
| numpy              | 1.26.2       | BSD License                                             |
| pandas             | 2.1.4        | BSD License                                             |
| param              | 2.0.1        | BSD License                                             |
| pycparser          | 2.21         | BSD License                                             |
| pyct               | 0.5.0        | BSD License                                             |
| rasterio           | 1.3.9        | BSD License                                             |
| scikit-learn       | 1.3.2        | BSD License                                             |
| scipy              | 1.11.4       | BSD License                                             |
| shapely            | 2.0.2        | BSD License                                             |
| threadpoolctl      | 3.2.0        | BSD License                                             |
| tinycss2           | 1.2.1        | BSD License                                             |
| toolz              | 0.12.0       | BSD License                                             |
| webencodings       | 0.5.1        | BSD License                                             |
| colorcet           | 3.0.1        | CC-BY License                                           |
| CairoSVG           | 2.7.1        | GNU Lesser General Public License v3 or later (LGPLv3+) |
| Cartopy            | 0.22.0       | GNU Lesser General Public License v3 or later (LGPLv3+) |
| PyQt5              | 5.15.10      | GPL v3                                                  |
| Pillow             | 10.1.0       | Historical Permission Notice and Disclaimer (HPND)      |
| PyQt5-Qt5          | 5.15.2       | LGPL v3                                                 |
| pyshp              | 2.3.1        | MIT                                                     |
| snuggs             | 1.4.7        | MIT                                                     |
| PyYAML             | 6.0.1        | MIT License                                             |
| QtPy               | 2.4.1        | MIT License                                             |
| attrs              | 23.1.0       | MIT License                                             |
| cffi               | 1.16.0       | MIT License                                             |
| cftime             | 1.6.3        | MIT License                                             |
| charset-normalizer | 3.3.2        | MIT License                                             |
| fonttools          | 4.47.0       | MIT License                                             |
| netCDF4            | 1.6.5        | MIT License                                             |
| pyparsing          | 3.1.1        | MIT License                                             |
| pyproj             | 3.6.1        | MIT License                                             |
| pytz               | 2023.3.post1 | MIT License                                             |
| six                | 1.16.0       | MIT License                                             |
| urllib3            | 2.1.0        | MIT License                                             |
| zipp               | 3.17.0       | MIT License                                             |
| certifi            | 2023.11.17   | Mozilla Public License 2.0 (MPL 2.0)                    |
| defusedxml         | 0.7.1        | Python Software Foundation License                      |
| matplotlib         | 3.8.2        | Python Software Foundation License                      |
| PyQt5-sip          | 12.13.0      | SIP                                                     |
