### 2023-12-21 62ce4a8
fix issues with contour-plot unittest
