### 2023-12-21 #138 @raphaelquast
@jhkennedy I had a look at the error you've encountered in the unittests (https://github.com/pyOpenSci/software-submission/issues/138#issuecomment-1858584307):

This is caused by the use of matplotlibs `ax.clabel` in the test to add labels to a contour-plot. The problem is the following:
- Using this function re-creates the artist internally which results in an artist that is no longer properly managed by EOmaps
   - the old artist is not properly deleted (e.g. resulting in an artist that is no longer attached to the figure)
   - the new artist is not put on the associated layer of the initial artist

Since this issue originates from using a non-EOmaps method, my solution for now is to simply omit this in the tests
(and also add a catch for artists that no longer exist in the figure)
This is done in: https://github.com/raphaelquast/EOmaps/pull/205/commits/62ce4a8d46a8c74adaaf971f489af6e02e8e64b3

In addition, I opened a dedicated issue to summarize the problems and hopefully implement a solution for adding contour-labels in the future: https://github.com/raphaelquast/EOmaps/issues/218

@banesullivan I hope it is OK for this review to postpone the development of a contour-level feature for now.
