### 2023-12-21 6b0717f
update contribution guides - add more comprehensive quick-info to CONTRIBUTING.md - update contribution guide in the docs to include testing etc. - rename `contribute_env.yml` to `environment.yml` and move to main dir
