+++ CONTRIBUTING.md
@@ -11,8 +11,6 @@ We greatly appreciate any ideas how we can improve the code and documentation.
 
 EOmaps aims to be a feature-rich, performant and user-friendly tool for geographic data visualization and analysis. We highly welcome pull-requests for bug-fixes, improvements or entirely new features to improve the package!
 
-A detailed introduction on how to setup a development environment to contribute code to EOmaps is provided in the [Contribution Guide](https://eomaps.readthedocs.io/en/latest/contribute.html) of the documentation.
-
 If you got some questions or need help to get started, don't hesitate to get in touch by opening a new [issue](https://github.com/raphaelquast/EOmaps/issues), or [discussion](https://github.com/raphaelquast/EOmaps/discussions) on GitHub!
 
 ## General infos
@@ -31,83 +29,54 @@ If you got some questions or need help to get started, don't hesitate to get in 
 
 ## Development Practices
 
-This section provides a guide to how we conduct development in the EOmaps repository. Please follow the practices outlined here when contributing directly to this repository.
+This section provides a **quick overview** how we conduct development in the EOmaps repository.
+
+**More detailed instructions** on how to setup a development environment and contribute code to EOmaps are provided in the [Contribution Guide](https://eomaps.readthedocs.io/en/latest/contribute.html) of the documentation!
+
 
 ### Testing
 
 After making changes, please test changes locally before creating a pull request. The following tests will be executed after any commit or pull request, so we ask that you perform the following sequence locally to track down any new issues from your changes.
 
-To run our comprehensive suite of unit tests, install all the dependencies listed in `tests/test_env.yml`:
-
-```bash
-# Install dependencies (to be filled in)
-```
-
-Then, if you have everything installed, you can run the various test suites.
-
-#### Unit Testing
-
 Run the primary test suite and generate coverage report:
 
-.. code:: bash
-
-   python -m pytest -v --cov eomaps
+```bash
+python -m pytest -v --cov eomaps
+```
 
 Unit testing can take some time, if you wish to speed it up, set the
-number of processors with the ``-n`` flag. This uses ``pytest-xdist`` to
-leverage multiple processes. Example usage:
+number of processors with the ``-n`` flag. This uses [pytest-xdist](https://github.com/pytest-dev/pytest-xdist) to leverage multiple processes. Example usage:
 
-.. code:: bash
+```bash
+python -m pytest -n <NUMCORE> --cov eomaps
+```
 
-   python -m pytest -n <NUMCORE> --cov eomaps
+> NOTE: During the tests, a lot of figures will be created and destroyed!
 
-#### Style Checking
+### Style Checking
 
-To ensure your code meets minimum code styling standards, run::
+To ensure your code meets minimum code styling standards, run:
 
-  pip install pre-commit
-  pre-commit run --all-files
-
-If you have issues related to ``setuptools`` when installing ``pre-commit``, see
-`pre-commit Issue #2178 comment <https://github.com/pre-commit/pre-commit/issues/2178#issuecomment-1002163763>`_
-for a potential resolution.
+```bash
+pip install pre-commit
+pre-commit run --all-files
+```
 
 You can also install this as a pre-commit hook by running::
 
-  pre-commit install
+```bash
+pre-commit install
+```
 
-This way, it's not possible for you to push code that fails the style
-checks. For example, each commit automatically checks that you meet the style
-requirements::
+This way, pre-commits will run automatically before each commit to ensure that you do not push code that fails the style checks.
 
-  $ pre-commit install
-  $ git commit -m "added my cool feature"
-  check python ast.........................................................Passed
-  check for merge conflicts................................................Passed
-  fix end of files.........................................................Passed
-  trim trailing whitespace.................................................Passed
[…]
