### 2023-12-21 2aeb6a3
minor - fix unexpected indent in docstring
