### 2023-12-22 64eb258
add entry point to pyproject.toml
