### 2023-12-30 2ffa9c3
run unittests on python 3.11 and 3.12
