### 2023-12-30 89f06b2
run unittests also on windows
