### 2023-12-30 45eef4c
omit python 3.12 for now
