+++ tests/test_env.yml
@@ -11,6 +11,7 @@ dependencies:
   - descartes
   - mapclassify
   - pyproj
+  - cartopy
   # --------------for data-shading
   - datashader
   # --------------for GeoTIFF and NetCDF files
@@ -31,6 +32,4 @@ dependencies:
   - docutils
   - jupyter
   - nbformat
-  - pip
-  - pip:
-    - cartopy
+
