+++ tests/test_env.yml
@@ -32,4 +32,3 @@ dependencies:
   - docutils
   - jupyter
   - nbformat
-
