### 2023-12-30 b6abc32
run pytest with -v (e.g. verbose) flag
