+++ .github/workflows/testMaps.yml
@@ -47,7 +47,7 @@ jobs:
         shell: bash -l {0}
         run: |
           pip install -e .[all]
-          python -m pytest --cov=eomaps --cov-report=xml
+          python -m pytest -v --cov=eomaps --cov-report=xml
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
         with:
