### 2023-12-30 439b724
make sure notebook-tests don't block the terminal
