+++ .github/workflows/testMaps.yml
@@ -23,7 +23,6 @@ jobs:
 
       - uses: mamba-org/setup-micromamba@v1
         with:
-          micromamba-version: '1.3.1-0'
           environment-file: tests/test_env.yml
           init-shell: >-
             bash
