+++ .github/workflows/testMaps.yml
@@ -11,9 +11,9 @@ jobs:
     strategy:
       matrix:
         # set operating systems to test
-        os: [windows-latest, ubuntu-latest]
+        os: [ubuntu-latest]
         # set python versions to test
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     name: test_Maps ${{ matrix.os }}  ${{ matrix.python-version }}
     steps:
