### 2024-01-02 bd88c28
add info on optional dependency groups to the docs
