### 2024-01-02 raphaelquast/EOmaps#211 @raphaelquast
The pending PR for EOmaps v8.0 (#205) now includes the following changes:

- minimal python version is now explicitly stated in `pyproject.toml` as `>3.8`
- unittests run on `3.8 - 3.11`
  - support for `python 3.12` is currently waiting for the 3.12 support PR from `numba` (required by `datashader`) which is about to be merged soon
- conda-forge recipe will be updated accordingly once `EOmaps v8.0` is released
