### 2024-01-02 raphaelquast/EOmaps#210 @raphaelquast
#216 is now merged into the `dev` branch!
EOmaps v8.0 (see #205) will incorporate the following changes:
- switch from `setup.py` to `pyproject.toml`
- minimal supported python version is `v3.8`
- version is now single-sourced to `pyproject.toml` file (`_version.py` is removed)
- optional dependency groups are implemented for `pip` install
