### 2024-01-02 raphaelquast/EOmaps#209 @raphaelquast
Just noticed that cartopy is also in the process of switching from LGPL to BSD-3: https://github.com/SciTools/cartopy/pull/2285.

This stripps the question for the EOmaps license down to using either `MIT` or also `BSD-3`.
@banesullivan do you have any comments concerning which of the 2 licenses is more suitable?
