+++ docs/api.rst
@@ -1,99 +1,0 @@
-🌍 Users Guide
-===============
-
-The following chapters provide comprehensive overview on the capabilities of EOmaps.
-
-|  :doc:`api_basics`
-|  The basic concepts how to create beautiful interactive maps.
-
-|  :doc:`api_data_visualization`
-|  How to visualize datasets (plot-shapes, classification, colorbars etc.).
-
-.. raw:: html
-
-    <hr style="height:1px;border:none;color:darkblue;background-color:darkblue;">
-
-|  :doc:`api_companion_widget`
-|  The graphical user interface that ships with EOmaps!
-
-|  :doc:`api_layout_editor`
-|  Interactively re-arrange and re-size axes of a figure.
-
-.. raw:: html
-
-    <hr style="height:1px;border:none;color:darkblue;background-color:darkblue;">
-
-|  :doc:`api_inset_maps`
-|  Create zoomed-in views on specific regions of a map.
-
-|  :doc:`api_callbacks`
-|  Turn your maps into interactive data-analysis widgets.
-
-.. raw:: html
-
-    <hr style="height:1px;border:none;color:darkblue;background-color:darkblue;">
-
-|  :doc:`api_naturalearth_features`
-|  Add basic map features (coastlines, ocean-coloring etc. ) to the map.
-
-|  :doc:`api_webmaps`
-|  Add imagery provided by WebMap services (ts, wms, wmts, xyz) to the map.
-
-|  :doc:`api_vector_data`
-|  Add vector geometries to the map.
-
-|  :doc:`api_annotations_markers_etc`
-|  Add markers, annotations, lines, logos etc. to the map.
-
-|  :doc:`api_scalebar`
-|  Add a scalebar to the map.
-
-|  :doc:`api_compass`
-|  Add a compass (or North Arrow) to the map.
-
-|  :doc:`api_gridlines`
-|  Add grid-lines (and optionally grid-labels) to the map.
-
-.. raw:: html
-
-    <hr style="height:1px;border:none;color:darkblue;background-color:darkblue;">
-
-|  :doc:`api_draw`
-|  Interactively draw geometries on a map and export them as shapefiles.
-
-|  :doc:`api_utils`
-|  A collection of utility widgets (layer-sliders, layer-selectors)
-
-|  :doc:`api_read_data`
-|  Read data from NetCDF, GeoTIFF or CSV files.
-
-|  :doc:`api_misc`
-|  Additional functions and properties that might come in handy.
-
-
-.. toctree::
-    :hidden:
-    :maxdepth: 3
-
-    api_basics
-    api_data_visualization
-
-    api_companion_widget
-    api_layout_editor
-
-    api_inset_maps
-    api_callbacks
-
-    api_naturalearth_features
-    api_webmaps
-    api_vector_data
-
-    api_annotations_markers_etc
-    api_scalebar
-    api_compass
-    api_gridlines
-
-    api_draw
-    api_utils
-    api_read_data
-    api_misc
