### 2024-01-02 e61ab45
avoid executing notebooks when compiling docs
