### 2024-01-03 raphaelquast/EOmaps#210 @banesullivan
Fantastic!
