### 2024-01-03 raphaelquast/EOmaps#209 @banesullivan
I think MIT and BSD-3 are both fantastic, permissive licenses and either would be a solid choice here.

In practice, I don't think there would be much difference between these two licenses here. In my mind, I typically equate the MIT and BSD 3 clause licenses, but since I'm not a lawyer/expert, I'd like to point you to this great answer: https://opensource.stackexchange.com/a/582

Typically, I go with the MIT license for everything, but in this case I think following the precedent set by the core upstream dependency of using BSD-3 would be a sound choice here.
