### 2024-01-03 raphaelquast/EOmaps#209 @raphaelquast
Thanks for your insights on this... they very much align with my opinion!

I'll start to implement relevant changes to switch to BSD-3 as soon as possible!

Since there have been no substantial contributions so far (only 1-liners and typos and most of them are no longer in the code),
I don't think it's required to get permission from the contributors to perform the license-change.
