### 2024-01-03 2dfac76
add info of major changes between v7.x and v8.x
