+++ docs/FAQ.rst
@@ -345,3 +345,17 @@ Important changes between major versions
     - ⚠️ The use of ``m.set_data_specs(...)`` is depreciated. Use ``m.set_data(...)`` instead!
 
     - Figure export routines have been completely re-worked (but should result in the exact same output as in v6.x)
+
+
+.. dropdown:: ⚙ From EOmaps v7.x to v8.x
+
+    - ⚠️ Some functions and classes have been re-named to better follow PEP8 naming conventions.
+      While this should not interfere with the public API, more extensive customizations might need to be
+      adjusted with respect to the new names.
+
+      If you encounter any problems, feel free to open an `issue <https://github.com/raphaelquast/EOmaps/issues>`_ , and I'll see what I can do!
+
+    - ⚠️ ``pip`` install has been updated to implement optional dependency groups
+      Have a look at the :ref:`installation` instructions for more details!
+
+    - ``setup.py`` and ``_version.py`` have been removed in favor of using a ``pyproject.toml`` file
