### 2024-01-03 4672d1d
update pyproject.toml with more metadata
