### 2024-01-03 c6945c4
fix usage of deprecated kwarg in docs
