### 2024-01-03 82a6f9b
add license info to all source files
