### 2024-01-03 raphaelquast/EOmaps#209 @raphaelquast
There is now a PR to implement the switch from GPL to BSD-3 (#220)
@banesullivan It would be awesome if you could have a look and let me know if you see any issues with the way I implemented the license there.
