### 2024-01-03 83228f6
make deprecation of m.set_data_specs final
