### 2024-01-03 fbc0fbe
make deprecation of InsetMaps.indicate_inset_extent final
