### 2024-01-03 218a564
make deprecation of m.add_wms.DLR_basemaps final
