+++ eomaps/scripts/open.py
@@ -207,9 +207,9 @@ def cli(crs=None, file=None, ne=None, wms=None, location=None, loglevel=None):
         elif wms in ["terrain_light"]:
             m.add_wms.S2_cloudless.add_layer.terrain_light_3857()
         elif wms in ["basemap"]:
-            m.add_wms.DLR_basemaps.add_layer.basemap()
+            m.add_wms.DLR.basemap.add_layer.basemap()
         elif wms in ["basemap_light"]:
-            m.add_wms.DLR_basemaps.add_layer.litemap()
+            m.add_wms.DLR.basemap.add_layer.litemap()
         elif wms in ["s1_vv"]:
             m.add_wms.S1GBM.add_layer.vv()
 
