### 2024-01-03 6c80db2
add info on removed (previously deprecated) methods to docs
