### 2024-01-05 20d56f2
update docs env - use more recent versions of docs-build dependencies - install only minimal eomaps dependencies required for import
