### 2024-01-07 75a386e
fix docstring for DLR webmap service
