### 2024-01-07 674f359
update NaturalEarth feature implementation - apply better naming conventions - make namespaces accessible to sphinx
