+++ eomaps/ne_features.py
@@ -12,12 +12,457 @@ from .helpers import register_modules
 _log = logging.getLogger(__name__)
 
 
-def combdoc(*args):
+# ------ Load available features from NE_features.json file.
+
+_NE_features_path = Path(__file__).parent / "NE_features.json"
+
+try:
+    with open(_NE_features_path, "r") as file:
+        _NE_features = json.load(file)
+
+        _NE_features_all = dict()
+        for scale, scale_items in _NE_features.items():
+            for category, category_items in scale_items.items():
+                _NE_features_all.setdefault(category, set()).update(category_items)
+
+
+except Exception:
+    _log.error(
+        "EOmaps: Could not load available NaturalEarth features from\n"
+        f"{_NE_features_path}",
+        exc_info=_log.getEffectiveLevel() <= logging.DEBUG,
+    )
+    _NE_features = dict()
+    _NE_features_all = dict()
+
+# ---------------------
+
+
+def _combdoc(*args):
     """Dedent and combine strings."""
     return "\n".join(dedent(str(i)) for i in args)
 
 
-class NaturalEarth_presets:
+class _Category:
+    """
+    Interface to features provided by NaturalEarth.
+
+    (see https://www.naturalearthdata.com)
+
+    The features are grouped into the categories "cultural" and "physical"
+    and available at 3 different scales:
+
+    - 10 : Large-scale data (1:10m)
+    - 50 : Medium-scale data (1:50m)
+    - 110 : Small-scale data (1:110m)
+
+    If you use scale="auto", the appropriate scale of the feature will be determined
+    based on the map-extent.
+
+    >>> m.add_feature.< category >.< feature-name >(scale=10, ... style-kwargs ...)
+
+    Examples
+    --------
+    - add black (coarse resolution) coastlines
+
+      >>> from eomaps import Maps
+      >>> m = Maps()
+      >>> m.add_feature.physical.coastline(scale=110, fc="none", ec="k")
+
+    - color all land red with 50% transparency and automatically determine the
+      appropriate scale if you zoom the map
+
+      >>> from eomaps import Maps
+      >>> m = Maps()
+      >>> m.add_feature.physical.land(scale="auto", fc="r", alpha=0.5)
+
+    - fetch features as geopandas.GeoDataFrame
+      (to color all countries with respect to the area)
+
+      >>> from eomaps import Maps
+      >>> m = Maps()
+      >>> countries = m.add_feature.cultural.admin_0_countries.get_gdf(scale=10)
+      >>> countries["area_rank"] = countries.area.rank()
+      >>> m.add_gdf(countries, column="area_rank")
+
+    """
+
+    _category = "???"
+
+    def __repr__(self):
+        return f"EOmaps interface for {self._category} " + "NaturalEarth features"
+
+    @classmethod
+    def _setup(cls, category):
+        # this is used to setup the namespace so that it is accessible for sphinx
+
+        for i in _NE_features_all[category]:
+            # we need to init new classes to avoid issues with immutuable attributes
+            setattr(cls, i, type(i, (_Feature,), {}))  # (category, i))
+
+        cls.__doc__ = _combdoc(
+            f"NaturalEarth feature interface for: **{category}**.\n",
+            _Category.__doc__,
+        )
+
+    def _set_map(self, m):
+        for feature_name in filter(lambda x: not x.startswith("_"), dir(self)):
+            try:
+                feature = _Feature(self._category, feature_name)
+                feature._set_map(m)
+                setattr(self, feature_name, feature)
+            except Exception:
+                _log.error(f"EOmaps: unable to set map for feature {feature}")
+
+
+class _Feature:
+    """
+    Natural Earth Feature.
+
[…]
