### 2024-01-07 13c82c2
make namespaces accessible to sphinx
