+++ docs/api/reference.rst
@@ -1,0 +1,5 @@
+
+.. toctree::
+    :maxdepth: 1
+
+    api_Maps
