+++ docs/_templates/obj_with_attributes_no_toc.rst
@@ -18,4 +18,3 @@
 {%- endfor %}
 {% endif %}
 {% endblock %}
-
