### 2024-01-07 1939e66
fix generation of stub-files for documents other than the api-docs
