### 2024-01-08 7e00d4d
fix issues for move-callbacks shared between multiple Maps
- avoid calling update on callbacks directly (update only from parent)
- remove "update_on_trigger" arg and simply check if move-callbacks are
  attached before triggering an update
