### 2024-01-08 5ca9a20
update autodocs... unify toctree for generated docs
