+++ docs/full_reference.rst
@@ -1,0 +1,9 @@
+:orphan:
+
+.. autosummary::
+   :toctree: generated
+   :nosignatures:
+   :template: custom-module-template.rst
+   :recursive:
+
+   eomaps
