### 2024-01-11 0295525
improve implementation of cb accessor
