+++ docs/conf.py
@@ -1,66 +1,18 @@
 # Configuration file for the Sphinx documentation builder.
 import sys, os
-from pathlib import Path
+from eomaps import Maps
 
 sys.path.insert(0, os.path.abspath(os.path.join("..", "..")))
 sys.path.insert(0, os.path.abspath(".."))
+sys.path.insert(0, os.path.abspath("."))
 
-from eomaps import Maps
-
-
-def make_feature_toctree_file():
-    preset = (f for f in dir(Maps.add_feature.preset) if not f.startswith("_"))
-    physical = (f for f in dir(Maps.add_feature.physical) if not f.startswith("_"))
-    cultural = (f for f in dir(Maps.add_feature.cultural) if not f.startswith("_"))
-    shapes = (f for f in dir(Maps.set_shape) if not f.startswith("_"))
-    draw = (f for f in dir(Maps.draw) if not f.startswith("_"))
-    wms = (f for f in dir(Maps.add_wms) if not f.startswith("_"))
-    files = (f for f in dir(Maps.from_file) if not f.startswith("_"))
-    read_files = (f for f in dir(Maps.read_file) if not f.startswith("_"))
-
-    s = (
-        ":orphan:\n"
-        ".. currentmodule:: eomaps.eomaps\n"
-        ".. autosummary::\n"
-        "    :toctree: generated\n"
-        "    :nosignatures:\n"
-        "    :template: obj_with_attributes_no_toc.rst\n\n"
-        "    Maps.add_wms\n"
-        "    Maps.set_shape\n"
-        "    Maps.config\n"
-        "    Maps.draw\n"
-        "    Maps.add_feature\n"
-        "    Maps.add_feature.preset\n"
-        "    Maps.add_feature.physical\n"
-        "    Maps.add_feature.cultural\n"
-        + "\n".join(
-            [
-                "\n".join([f"    Maps.add_feature.preset.{f}" for f in preset]),
-                "\n".join([f"    Maps.add_feature.physical.{f}" for f in physical]),
-                "\n".join(
-                    [f"    Maps.add_feature.physical.{f}.get_gdf" for f in physical]
-                ),
-                "\n".join([f"    Maps.add_feature.cultural.{f}" for f in cultural]),
-                "\n".join(
-                    [f"    Maps.add_feature.cultural.{f}.get_gdf" for f in cultural]
-                ),
-                "\n".join([f"    Maps.set_shape.{f}" for f in shapes]),
-                "\n".join([f"    Maps.draw.{f}" for f in draw]),
-                "\n".join([f"    Maps.add_wms.{f}" for f in wms]),
-                "\n".join([f"    Maps.from_file.{f}" for f in files]),
-                "\n".join([f"    Maps.read_file.{f}" for f in read_files]),
-            ]
-        )
-    )
-
-    with open("autodoc_additional_props.rst", "w") as file:
-        file.write(s)
-
+from gen_autodoc_file import make_feature_toctree_file
 
 make_feature_toctree_file()
 
 
 def setup(app):
+
     app.add_css_file("custom_css.css")
 
     # need to assign the names here, otherwise autodoc won't document these classes,
@@ -73,6 +25,12 @@ def setup(app):
     Maps.set_shape.__name__ = "set_shape"
     Maps.draw.__name__ = "draw"
     Maps.add_wms.__name__ = "add_wms"
+    Maps.util.__name__ = "util"
+    Maps.cb.__name__ = "cb"
+    Maps.cb.click.__name__ = "click"
+    Maps.cb.pick.__name__ = "pick"
+    Maps.cb.keypress.__name__ = "keypress"
[…]
