+++ docs/api/reference.rst
@@ -1,5 +1,115 @@
+======
+eomaps
+======
 
-.. toctree::
-    :maxdepth: 1
+.. currentmodule:: eomaps.eomaps
 
-    api_Maps
+.. card:: The Maps object
+    :link: ../generated/eomaps.eomaps.Maps
+    :link-type: doc
+
+    :py:class:`Maps` objects are used as the primary access point to create, edit and interact with maps.
+
+    .. autosummary::
+        :toctree: ../generated
+        :template: custom-class-template.rst
+        :nosignatures:
+
+        Maps
+
+
+Feature objects
+===============
+
+.. card::
+    :shadow: none
+
+    .. card:: ColorBar
+        :link: ../generated/eomaps.colorbar.ColorBar
+        :link-type: doc
+        :margin: 0
+
+        .. currentmodule:: eomaps.colorbar
+
+        .. autosummary::
+            :toctree: ../generated
+            :template: custom-class-template.rst
+            :nosignatures:
+
+            ColorBar
+
+    +++
+
+    .. currentmodule:: eomaps.eomaps
+
+    See :py:meth:`Maps.add_colorbar` on how to add a colorbar to a map!
+
+.. card::
+    :shadow: none
+
+    .. card:: GridLines
+        :link: ../generated/eomaps.grid.GridLines
+        :link-type: doc
+        :margin: 0
+
+        .. currentmodule:: eomaps.grid
+
+        .. autosummary::
+            :toctree: ../generated
+            :template: custom-class-template.rst
+            :nosignatures:
+
+            GridLines
+
+    +++
+
+    .. currentmodule:: eomaps.eomaps
+
+    See :py:meth:`Maps.add_gridlines` on how to add grid-lines to a map!
+
+
+.. card::
+    :shadow: none
+
+    .. card:: Compass
+        :link: ../generated/eomaps.compass.Compass
+        :link-type: doc
+        :margin: 0
+
+        .. currentmodule:: eomaps.compass
+
+        .. autosummary::
+            :toctree: ../generated
+            :template: custom-class-template.rst
+            :nosignatures:
+
+            Compass
+
+    +++
+
+    .. currentmodule:: eomaps.eomaps
+
+    See :py:meth:`Maps.add_compass` on how to add a compass (or North-arrow) to a map!
+
+.. card::
+    :shadow: none
+
+    .. card:: ScaleBar
+        :link: ../generated/eomaps.scalebar.ScaleBar
+        :link-type: doc
+        :margin: 0
+
+        .. currentmodule:: eomaps.scalebar
+
+        .. autosummary::
+            :toctree: ../generated
+            :template: custom-class-template.rst
+            :nosignatures:
+
+            ScaleBar
+
+    +++
+
+    .. currentmodule:: eomaps.eomaps
+
+    See :py:meth:`Maps.add_scalebar` on how to add a scalebar to a map!

[…]
