### 2024-01-11 6529fe7
include package data in pyproject.toml
