### 2024-01-11 81bcbfe
fix including package data in pyproject.toml
