### 2024-01-11 #138 @raphaelquast
Response to comment form @jhkennedy

> The [API reference](https://eomaps.readthedocs.io/en/latest/generated/eomaps.html) is hard to navigate
>
> - it's not in alphabetical order, or is at least missing headings, and the add_feature and add_wms methods were hard to find
> - add_feature.preset isn't documented and seems to be a NaturalEarth_features class. In fact, the add_feature reference doesn't seem to match at all
> - IDE tab completion also doesn't discover NE categories; hard to know what I'm able to access
> - I ended up just opening NE_features.json, reformatting it so it's not on a single line, and looking at what's available there.

I implemented major changes to the API docs to reflect the actual public API and not the (internal) module-structure...
I hope you'll find them easier to navigate now!
You can have a look at the current state here: https://eomaps.readthedocs.io/en/dev/api/reference.html
I'm happy about any additional comments how to improve the docs!
