+++ docs/api_draw.rst
@@ -40,14 +40,14 @@ Starting with EOmaps v5.0 it is possible to draw simple shapes on the map using 
 
 
 
-.. currentmodule:: eomaps.draw
+.. currentmodule:: eomaps.eomaps.Maps.draw
 
 .. autosummary::
     :nosignatures:
 
-    ShapeDrawer.new_drawer
-    ShapeDrawer.rectangle
-    ShapeDrawer.circle
-    ShapeDrawer.polygon
-    ShapeDrawer.save_shapes
-    ShapeDrawer.remove_last_shape
+    new_drawer
+    rectangle
+    circle
+    polygon
+    save_shapes
+    remove_last_shape
