+++ docs/index.rst
@@ -12,7 +12,7 @@
     <p>
 
 
-Here you can find detailed explanations on all the functionalities of EOmaps.
+Here you can find detailed explanations on all the features and functionalities of EOmaps.
 
 .. admonition:: Interested in contributing to EOmaps?
 
@@ -37,44 +37,58 @@ Want to visualize some data? Have a look at the :doc:`api_data_visualization` se
 EOmaps provides a variety of plot-shapes so you can select a shape that suits the structure, size and spatial representativeness of your data:
 
 
-.. raw:: html
+.. grid:: 2 2 5 5
+    :gutter: 1
 
-    <table width="100%" style="font-size:90%; table-layout:fixed; overflow:scroll;">
-    <tr>
-    <th width="20%">Ellipses</th>
-    <th width="20%">Rectangles</th>
-    <th width="20%">Geodesic Circles</th>
-    <th width="20%">Raster</th>
-    <th width="20%">Scatter Points</th>
-    </tr>
-    <tr>
-    <th width="20%"><a href=api_data_visualization.html#ellipses><img src="_static/shape_imgs/ellipses.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#rectangles><img src="_static/shape_imgs/rectangles.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#geodesic-circles><img src="_static/shape_imgs/geod_circles.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#raster><img src="_static/shape_imgs/raster.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#scatter-points><img src="_static/shape_imgs/scatter_points.png"/></a></th>
-    </tr>
-    <tr>
-    <th width="20%">Contour</th>
-    <th width="20%">Voronoi Diagram</th>
-    <th width="20%">Delaunay Triangulation</th>
-    <th width="20%">Shade Raster</th>
-    <th width="20%">Shade Points</th>
-    </tr>
-    <tr>
-    <th width="20%"><a href=api_data_visualization.html#contour><img src="_static/shape_imgs/contour_filled.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#voronoi-diagram><img src="_static/shape_imgs/voronoi_diagram.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#delaunay-triangulation><img src="_static/shape_imgs/delaunay_triangulation.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#shade-raster><img src="_static/shape_imgs/shade_raster.png"/></a></th>
-    <th width="20%"><a href=api_data_visualization.html#shade-points><img src="_static/shape_imgs/shade_points.png"/></a></th>
-    </tr>
-    </table>
+    .. grid-item-card:: Ellipses
+        :img-bottom: _static/shape_imgs/ellipses.png
+        :link: shp_ellipses
+        :link-type: ref
 
+    .. grid-item-card:: Rectangles
+        :img-bottom: _static/shape_imgs/rectangles.png
+        :link: shp_rectangles
+        :link-type: ref
 
-.. raw:: html
+    .. grid-item-card:: Geodesic Circles
+        :img-bottom: _static/shape_imgs/geod_circles.png
+        :link: shp_geod_circles
+        :link-type: ref
 
-    </p>
+    .. grid-item-card:: raster
+        :img-bottom: _static/shape_imgs/raster.png
+        :link: shp_raster
+        :link-type: ref
 
+    .. grid-item-card:: Scatter Points
+        :img-bottom: _static/shape_imgs/scatter_points.png
+        :link: shp_scatter
+        :link-type: ref
+
[…]
