### 2024-01-13 902e841
address final docs-warnings - implement custom rc role to address warnings from inherited docstrings - link to matplotlib docs via intersphinx to resolve external labels
