+++ docs/FAQ.rst
@@ -16,8 +16,15 @@ Need some help to setup python for EOmaps?
    quickstart_quide
 
 
-Have a look at the :ref:`quickstart_guide` to learn how to setup a python environment that can be
+Have a look at the :ref:`Quickstart Guide <quickstart_guide>` to learn how to setup a python environment that can be
 used to create EOmaps maps!
+
+
+Interested in contributing to EOmaps?
+*************************************
+
+Have a look at the :ref:`Contribution Guide <contribute>` on how to setup a development environment to start contributing to EOmaps!
+Any contributions are welcome!
 
 
 .. _configuring_the_editor:
@@ -187,6 +194,26 @@ The best way to record interactions on a EOmaps map is with the
 free and open source `ScreenToGif <https://www.screentogif.com/>`_ software.
 
 All animated gifs in this documentation have been created with this awesome piece of software.
+
+
+LICENSING and redistribution
+****************************
+
+Information is provided WITHOUT WARRANTY FOR CORRECTNESS!
+
+Since v8.0, the source-code of EOmaps is licensed under a `BSD 3-clause license <https://choosealicense.com/licenses/bsd-3-clause/>`_.
+
+However, it must be notet that EOmaps has a number of required and optional dependencies whose licenses must be taken into account when packaging
+and redistributing works that build on top of EOmaps. While most dependencies are either **BSD** or **MIT** licensed, there are some dependencies 
+that might require additional considerations.
+
+Most notably (start of 2024), the **required dependency** `cartopy <https://github.com/SciTools/cartopy>`_ is licensed under **LGPLv3+** (however they are in the process of `re-licensing to BSD-3-clause <https://github.com/SciTools/cartopy/pull/2285>`_) and the **optional** `Qt <https://www.qt.io/licensing>`_ GUI framework is available via commertial and open-source (**LGPLv3+**) licenses and the used python-bindings (`PyQt5 <https://github.com/pyqt/python-qt5>`_) are licensed under **GPL v3**.
+
+
+.. tip::
+
+    If you want to get a quick overview of the licenses of an existing pyhton environment, I recommend having a look at the `pip-licenses <https://github.com/raimon49/pip-licenses>`_ package!
+
 
 Important changes between major versions
 ****************************************
