### 2024-01-14 raphaelquast/EOmaps#209 @raphaelquast
@banesullivan The change to BSD 3-clause has now been merged into the dev-branch to be included in the upcoming v8.0 release!

In addition I also added a section on Licensing to the FAQ to clarify that some upstream dependencies
have more restrictive licensing that might need to be considered more carefully (https://eomaps.readthedocs.io/en/dev/FAQ.html#licensing-and-redistribution), e.g.:

- **required** `cartopy` dependency is still LGPL (but will be BSD soon)
- **optional** `Qt / PyQT` gui framework bindings offers flexible licensing but requires GPL/LGPL for open-source projects

The good thing is that once the switch to BSD for `cartopy` is released, all core dependencies of EOmaps are BSD-compatible.

In my opinion this should address any Licensing concerns for EOmaps mentioned in this issue.

Please let me know if you see any further issues that should be addressed. (otherwise, I think this can be closed)
