### 2024-01-14 2884502
remove cairosvg as dependency (allow usage for legends if installed)
