### 2024-01-14 e553a1a
remove obsolete dependencies from test-env
