+++ tests/test_env.yml
@@ -3,7 +3,6 @@ channels:
   - conda-forge
 
 dependencies:
-  - rtree
   - numpy
   - scipy
   - pandas
@@ -23,8 +22,6 @@ dependencies:
   # --------------for WebMaps
   - owslib
   - requests
-  - xmltodict
-  - cairosvg
   # --------------for testing
   - nbformat     # to parse Jupyter Notebooks
   - coveralls
