### 2024-01-14 983283c
updates for examples galery - examples are now located in "docs/examples" - unittests now automatically test all example-files - naming-convention is improved (avoid numbering and use descriptive names)
