### 2024-01-14 54530ba
update docs: make layer-intro image larger
