+++ docs/api_annotations_markers_etc.rst
@@ -50,62 +50,66 @@ To dynamically add annotations if you click on the map, checkout the associated 
       :width: 50%
 
 
-.. table::
-    :widths: 50 50
-    :align: center
 
-    +-----------------------------------------------------------------------------------+---------------------------------------------+
-    | .. code-block:: python                                                            | .. image:: _static/minigifs/annotations.png |
-    |     :name: test_add_annotations                                                   |   :align: center                            |
-    |                                                                                   |                                             |
-    |     from eomaps import Maps                                                       | |img_minsize|                               |
-    |     import numpy as np                                                            |                                             |
-    |     x, y = np.mgrid[-45:45, 20:60]                                                |                                             |
-    |                                                                                   |                                             |
-    |     m = Maps(crs=4326)                                                            |                                             |
-    |     m.set_data(x+y, x, y)                                                         |                                             |
-    |     m.add_feature.preset.coastline(ec="k", lw=.75)                                |                                             |
-    |     m.plot_map()                                                                  |                                             |
-    |                                                                                   |                                             |
-    |     # annotate any point in the dataset via the data-index                        |                                             |
-    |     m.add_annotation(ID=345)                                                      |                                             |
-    |     # annotate an arbitrary position (in the plot-crs)                            |                                             |
-    |     m.add_annotation(                                                             |                                             |
-    |         xy=(20,25), text="A formula:\n $z=\sqrt{x^2+y^2}$",                       |                                             |
-    |         fontweight="bold", bbox=dict(fc=".6", ec="none", pad=0.2))                |                                             |
-    |     # annotate coordinates defined in arbitrary crs                               |                                             |
-    |     m.add_annotation(                                                             |                                             |
-    |         xy=(2873921, 6527868), xy_crs=3857, xytext=(5,5),                         |                                             |
-    |         text="A location defined \nin epsg 3857", fontsize=8,                     |                                             |
-    |         rotation=-45, bbox=dict(fc="skyblue", ec="k", ls="--", pad=0.2))          |                                             |
-    |                                                                                   |                                             |
-    |     # functions can be used for more complex text                                 |                                             |
-    |     def text(m, ID, val, pos, ind):                                               |                                             |
-    |         return f"lon={pos[0]}\nlat={pos[1]}"                                      |                                             |
-    |                                                                                   |                                             |
-    |     props = dict(xy=(-1.5, 38.45), text=text,                                     |                                             |
-    |                  arrowprops=dict(arrowstyle="-|>", fc="fuchsia",                  |                                             |
-    |                                  mutation_scale=15))                              |                                             |
-    |                                                                                   |                                             |
-    |     m.add_annotation(**props, xytext=(20, 20), color="darkred")                   |                                             |
-    |     m.add_annotation(**props, xytext=(-60, 20), color="purple")                   |                                             |
[…]
