+++ docs/api_basics.rst
@@ -375,105 +375,142 @@ To position the map in a (virtual) grid, one of the following options are possib
   - ``index`` can also be a two-tuple specifying the (first, last)
     indices (1-based, and including last) of the map, e.g., ``Maps(ax=(3, 1, (1, 2)))`` makes a map that spans the upper 2/3 of the figure.
 
-.. table::
 
-    +----------------------------------------------------+------------------------------------+
-    | .. code-block:: python                             | .. image:: _static/grids/grid1.png |
-    |     :name: test_gridpos_1                          |   :align: center                   |
-    |                                                    |                                    |
-    |     from eomaps import Maps                        | |img_minsize|                      |
-    |     # ----- initialize a figure with an EOmaps map |                                    |
-    |     # position = item 1 of a 2x1 grid              |                                    |
-    |     m = Maps(ax=(2, 1, 1))                         |                                    |
-    |     # ----- add a normal matplotlib axes           |                                    |
-    |     # position = item 2 of a 2x1 grid              |                                    |
-    |     ax = m.f.add_subplot(2, 1, 2)                  |                                    |
-    +----------------------------------------------------+------------------------------------+
+.. grid:: 1 1 2 2
 
-.. table::
+    .. grid-item::
 
-    +----------------------------------------------------+------------------------------------+
-    | .. code-block:: python                             | .. image:: _static/grids/grid2.png |
-    |     :name: test_gridpos_2                          |   :align: center                   |
-    |                                                    |                                    |
-    |     from eomaps import Maps                        | |img_minsize|                      |
-    |     # ----- initialize a figure with an EOmaps map |                                    |
-    |     # position = item 1 of a 2x2 grid              |                                    |
-    |     m = Maps(ax=(2, 2, 1))                         |                                    |
-    |     # ----- add another Map to the same figure     |                                    |
-    |     # position = item 3 of a 2x2 grid              |                                    |
-    |     m2 = m.new_map(ax=(2, 2, 3))                   |                                    |
-    |     # ----- add a normal matplotlib axes           |                                    |
-    |     # position = second item of a 1x2 grid         |                                    |
-    |     ax = m.f.add_subplot(1, 2, 2)                  |                                    |
-    +----------------------------------------------------+------------------------------------+
+         .. code-block:: python
+             :name: test_gridpos_1
 
-.. table::
+             from eomaps import Maps
+             # ----- initialize a figure with an EOmaps map
+             # position = item 1 of a 2x1 grid
+             m = Maps(ax=(2, 1, 1))
+             # ----- add a normal matplotlib axes
+             # position = item 2 of a 2x1 grid
+             ax = m.f.add_subplot(2, 1, 2)
 
-    +----------------------------------------------------+------------------------------------+
-    | .. code-block:: python                             | .. image:: _static/grids/grid3.png |
-    |     :name: test_gridpos_3                          |   :align: center                   |
-    |                                                    |                                    |
-    |     from eomaps import Maps                        | |img_minsize|                      |
-    |     # ----- initialize a figure with an EOmaps map |                                    |
-    |     # position = span 2 rows of a 3x1 grid         |                                    |
-    |     m = Maps(ax=(3, 1, (1, 2)))                    |                                    |
-    |     # ----- add a normal matplotlib axes           |                                    |
-    |     # position = item 3 of a 3x1 grid              |                                    |
-    |     ax = m.f.add_subplot(3, 1, 3)                  |                                    |
-    +----------------------------------------------------+------------------------------------+
+    .. grid-item-card::
+        :width: 50%
+        :img-background: _static/grids/grid1.png
+
+
+.. grid:: 1 1 2 2
+
+    .. grid-item::
+
[…]

[…]
