+++ docs/api_gridlines.rst
@@ -39,37 +39,39 @@ If no explicit grid-spacing is provided (e.g. ``d=None``), the grid is dynamical
     Maps.add_gridlines
 
 
-.. table::
-    :widths: 50 50
-    :align: center
+.. grid:: 1 1 2 2
 
-    +-------------------------------------------------------------------+-----------------------------------------+
-    | .. code-block:: python                                            | .. image:: _static/minigifs/grid_01.png |
-    |   :name: test_add_gridlines                                       |     :align: center                      |
-    |                                                                   |                                         |
-    |   from eomaps import Maps                                         | |img_minsize|                           |
-    |   m = Maps(Maps.CRS.Mollweide(), frameon=False)                   |                                         |
-    |   m.add_feature.preset.ocean()                                    |                                         |
-    |                                                                   |                                         |
-    |   # add gridlines with a fixed grid-spacing                       |                                         |
-    |   mg = m.new_layer("grid")                                        |                                         |
-    |   g0 = mg.add_gridlines(d=40, ec="orange", lw=3, zorder=2)        |                                         |
-    |   g1 = mg.add_gridlines(d=(10, 20), ec="orange", lw=.5, zorder=1) |                                         |
-    |                                                                   |                                         |
-    |   # add fine-grained gridlines in a specific area                 |                                         |
-    |   g2 = mg.add_gridlines(d=2, ec="darkred", lw=0.5, zorder=0,      |                                         |
-    |                         bounds=(-20, 20, -10, 30))                |                                         |
-    |   g3 = mg.add_gridlines(d=2, ec="b", lw=0.5, zorder=0,            |                                         |
-    |                         bounds=(60, 100, 30, 70))                 |                                         |
-    |                                                                   |                                         |
-    |   # add dedicated gridlines at specific coordinates               |                                         |
-    |   g4 = mg.add_gridlines(([-123, -112, -75], [35, 65]),            |                                         |
-    |                         ec="k", lw=2, ls="--", zorder=20,         |                                         |
-    |                         bounds=(-140, 20, -50, 70)                |                                         |
-    |                         )                                         |                                         |
-    |                                                                   |                                         |
-    |   m.show_layer(m.layer, "grid")                                   |                                         |
-    +-------------------------------------------------------------------+-----------------------------------------+
+    .. grid-item::
+
+         .. code-block:: python
+            :name: test_add_gridlines
+
+            from eomaps import Maps
+            m = Maps(Maps.CRS.Mollweide(), frameon=False)
+            m.add_feature.preset.ocean()
+
+            # add gridlines with a fixed grid-spacing
+            mg = m.new_layer("grid")
+            g0 = mg.add_gridlines(d=40, ec="orange", lw=3, zorder=2)
+            g1 = mg.add_gridlines(d=(10, 20), ec="orange", lw=.5, zorder=1)
+
+            # add fine-grained gridlines in a specific area
+            g2 = mg.add_gridlines(d=2, ec="darkred", lw=0.5, zorder=0,
+                                  bounds=(-20, 20, -10, 30))
+            g3 = mg.add_gridlines(d=2, ec="b", lw=0.5, zorder=0,
+                                  bounds=(60, 100, 30, 70))
+
+            # add dedicated gridlines at specific coordinates
+            g4 = mg.add_gridlines(([-123, -112, -75], [35, 65]),
+                                  ec="k", lw=2, ls="--", zorder=20,
+                                  bounds=(-140, 20, -50, 70)
+                                  )
+
+            m.show_layer(m.layer, "grid")
+
+    .. grid-item::
+
+        .. image:: _static/minigifs/grid_01.png
 
 
 .. currentmodule:: eomaps.grid
[…]
