+++ docs/api_data_visualization.rst
@@ -542,33 +542,35 @@ To assign a classification scheme to a :py:class:`Maps` object, use ``m.set_clas
     Maps.set_classify
 
 
-.. table::
-    :widths: 70 30
-    :align: center
+.. grid:: 1 1 2 2
 
-    +------------------------------------------------------------------+--------------------------------------------------+
-    | .. code-block:: python                                           | .. image:: _static/minigifs/classify_data_01.png |
-    |     :name: test_classify_data                                    |     :align: center                               |
-    |                                                                  |                                                  |
-    |     from eomaps import Maps                                      | |img_minsize|                                    |
-    |     import numpy as np                                           |                                                  |
-    |                                                                  |                                                  |
-    |     data = np.random.normal(0, 1, (50, 50))                      |                                                  |
-    |     x = np.linspace(-45, 45, 50)                                 |                                                  |
-    |     y = np.linspace(-45, 45, 50)                                 |                                                  |
-    |                                                                  |                                                  |
-    |     m = Maps(figsize=(4, 5))                                     |                                                  |
-    |     m.add_feature.preset.coastline(lw=2)                         |                                                  |
-    |     m.add_feature.preset.ocean(zorder=99, alpha=0.5)             |                                                  |
-    |     m.set_data(data, x, y)                                       |                                                  |
-    |     m.set_shape.ellipses()                                       |                                                  |
-    |     m.set_classify.StdMean(multiples=[-1.5, -.5, .5, 1.5])       |                                                  |
-    |     m.plot_map(vmin=-3, vmax=3)                                  |                                                  |
-    |                                                                  |                                                  |
-    |     cb = m.add_colorbar(pos=0.2, label="StdMean classification") |                                                  |
-    |     cb.tick_params(labelsize=7)                                  |                                                  |
-    +------------------------------------------------------------------+--------------------------------------------------+
+    .. grid-item::
 
+         .. code-block:: python
+            :name: test_classify_data
+
+            from eomaps import Maps
+            import numpy as np
+
+            data = np.random.normal(0, 1, (50, 50))
+            x = np.linspace(-45, 45, 50)
+            y = np.linspace(-45, 45, 50)
+
+            m = Maps(figsize=(4, 5))
+            m.add_feature.preset.coastline(lw=2)
+            m.add_feature.preset.ocean(zorder=99, alpha=0.5)
+            m.set_data(data, x, y)
+            m.set_shape.ellipses()
+            m.set_classify.StdMean(multiples=[-1.5, -.5, .5, 1.5])
+            m.plot_map(vmin=-3, vmax=3)
+
+            cb = m.add_colorbar(pos=0.2, label="StdMean classification")
+            cb.tick_params(labelsize=7)
+
+    .. grid-item::
+
+        .. image:: _static/minigifs/classify_data_01.png
+            :width: 75%
 
 
 Currently available classification-schemes are (see `mapclassify <https://github.com/pysal/mapclassify>`_ for details):
@@ -828,43 +830,46 @@ Once a dataset has been plotted, a colorbar with a colored histogram on top can 
     Colorbars are only visible if the layer at which the data was plotted is visible!
 
 
-.. table::
-    :widths: 60 40
-    :align: center
+.. grid:: 1 1 2 2
 
-    +-----------------------------------------------------------------+------------------------------------------+
-    | .. code-block:: python                                          | .. image:: _static/minigifs/colorbar.gif |
-    |     :name: test_colorbars                                       |     :align: center                       |
-    |                                                                 |                                          |
-    |     from eomaps import Maps                                     | |img_minsize|                            |
-    |     import numpy as np                                          |                                          |
-    |                                                                 |                                          |
-    |     data = np.random.normal(0, 1, (50, 50))                     |                                          |
[…]
