### 2024-01-14 43f2253
remove rst_prolog (now responsive grids are used instead of tables)
