### 2024-01-15 raphaelquast/EOmaps#209 @banesullivan
This all looks excellent and sounds good to me! Glad you found a compatible license and are happy with the implementation!
