### 2024-01-19 1d45fee
Internal re-structuring of pre-defined callback machinery - `<method>Callback` classes have been re-structured as mixins - `_cb` and `_cb_list` are removed from `_CallbackContainer` _ `_attach` classes now use mixins to define available callbacks   (all callbacks are decorated in the respective container so that    calling them will attach them as callbacks to the map)
