+++ eomaps/cb_container.py
@@ -124,12 +124,6 @@ class _CallbackContainer(object):
         else:
             self._temporary_artists = self._parent_container._temporary_artists
 
-        self._cb = cb_class(m, self._temporary_artists)
-        self._cb_list = cb_class._cb_list
-
-        self.attach = self._attach(self)
-        self.get = self._get(self)
-
         self._fwd_cbs = dict()
 
         self._method = method
@@ -196,10 +190,11 @@ class _CallbackContainer(object):
             self._m.BM._artists_to_clear.setdefault(self._method, []).append(art)
 
     def _sort_cbs(self, cbs):
+        _cb_list = self._attach._available_callbacks()
         if not cbs:
             return set()
         cbnames = set([i.rsplit("__", 1)[0].rsplit("_", 1)[0] for i in cbs])
-        sortp = self._cb_list + list(set(self._cb_list) ^ cbnames)
+        sortp = _cb_list + list(set(_cb_list) ^ cbnames)
         return sorted(
             list(cbs), key=lambda w: sortp.index(w.rsplit("__", 1)[0].rsplit("_", 1)[0])
         )
@@ -311,6 +306,7 @@ class _CallbackContainer(object):
 
     @property
     def execute_on_all_layers(self):
+        """Indicator if callbacks of this container are executed on all layers."""
         if self._parent_container is not None:
             return self._parent_container._execute_on_all_layers
 
@@ -370,6 +366,22 @@ class _CallbackContainer(object):
         """
         self._execute_while_toolbar_active = q
 
+    def _attach_decorator(self, func):
+        def inner(*args, **kwargs):
+            return self._add_callback(callback=func, *args, **kwargs)
+
+        return inner
+
+
+def _apply_decorator_to_all_public_methods(decorator):
+    def decorate(cls):
+        for attr in filter(lambda x: not x.startswith("_"), dir(cls)):
+            if callable(getattr(cls, attr)):
+                setattr(cls, attr, decorator(getattr(cls, attr)))
+        return cls
+
+    return decorate
+
 
 class _ClickContainer(_CallbackContainer):
     """
@@ -395,11 +407,18 @@ class _ClickContainer(_CallbackContainer):
         # the default button to use when attaching callbacks
         self._default_button = default_button
 
+        self.attach = self._attach(self)
+        self.attach = _apply_decorator_to_all_public_methods(self._attach_decorator)(
+            self.attach
+        )
+
+        self.get = self._get(self)
+
     class _attach:
         """
         Attach custom or pre-defined callbacks to the map.
 
-        Each callback-function takes 2 additional keyword-arguments:
+        Callback-functions accept the following additional keyword-arguments:
 
         double_click : bool
             Indicator if the callback should be executed on double-click (True)
@@ -457,17 +476,8 @@ class _ClickContainer(_CallbackContainer):
 
         def __init__(self, parent):
             self._parent = parent
-
-            # attach pre-defined callbacks
-            for cb in self._parent._cb_list:
-                setattr(
-                    self,
-                    cb,
-                    update_wrapper(
-                        partial(self._parent._add_callback, callback=cb),
-                        getattr(self._parent._cb, cb),
-                    ),
-                )
+            self.m = parent._m
+            self._temporary_artists = self._parent._temporary_artists
 
         def __call__(self, f, double_click=False, button=None, modifier=None, **kwargs):
             """
@@ -536,19 +546,24 @@ class _ClickContainer(_CallbackContainer):
                 **kwargs,
             )
 
+        @classmethod
+        def _available_callbacks(cls):
+            return list(filter(lambda x: not x.startswith("_"), dir(cls)))
+
     class _get:
         """Accessor for objects generated/retrieved by callbacks."""
 
         def __init__(self, parent):
-            self.m = parent._m
-            self.cb = parent._cb
[…]
