### 2024-01-19 5b30dd4
update docs to properly handle `Maps.cb`
