### 2024-01-19 7a1267a
use @wraps for callback decorators and fix default value for "on_motion"
