+++ tests/test_env.yml
@@ -23,11 +23,11 @@ dependencies:
   - owslib
   - requests
   # --------------for testing
-  - nbformat     # to parse Jupyter Notebooks
   - coveralls
   - pytest
   - pytest-cov
   # --------------for testing the docs
+  # (e.g. parsing .rst code-blocks and Jupyter Notebooks)
   - docutils
   - jupyter
   - nbformat
