### 2024-01-20 #138 @raphaelquast
@banesullivan @jhkennedy @yeelauren
Once again, thank you very much for all your helpful and productive comments! They really helped to pinpoint aspects of the package that required improvements (which are sometimes hard to spot if you know almost any line of code by heart).

I have now collected all major changes into a pending PR for EOmaps v8.0 [EOmaps#205](https://github.com/raphaelquast/EOmaps/pull/205) which should address the majority of comments mentioned here.

Here's an overview of the most relevant changes/fixes relevant for this review:

- EOmaps license will change from GPL to **BSD-3-clause**
- All links in readme and github now point to the [start-page](https://eomaps.readthedocs.io/en/latest/) of the docs (instead of [github.io](https://raphaelquast.github.io/EOmaps/)).
- There is now a top-level `environment.yml` containing all relevant dependencies
  - `pre-commit` and `git` are no longer included explicitly but mentioned in the contribution-guide
  - `tests/test_env.yml` is kept for now, since the test-runs do not require packages relevant to build the docs
  - `docs/docs_env.yml` now pinns docs-build dependencies and installs only minimal EOmaps requirements
- Installation instructions are now provided via `pyproject.toml`
  - Minimal python support is now clearly stated and >3.8
  - Obsolete dependencies have been removed
  - Optional dependency groups are now available for pip install
    (`pip install eomaps` now installs only bare minimum requirements)
- Major updates for the Docs:
  - Docs now build without any warnings and explicitly indicate the associated eomaps-version
  - Contribution guide has been re-worked to include info on testing, docs-build etc.
  - API docs have been completely re-worked to better represent the public API actually relevant for users
    (including access to available feature-names, pre-defined callbacks etc. )
  - Quickstart-guides now recommend `miniforge` as starting point for python+conda/mamba setup
  - Info on how to setup VSCode/VSCodium has been added to the docs
  - Example files are now located under "docs/examples" using a descriptive naming-convention (aligning with the associated figures and images)
  - A more responsive design is used to make the docs look nice also on small screens (phones etc.)
- In addition, there are a lot of fixes as well as internal changes and re-structuring to cleanup old code etc.

Since my personal todos tagged for EOmaps v8.0 are more or less implemented, I wanted to ask how we proceed with the review-process since it would be nice if the release of v8.0 can be aligned with the outcomes of the PyOpenSci review.

- All pending changes are summarized in [EOmaps#205](https://github.com/raphaelquast/EOmaps/pull/205)
- ...and the associated docs can be viewed [here](https://eomaps.readthedocs.io/en/dev/)

@banesullivan In addition, I also wanted to ask if it is feasible to still append a JOSS paper at this stage of the review or if you prefer that this is done post-review to avoid stretching the time for this review.
