### 2024-01-23 ac6c121
fix dynamic-shade indicator colorbars make datasets visible on all layers
