### 2024-01-23 04a9f79
fix vmin/vmax assignment for datashader "count" aggregations
