### 2024-01-24 bc62297
fix handling of masked values during vmin/vmax evaluation from files
