### 2024-01-24 #138 @banesullivan
Hey, @raphaelquast, I want to leave a quick note that I haven't forgotten about this! Again fantastic work addressing all of the feedback!

In the coming days, I will sit down and review all of this to take inventory of anything that may remain and answer you questions about JOSS

I appreciate your patience!
