### 2024-01-24 f2400ff
Fix issues with help-popups of the LayerTabBar of the companion-widget invalid result from LayerTabBar.event(), 'bool' expected not None
