### 2024-01-26 eb146e9
push current view to toolbar nav-stack after layout-restore (if possible)
