### 2024-01-27 afdf337
major colorbar updates
