+++ eomaps/helpers.py
@@ -100,6 +100,12 @@ class _TransformedBoundsLocator:
         # Subtracting transSubfigure will typically rely on inverted(),
         # freezing the transform; thus, this needs to be delayed until draw
         # time as transSubfigure may otherwise change after this is evaluated.
+        if ax.figure is None:
+            return TransformedBbox(
+                Bbox.from_bounds(*self._bounds),
+                self._transform,
+            )
+
         return TransformedBbox(
             Bbox.from_bounds(*self._bounds),
             self._transform - ax.figure.transSubfigure,
@@ -689,7 +695,6 @@ class LayoutEditor:
 
         self._ax_picked = []
         self._m_picked = []
-        self._cb_picked = []
 
         self._modifier_pressed = False
 
@@ -792,23 +797,8 @@ class LayoutEditor:
         return [m.ax for m in self.ms]
 
     @property
-    def cbaxes(self):
-        axes = list()
-        for m in self.ms:
-            axes.extend((i._ax for i in m._colorbars))
-        return axes
-
-    @property
     def axes(self):
         return self.f.axes
-
-    @property
-    def cbs(self):
-        # get all colorbars
-        cbs = list()
-        for m in self.ms:
-            cbs.extend(m._colorbars)
-        return cbs
 
     @staticmethod
     def roundto(x, base=10):
@@ -929,24 +919,13 @@ class LayoutEditor:
         for ax in self.axes:
             self._color_unpicked(ax)
 
-        for cb in self.cbs:
-            for ax in (cb.ax_cb, cb.ax_cb_plot):
-                self._color_unpicked(ax)
-
         for ax in self._ax_picked:
             if ax is not None:
                 self._color_picked(ax)
 
-        for cb in self._cb_picked:
-            for ax in (cb.ax_cb, cb.ax_cb_plot):
-                self._color_picked(ax)
-
     def _set_startpos(self, event):
         self._start_position = (event.x, event.y)
-        self._start_ax_position = {
-            i: (i.bbox.x0, i.bbox.y0)
-            for i in (*self._ax_picked, *(cb._ax for cb in self._cb_picked))
-        }
+        self._start_ax_position = {i: (i.bbox.x0, i.bbox.y0) for i in self._ax_picked}
 
     def _add_to_history(self):
         self._history_undone.clear()
@@ -987,7 +966,7 @@ class LayoutEditor:
 
         if eventax not in self.axes:
             # if no axes is clicked "unpick" previously picked axes
-            if len(self._ax_picked) + len(self._cb_picked) == 0:
+            if len(self._ax_picked) == 0:
                 # if there was nothing picked there's nothing to do
                 # except updating the info-text visibility
 
@@ -997,7 +976,6 @@ class LayoutEditor:
                 return
 
             self._ax_picked = []
-            self._cb_picked = []
             self._m_picked = []
             self._color_axes()
             self._remove_snap_grid()
@@ -1017,38 +995,22 @@ class LayoutEditor:
                     self._m_picked.remove(m)
                 else:
                     self._m_picked.append(m)
-            elif eventax in self.cbaxes:
-                cb = self.cbs[self.cbaxes.index(eventax)]
-                if cb in self._cb_picked:
-                    self._cb_picked.remove(cb)
-                else:
-                    self._cb_picked.append(cb)
             else:
                 if eventax in self._ax_picked:
                     self._ax_picked.remove(eventax)
                 else:
                     self._ax_picked.append(eventax)
         else:
-            selected = eventax in self._ax_picked
-            if eventax in self.cbaxes:
-                selected = (
-                    selected or self.cbs[self.cbaxes.index(eventax)] in self._cb_picked
-                )
-
-            if not selected:
+            if eventax not in self._ax_picked:
                 self._m_picked = []
[…]
