+++ eomaps/colorbar.py
@@ -1,28 +1,16 @@
-# Copyright EOmaps Contributors
-#
-# This file is part of EOmaps and is released under the BSD 3-clause license.
-# See LICENSE in the root of the repository for full licensing details.
-
-"""Interactive Colorbar."""
-
-import logging
-from functools import partial, lru_cache
-from itertools import cycle
+from itertools import pairwise, cycle
+from functools import partial
 from textwrap import dedent
-import copy
 
+import matplotlib.pyplot as plt
+from matplotlib.gridspec import GridSpecFromSubplotSpec, SubplotSpec
+import matplotlib.transforms as mtransforms
+from matplotlib.colors import LinearSegmentedColormap
+from matplotlib.patches import Rectangle
 
 import numpy as np
 
-from matplotlib.gridspec import GridSpecFromSubplotSpec, SubplotSpec
-import matplotlib.transforms as mtransforms
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-from matplotlib.colors import LinearSegmentedColormap
-
-from .helpers import pairwise, _TransformedBoundsLocator, register_modules
-
-_log = logging.getLogger(__name__)
+from eomaps.helpers import _TransformedBoundsLocator
 
 
 def get_named_bins_formatter(bins, names, show_values=False):
@@ -64,298 +52,511 @@ def get_named_bins_formatter(bins, names, show_values=False):
     return formatter
 
 
-class ColorBar:
-    """Class to draw colorbars with a histogram on top."""
-
+class ColorBarBase:
     def __init__(
         self,
-        m,
-        pos=0.4,
-        inherit_position=None,
-        margin=None,
-        hist_size=0.8,
-        hist_bins=256,
-        extend=None,
-        extend_frac=0.025,
         orientation="horizontal",
-        dynamic_shade_indicator=False,
-        show_outline=False,
-        tick_precision=2,
-        tick_formatter=None,
-        log=False,
-        out_of_range_vals="clip",
+        extend_frac=0.025,
+        dynamic=True,
+        padding=0.1,
         hist_kwargs=None,
-        label=None,
-        ylabel=None,
-        **kwargs,
+        tick_precision=2,
     ):
-        """
-        Add a colorbar to the map.
 
-        The colorbar always represents the data of the associated Maps-object
-        that was assigned in the last call to `m.plot_map()`.
+        self._hist_size = 0.9
 
-        By default, the colorbar will only be visible on the layer of the associated
-        Maps-object.
-
-        After the colorbar has been created, it can be accessed via:
-
-            >>> cb = m.colorbar
-
-        Parameters
-        ----------
-        pos : float or 4-tuple, optional
-
-            - float: fraction of the axis size that is used to create the colorbar.
-              The axes of the Maps-object will be shrunk accordingly to make space
-              for the colorbar.
-            - 4-tuple (x0, y0, width, height):
-              Absolute position of the colorbar in relative figure-units (0-1).
-              In this case, existing axes are NOT automatically re-positioned!
-
-            Note: By default, multiple colorbars on different layers share their
-            position! To force placement of a colorbar, use "inherit_position=False".
-
-            The default is 0.4.
-        inherit_position : bool or None optional
-            Indicator if the colorbar should share its position with other colorbars
-            that represent datasets on the same plot-axis.
-
-            - If True, and there is already another colorbar for the given plot-axis,
-              the value of "pos" will be ignored and the new colorbar will share its
-              position with the parent-colorbar. (e.g. all colorbars for a given axis will
-              overlap and moving a colorbar in one layer will move all other relevant
-              colorbars accordingly).
-            - If None: If the colorbar is added on a different layer than the parent
-              colorbar, use "inherit_position=True", else use "inherit_position=False".
[…]
