### 2024-01-29 53d157a
colorbar updates - re-introduce marings to preserve layout with respect to v7.x - fix identifying relevant colorbars with the layout-editor - introduce additional kwargs - fix inverting axis after ax.clear - avoid recursions in xlim callbacks
