+++ eomaps/colorbar.py
@@ -57,35 +57,29 @@ class ColorBarBase:
         self,
         orientation="horizontal",
         extend_frac=0.025,
-        dynamic=True,
-        padding=0.1,
         hist_kwargs=None,
         tick_precision=2,
+        margin=None,
     ):
 
         self._hist_size = 0.9
-
-        self._extend_frac = extend_frac
-        self._dynamic = dynamic
-
-        self.orientation = orientation
-
-        self._log = False
-
-        self._padding = padding
-
         if hist_kwargs is not None:
             self._hist_kwargs = hist_kwargs
         else:
             self._hist_kwargs = {}
+
+        self._extend_frac = extend_frac
+
+        self.orientation = orientation
+
+        self._tick_precision = tick_precision
+        self._margin = margin
 
         self._vmin = None
         self._vmax = None
         self._norm = None
         self._cmap = None
         self._data = None
-
-        self._tick_precision = tick_precision
 
     @property
     def _scm(self):
@@ -124,29 +118,27 @@ class ColorBarBase:
                 if self.orientation == "horizontal":
                     gs = GridSpecFromSubplotSpec(
                         2,
-                        3,
+                        1,
                         parent_ax.get_subplotspec(),
                         height_ratios=(1, pos),
-                        width_ratios=(self._padding, 1, self._padding),
                     )
                     parent_ax.set_subplotspec(gs[0, :])
                     self._ax = f.add_subplot(
-                        gs[1, 1],
+                        gs[1, :],
                         label="cb",
                         zorder=zorder,
                     )
                 else:
                     gs = GridSpecFromSubplotSpec(
-                        3,
+                        1,
                         2,
                         parent_ax.get_subplotspec(),
                         width_ratios=(1, pos),
-                        height_ratios=(self._padding, 1, self._padding),
                     )
 
-                    parent_ax.set_subplotspec(gs[1, 0])
+                    parent_ax.set_subplotspec(gs[:, 0])
                     self._ax = f.add_subplot(
-                        gs[1, 1],
+                        gs[:, 1],
                         label="cb",
                         zorder=zorder,
                     )
@@ -193,8 +185,8 @@ class ColorBarBase:
         else:
             self.ax_cb_plot.sharey(self.ax_cb)
 
-            # for vertical colorbars, histogram-axis must be inverted!
-            self.ax_cb_plot.invert_xaxis()
+            # # for vertical colorbars, histogram-axis must be inverted!
+            self.ax_cb_plot.xaxis.set_inverted(True)
 
         # keep the background of the plot-axis but remove the outer frame
         self.ax_cb_plot.spines["top"].set_visible(False)
@@ -213,11 +205,14 @@ class ColorBarBase:
 
         def ychanged(event):
             if self.orientation == "horizontal":
-                self.ax_cb_plot.set_ylim(0, None, emit=False)
+                with self.ax_cb_plot.callbacks.blocked(signal="ylim_changed"):
+                    self.ax_cb_plot.set_ylim(0, None, emit=False)
 
         def xchanged(event):
             if self.orientation == "vertical":
-                self.ax_cb_plot.set_xlim(None, 0, emit=False)
+                with self.ax_cb_plot.callbacks.blocked(signal="xlim_changed"):
+                    self.ax_cb_plot.xaxis.set_inverted(True)
[…]
