### 2024-01-29 f1f580e
update m.add_colorbar
