### 2024-01-29 66bc47e
add docstrings and missing methods to colorbar
