+++ eomaps/colorbar.py
@@ -127,6 +127,7 @@ class ColorBarBase:
                         gs[1, :],
                         label="cb",
                         zorder=zorder,
+                        navigate=False,
                     )
                 else:
                     gs = GridSpecFromSubplotSpec(
@@ -141,12 +142,14 @@ class ColorBarBase:
                         gs[:, 1],
                         label="cb",
                         zorder=zorder,
+                        navigate=False,
                     )
             elif isinstance(pos, SubplotSpec):
                 self._ax = f.add_subplot(
                     pos,
                     label="cb",
                     zorder=zorder,
+                    navigate=False,
                 )
             elif isinstance(pos, (list, tuple)):
                 x0, y0, w, h = pos
@@ -156,7 +159,7 @@ class ColorBarBase:
 
                 # the parent axes holding the 2 child-axes
                 self._ax = plt.Axes(f, bbox, label="cb", zorder=zorder)
-                f.add_axes(self._ax)
+                f.add_axes(self._ax, navigate=False)
 
             # make all spines, labels etc. invisible for the base-axis
             self._ax.set_axis_off()
@@ -166,6 +169,7 @@ class ColorBarBase:
             self._ax.get_position(),
             label="EOmaps_cb",
             zorder=zorder - 1,  # make zorder 1 lower than container axes for picking
+            navigate=False,
         )
 
         # histogram axes
@@ -173,6 +177,7 @@ class ColorBarBase:
             self._ax.get_position(),
             label="EOmaps_cb_hist",
             zorder=zorder - 1,  # make zorder 1 lower than container axes for picking
+            navigate=False,
         )
 
         # add axes as child-axes
@@ -263,30 +268,17 @@ class ColorBarBase:
         self._set_axes_locators(l_cb_bounds, l_hist_bounds)
         self._style_hist_ticks()
 
-    def get_extend_fracs(self):
-        # if no colorbar is found, use the full axis
-        if not hasattr(self, "cb"):
-            return 0, 1
-
-        return 0, 1
-
-        extend = self.cb.extend
-
-        # extend fraction is defined as % of the interior colorbar length!
-        getfrac = lambda n: self._extend_frac / (1 + n * self._extend_frac)
-
-        if extend == "both":
-            frac = getfrac(2)
-            return frac, 1 - 2 * frac
-        elif extend == "min":
-            frac = getfrac(1)
-            return frac, 1 - frac
-        elif extend == "max":
-            return 0, 1 - getfrac(1)
-        else:
-            return 0, 1
-
     def set_scale(self, log=False):
+        """
+        Set the scale of the colorbar histogram. (e.g. logarithmic or linear)
+
+        Parameters
+        ----------
+        log : bool, optional
+            If True, use a logarithmic scale for the histogram.
+            The default is False.
+
+        """
         if self.orientation == "horizontal":
             # set axis scale
             if log is True:
@@ -603,7 +595,7 @@ class ColorBarBase:
 
     def set_labels(self, cb_label=None, hist_label=None, **kwargs):
         """
-        Set the labels (and the styling) for the colorbar (and the histogram).
+        Set the labels (and the label-style) for the colorbar (and the histogram).
 
         For more details, see `ColorBar.ax_cb.set_xlabel(..)` and matplotlib's `.Text`
         properties.
@@ -672,6 +664,8 @@ class ColorBarBase:
 
 
 class ColorBar(ColorBarBase):
+    """Class to draw colorbars with a histogram on top"""
+
     def __init__(self, *args, inherit_position=True, **kwargs):
         super().__init__(*args, **kwargs)
 
@@ -798,7 +792,7 @@ class ColorBar(ColorBarBase):
[…]
