### 2024-01-29 208bd84
unify treatment of parsing of layer-names - introduce helper-mixin LayerParser to collect all relevant methods
