+++ eomaps/helpers.py
@@ -280,6 +280,153 @@ def _add_to_docstring(prefix=None, suffix=None, insert=None):
     return decorator
 
 
+class LayerParser:
+    @staticmethod
+    def _parse_single_layer_str(layer):
+        """
+        Parse a single layer-string (with optional transparency assignment).
+
+        Parameters
+        ----------
+        layer : str
+            A layer-string (with transparency provided in curly brackets).
+
+        Returns
+        -------
+        name: str
+            The name of the layer.
+        alpha:
+            The transparency of the layer.
+
+        """
+        # split transparency
+        t_split = layer.find("{")
+        if t_split > 0:
+            name = layer[:t_split]
+            alpha = layer[t_split + 1 :]
+            if not alpha.endswith("}"):
+                raise TypeError(
+                    f"EOmaps: unable to parse multilayer-transparency for '{layer}'"
+                )
+            return name, float(alpha[:-1])
+        else:
+            return layer, 1
+
+    @classmethod
+    def _parse_multi_layer_str(cls, layer=None):
+        layers, alphas = zip(*map(cls._parse_single_layer_str, layer.split("|")))
+        return list(layers), list(alphas)
+
+    @classmethod
+    def _layer_is_subset(cls, layer1, layer2):
+        """
+        Return True if combined layer-name 'layer2' is a subset of 'layer1'.
+
+        - Transparency assignments are stripped off before comparison
+
+        Parameters
+        ----------
+        layer1, layer2 : str
+            The combined layer-names to check.
+
+        Returns
+        -------
+        subset: bool
+            True if layer2 is a subset of layer1, False otherwise
+
+        """
+        # get a list of the currently visible layers
+        layers1, _ = cls._parse_multi_layer_str(layer1)
+        layers2, _ = cls._parse_multi_layer_str(layer2)
+
+        return set(layers1).issubset(layers2)
+
+    @staticmethod
+    def _get_combined_layer_name(*args):
+        """
+        Create a combine layer name from layer-names or tuples (name, transparency).
+
+        Parameters
+        ----------
+        *args : str or tuple
+            The layers to combine. (e.g. `"A"`, `"B"` or `("A", .5)`, `("B", .23)`, ...)
+
+        Returns
+        -------
+        str
+            The combined layer-name.
+
+        """
+        try:
+            combnames = []
+            for i in args:
+                if isinstance(i, str):
+                    combnames.append(i)
+                elif isinstance(i, (list, tuple)):
+                    assert (
+                        len(i) == 2
+                        and isinstance(i[0], str)
+                        and i[1] >= 0
+                        and i[1] <= 1
+                    ), (
+                        f"EOmaps: unable to identify the layer-assignment: {i} .\n"
+                        "You can provide either a single layer-name as string, a list "
+                        "of layer-names or a list of tuples of the form: "
+                        "(< layer-name (str) >, < layer-transparency [0-1] > )"
+                    )
+
+                    if i[1] < 1:
+                        combnames.append(i[0] + "{" + str(i[1]) + "}")
+                    else:
+                        combnames.append(i[0])
+                else:
+                    raise TypeError(
+                        f"EOmaps: unable to identify the layer-assignment: {i} .\n"
+                        "You can provide either a single layer-name as string, a list "
+                        "of layer-names or a list of tuples of the form: "
+                        "(< layer-name (str) >, < layer-transparency [0-1] > )"
+                    )
+            return "|".join(combnames)
+        except Exception:
+            raise TypeError(f"EOmaps: Unable to combine the layer-names {args}")
+
+    @staticmethod
+    def _check_layer_name(layer):
+        if not isinstance(layer, str):
[…]

[…]
