### 2024-01-29 528d263
unify treatment of parsing of layer-names - introduce helper-mixin LayerParser to collect all relevant methods
