+++ eomaps/layout_editor.py
@@ -1,0 +1,1040 @@
+# Copyright EOmaps Contributors
+#
+# This file is part of EOmaps and is released under the BSD 3-clause license.
+# See LICENSE in the root of the repository for full licensing details.
+
+"""LayoutEditor used to handle interactive re-arrangement of axes."""
+
+import logging
+from contextlib import ExitStack
+from pathlib import Path
+import json
+
+import numpy as np
+from matplotlib.lines import Line2D
+from matplotlib.transforms import Bbox
+from matplotlib.axis import XAxis, YAxis
+from matplotlib.spines import Spine
+
+_log = logging.getLogger(__name__)
+
+
+class LayoutEditor:
+    """Class to handle interactive re-positioning of figure-objects."""
+
+    def __init__(self, m, modifier="alt+d", cb_modifier="control"):
+        self.modifier = modifier
+        self.cb_modifier = cb_modifier
+
+        self.m = m
+        self.f = self.m.parent.f
+
+        self._ax_picked = []
+        self._m_picked = []
+        self._cb_picked = []
+
+        self._modifier_pressed = False
+
+        self.cids = []
+
+        # indicator if the pick-callback should be re-attached or not
+        self._reattach_pick_cb = False
+
+        self.f.canvas.mpl_connect("key_press_event", self.cb_key_press)
+        self.f.canvas.mpl_connect("resize_event", self._on_resize)
+
+        # the snap-to-grid interval (0 means no snapping)
+        self._snap_id = 5
+
+        # an optional filepath that will be used to store the layout once the
+        # editor exits
+        self._filepath = None
+
+        # indicator if scaling should be in horizontal or vertical direction
+        self._scale_direction = "both"
+
+        # indicator if multiple-axis select key is pressed or not (e.g. "shift")
+        self._shift_pressed = False
+
+        self._max_hist_steps = 1000
+        self._history = list()
+        self._history_undone = list()
+
+        self._current_bg = None
+
+        self._info_text = None
+        self._info_text_hidden = False
+
+    def add_info_text(self):
+        self._info_text_hidden = False
+
+        a = self.m.f.text(
+            0.72,
+            0.98,
+            (
+                "LayoutEditor Controls:\n\n"
+                "0 - 9:  Snap-grid spacing\n"
+                "SHIFT:  Multi-select\n"
+                "P:      Print to console\n"
+                "ESCAPE (or ALT + L): Exit\n"
+                "\n"
+                "ARROW-KEYS:   Move\n"
+                "SCROLL (+/-): Resize\n"
+                "  H:    horizontal\n"
+                "  V:    vertical\n"
+                "  ctrl: histogram"
+                "\n\n(right-click to hide info)"
+            ),
+            transform=self.m.f.transFigure,
+            ha="left",
+            va="top",
+            fontsize=min(self.m.f.bbox.width * 72 / self.m.f.dpi / 60, 12),
+            bbox=dict(
+                boxstyle="round", facecolor=".8", edgecolor="k", lw=0.5, alpha=0.9
+            ),
+            zorder=1e6,
+            fontfamily="monospace",
+        )
+        return a
+
+    def _update_info_text(self):
+        if getattr(self, "_info_text", None) is not None:
+            self._info_text.set_fontsize(
+                min(self.m.f.bbox.width * 72 / self.m.f.dpi / 60, 15)
+            )
+
+    def _on_resize(self, *args, **kwargs):
+        # update snap-grid on resize
+        if self.modifier_pressed:
+            self._add_snap_grid()
+            self._update_info_text()
+
+    @property
+    def modifier_pressed(self):
+        return self._modifier_pressed
+
+    @modifier_pressed.setter
+    def modifier_pressed(self, val):
+        self._modifier_pressed = val
+        self.m.cb.execute_callbacks(not val)
+
[…]
