+++ eomaps/shapes.py
@@ -1351,6 +1351,178 @@ class Shapes(object):
         def radius_crs(self):
             return "in"
 
+    class _SphericalVoronoiDiagram(object):
+        name = "spherical_voronoi_diagram"
+
+        def __init__(self, m):
+            self._m = m
+            self._mask_radius = None
+
+        def __call__(self, masked=True, mask_radius=None):
+            """
+            Draw a Spherical Voronoi-Diagram of the data.
+
+            Parameters
+            ----------
+            masked : bool
+                Indicator if the voronoi-diagram should be masked or not
+
+            mask_radius : float, optional
+                The radius used for masking the voronoi-diagram
+                (in units of the plot-crs)
+                The default is 4 times the estimated data-radius.
+            """
+            from . import MapsGrid  # do this here to avoid circular imports!
+
+            for m in self._m if isinstance(self._m, MapsGrid) else [self._m]:
+                shape = self.__class__(m)
+
+                shape.mask_radius = mask_radius
+                shape.masked = masked
+
+                m._shape = shape
+
+        @property
+        def _initargs(self):
+            return dict(mask_radius=self.mask_radius, masked=self.masked)
+
+        def __repr__(self):
+            try:
+                s = f"voronoi_diagram(mask_radius={self.mask_radius}, masked={self.masked})"
+            except AttributeError:
+                s = "voronoi_diagram(mask_radius, masked)"
+
+            return s
+
+        @property
+        def mask_radius(self):
+            r = Shapes._get_radius(self._m, self._mask_radius, "out")
+            if self._mask_radius is None:
+                return [i * 4 for i in r]
+            else:
+                return r
+
+        @mask_radius.setter
+        def mask_radius(self, val):
+            self._mask_radius = val
+
+        def _get_voronoi_verts_and_mask(self, x, y, crs, radius, masked=True):
+            try:
+                from scipy.spatial import SphericalVoronoi
+                from itertools import zip_longest
+            except ImportError:
+                raise ImportError("'scipy' is required for 'voronoi'!")
+
+            # transform from crs to the lon/lat
+
+            t_in_lonlat = self._m._get_transformer(
+                self._m.get_crs(crs), self._m.get_crs(crs).geodetic_crs
+            )
+
+            t_lonlat_plot = self._m._get_transformer(
+                self._m.get_crs(crs).geodetic_crs, self._m.crs_plot
+            )
+
+            xy_orig = np.column_stack((x, y))  # remember orig input coords
+
+            x0, y0 = t_in_lonlat.transform(x, y)
+
+            x0 = (x0 + 360) % 360
+            y0 = (y0 + 180) % 180
+
+            datamask = np.isfinite(x0) & np.isfinite(y0)
+            [radiusx, radiusy] = radius
+
+            maxdist = 2 * np.mean(np.sqrt(radiusx**2 + radiusy**2))
+
+            xy = np.column_stack((x0[datamask], y0[datamask]))
+
+            # plot the unit sphere for reference (optional)
+            r = 1
+
+            theta, phi = np.deg2rad(xy.T)
+            x = r * np.cos(theta) * np.sin(phi)
+            y = r * np.sin(theta) * np.sin(phi)
+            z = r * np.cos(phi)
+
+            vor = SphericalVoronoi(np.column_stack((x, y, z)), r, [0, 0, 0])
+            vor.sort_vertices_of_regions()
+
+            rect_regions = np.array(list(zip_longest(*vor.regions, fillvalue=-2))).T
+            # (use -2 instead of None to make np.take work as expected)
+
+            # rect_regions = rect_regions[vor.point_region]
[…]
