### 2024-01-29 ce610e7
update docstrings
