### 2024-01-29 6d912c3
fix attaching custom click callbacks with "on_motion=True"
