### 2024-01-29 16f4cf6
fix hist kwargs and range of dynamic colorbar
