### 2024-01-30 f5a7085
add remove method for colorbar and fix unwanted draws
