### 2024-01-30 35597da
move methods and fix updating layer on set_scale
