+++ eomaps/layout_editor.py
@@ -1,3 +1,10 @@
+# Copyright EOmaps Contributors
+#
+# This file is part of EOmaps and is released under the BSD 3-clause license.
+# See LICENSE in the root of the repository for full licensing details.
+
+"""Definition of the LayoutEditor used to interactively re-position axes."""
+
 import logging
 from contextlib import ExitStack
 from pathlib import Path
