### 2024-01-30 feee183
move BlitManager to dedicated module
