### 2024-01-30 a3ba7ce
avoid setting limit to 0 on logarithmic axes
