### 2024-01-30 9125c6a
update docs - explicitly skip __init__ methods from class docstrings
