### 2024-01-31 9c5f4db
fix call to "_get_clip_path()" for peek-layer callbacks
