### 2024-02-01 5aa28f4
fix issues with figure-exports after `m.apply_layout(...)` (see #226)
