### 2024-02-04 #138 @banesullivan
@raphaelquast, this has truly come together fantastically! Thank you for diligently working to address all of the feedback our excellent reviewers brought up!

I've read this review thread up and down a few times now... (lots of amazing work happening here!!) and I believe you have addressed all of the feedback raised. That said, I'd like to give @yeelauren and @jhkennedy time to double check and sign off on the review.

@yeelauren and @jhkennedy, to make this process a little easier for you both I've tried to sift through your open review request/comments to track down the state of things. The following are my findings:

@yeelauren's comments:

- Explicitly calling out the statement of need and why users may turn to EOmaps: I'm not sure if this has been fully addressed. @raphaelquast, is there a line you can point us to that you feel captures the statement of need? If so, perhaps this could come right after the "EOmaps is a python package to visualize and analyze geographical datasets" line in the README?
- @yeelauren raised a great question on the why EOmaps is a separate package versus adding a PR/additional functionality to cartopy directly. I think @raphaelquast reponded well to this in [this comment](https://github.com/pyOpenSci/software-submission/issues/138#issuecomment-1844941504). I'm wondering if this explanation should be added in a new section to the CONTRIBUTING guide of EOmaps?
- Set up and testing: I think @raphaelquast has addressed on these and my latest run of the tests under the `dev` branch show everything passing.
- Testing warnings: @yeelauren raised some concerns about specific testing warnings that seemed of high importance and I think some of these have been addressed. In my latest run of the test suite from the `dev` branch, there are still a high number of warnings thrown, however they do not greatly concern me for the sake of this review at this time. I've raised a new issue in EOmap https://github.com/raphaelquast/EOmaps/issues/227 to track these and feel comfortable not having those block this review.

@jhkennedy's comments:

- Supported python versions: I believe this has been cleaned up and now EOmaps is supporting python>=3.8
- Setting up a development environment and contributing guide: @raphaelquast has included a new contributing guide that I think address your concerns.
- Documentation versioning: I think this is handled well now via RTD's versioning mechanisms
- Testing failures and hangs: On the latest `dev` branch, I'm having no issues about tests
- Testing figure windows: there is a note now in the contributing guid about this
- Metadata in pyproject.toml: I think all of this has been added!

@raphaelquast, on your question about the JOSS review: I'm leaning towards following up with the JOSS paper post-review at this stage, but I will talk with some other's in PyOpenSci to fully assess this.... stand by.
