### 2024-02-05 raphaelquast/EOmaps#227 @raphaelquast
@banesullivan Thank you very much for opening this detailed issue!

I had a quick look where the extremely large number of warnings is coming from and realized that almost all warnings relate to the `numpy DeprecationWarning` about conversion of an array with ndim>0 to scalar.

This warning is ultimately related to the way how `PyProj` handles 1-element arrays as discussed here: https://github.com/pyproj4/pyproj/issues/130.

- [ ] I'll try to figure out if the numpy-warning can be ignored or if changes in eomaps are required to ensure compatibility with upcoming numpy versions!
- [ ] I'll implement a filter for all warnings that can savely be ignored (e.g. download warnings, warnings about in correct buffers in geographic-crs etc.)
