### 2024-02-05 e5011ef
filter warnings that can savely be ignored for pytest (#227)
