### 2024-02-05 raphaelquast/EOmaps#227 @raphaelquast
`PyProj` devs confirmed that the numpy-warning can safely be ignored (which will get rid of ~900 of the warnings).
I've implemented a filter for all other warnings that can be ignored and now we're down to 25 again!

I'll report back once I find the time to check the remaining warnings!
