### 2024-02-05 2f1fb40
ignore more warnings with pytest
