+++ pyproject.toml
@@ -82,7 +82,8 @@ Repository = "https://github.com/raphaelquast/eomaps"
 
 [tool.pytest.ini_options]
 filterwarnings = [
-    "ignore:DownloadWarning",
+    "ignore::DownloadWarning",
+    "ignore:Pyarrow will become a required dependency of pandas:DeprecationWarning",
     "ignore:Geometry is in a geographic CRS. Results from *:UserWarning",
     "ignore:Back azimuth is being returned by default *:UserWarning",
     "ignore:Conversion of an array with ndim > 0:DeprecationWarning"
