### 2024-02-05 cbc99aa
fix filtering custom warning category
