+++ pyproject.toml
@@ -82,7 +82,7 @@ Repository = "https://github.com/raphaelquast/eomaps"
 
 [tool.pytest.ini_options]
 filterwarnings = [
-    "ignore::DownloadWarning",
+    "ignore:cartopy.io.DownloadWarning",
     "ignore:Pyarrow will become a required dependency of pandas:DeprecationWarning",
     "ignore:Geometry is in a geographic CRS. Results from *:UserWarning",
     "ignore:Back azimuth is being returned by default *:UserWarning",
