### 2024-02-05 c5f791c
avoid \* escape sequence in docstrings
