### 2024-02-06 #138 @raphaelquast
@banesullivan Thank you very much for your thorough review and the kind words!
I've already started investigating the extensive number of warnings (in raphaelquast/EOmaps#227) and I agree that the README can still be improved... I'll try to come up with a new README as soon as possible!
