### 2024-02-06 6aac958
fix using reprojected 1d datasets with shade_raster
