### 2024-02-06 1dfcb2d
fix parsing maxsize argument in companion widget for raster-shape
