### 2024-02-07 501989b
fix passing kwargs to colorbar creation
