### 2024-02-07 6df4154
trim very long parameter names in annotation text
