### 2024-02-07 05c4919
fix treatment of infinite values in weighted colorbar histograms
