### 2024-02-07 30b7249
fix histogram tick-styles are resetted on layout-change
