### 2024-02-07 2a995be
allow passing transform-kwarg to m.add_text
