### 2024-02-08 d61e5e9
fix all-layer check in BM._layer_visible
