### 2024-02-08 31a6fad
fix selecting multiple layers in layer-dropdown - make sure tab-bar is properly updated - make sure transparencies are kept as-is
