### 2024-02-08 8116c28
make sure dataset-dropdown in companion-widget does not expand main window (for very long layer names)
