### 2024-02-08 4577a61
make sure window-title (e.g. layer-name) is properly truncated if too long
