### 2024-02-11 0d308d4
add pyhton 3.12 to unittest matrix
