### 2024-02-12 c1dfe8d
add kwarg to adjust the style of the line between colorbar and histogram
