### 2024-02-12 8ecb2bb
warn if weights are used for dynamic-shade indicators (not yet supported)
