### 2024-02-12 6796b90
deprecate colorbar kwargs - "show_outline" -> "outline" - "ylabel" -> "hist_label"
