### 2024-02-13 8691c96
avoid memory-leaks caused by references to artists in pick-containers
