### 2024-02-13 fce0011
include cleanup exception tracebacks for debug logs
