### 2024-02-13 08f82ac
move "_identify_data(..)" to _data_manager module
