+++ eomaps/eomaps.py
@@ -4217,139 +4217,6 @@ class Maps(metaclass=_MapsMeta):
             except Exception:
                 _log.exception("EOmaps: Problem executing 'on_add_child' action:")
 
-    def _identify_data(self, data=None, x=None, y=None, parameter=None):
-        # identify the way how the data has been provided and convert to the internal
-        # structure
-
-        if data is None:
-            data = self.data_specs.data
-        if x is None:
-            x = self.data_specs.x
-        if y is None:
-            y = self.data_specs.y
-        if parameter is None:
-            parameter = self.data_specs.parameter
-
-        # check other types before pandas to avoid unnecessary import
-        if data is not None and not isinstance(data, (list, tuple, np.ndarray)):
-            (pd,) = register_modules("pandas", raise_exception=False)
-
-            if pd is None:
-                raise TypeError(
-                    f"EOmaps: Unable to handle the input-data type: {type(data)}"
-                )
-
-            if isinstance(data, pd.DataFrame):
-                if parameter is not None:
-                    # get the data-values
-                    z_data = data[parameter].values
-                else:
-                    z_data = np.repeat(np.nan, len(data))
-
-                # get the index-values
-                ids = data.index.values
-
-                if isinstance(x, str) and isinstance(y, str):
-                    # get the data-coordinates
-                    xorig = data[x].values
-                    yorig = data[y].values
-                else:
-                    assert isinstance(x, (list, np.ndarray, pd.Series)), (
-                        "'x' must be either a column-name, or explicit values "
-                        " specified as a list, a numpy-array or a pandas"
-                        + f" Series object if you provide the data as '{type(data)}'"
-                    )
-                    assert isinstance(y, (list, np.ndarray, pd.Series)), (
-                        "'y' must be either a column-name, or explicit values "
-                        " specified as a list, a numpy-array or a pandas"
-                        + f" Series object if you provide the data as '{type(data)}'"
-                    )
-
-                    xorig = np.asanyarray(x)
-                    yorig = np.asanyarray(y)
-
-                return z_data, xorig, yorig, ids, parameter
-
-        # identify all other types except for pandas.DataFrames
-        # lazily check if pandas was used
-        pandas_series_data = False
-        for iname, i in zip(("x", "y", "data"), (x, y, data)):
-            if iname == "data" and i is None:
-                # allow empty datasets
-                continue
-
-            if not isinstance(i, (list, tuple, np.ndarray)):
-                (pd,) = register_modules("pandas", raise_exception=False)
-
-                if pd and not isinstance(i, pd.Series):
-                    raise AssertionError(
-                        f"{iname} values must be a list, numpy-array or pandas.Series"
-                    )
-                else:
-                    if iname == "data":
-                        pandas_series_data = True
-
-        # set coordinates by extent
-        if isinstance(x, tuple) and isinstance(y, tuple):
-            assert data is not None, (
-                "EOmaps: If x- and y are provided as tuples, the data must be a 2D list"
-                " or numpy-array!"
-            )
-
-            shape = np.shape(data)
-            assert len(shape) == 2, (
-                "EOmaps: If x- and y are provided as tuples, the data must be a 2D list"
-                " or numpy-array!"
-            )
-
-            # get the data-coordinates
-            xorig = np.linspace(*x, shape[0])
-            yorig = np.linspace(*y, shape[1])
-
-        else:
-            # get the data-coordinates
-            xorig = np.asanyarray(x)
-            yorig = np.asanyarray(y)
-
-        if data is not None:
-            # get the data-values
[…]
