### 2024-02-13 bb32839
make sure "_ensure_same_pick_id" is set on all involved pick-containers
