### 2024-02-13 667e76b
fix inheriting hist_size from parent colorbars
