+++ README.md
@@ -1,7 +1,7 @@
 
 <p align="center">
     <a href=https://github.com/raphaelquast/EOmaps>
-    <img src="https://github.com/raphaelquast/EOmaps/blob/master/docs/_static/logo.png?raw=true" alt="EOmaps logo" width="55%">
+    <img src="https://github.com/raphaelquast/EOmaps/blob/master/docs/_static/logo.png?raw=true" alt="EOmaps logo" width="50%">
     </a>
 </p>
 
@@ -13,64 +13,113 @@
 <a href="https://app.gitter.im/#/room/#EOmaps:gitter.im" target="_blank"><img src="https://img.shields.io/gitter/room/raphaelquast/EOmaps?style=social" alt="chat on gitter" align="left" style="height: 20px !important;" ></a>
 
 ----
+  
 
-# EOmaps - Interactive maps in python!
+<h3 align="center">A python package to visualize and analyze geographical datasets.</h3>
 
-**EOmaps** is a <tt>python</tt> package to visualize and analyze geographical datasets.
+> **EOmaps** aims to provide a comprehensive, flexible, well-documented and easy-to-use API to create publication-ready maps that can directly be used for interactive data analysis.
 
-It is built on top of [matplotlib](https://matplotlib.org/) and [cartopy](https://scitools.org.uk/cartopy/docs/latest/) and provides an intuitive and easy-to-use interface to speed up and simplify the creation and comparison of maps.
+## How does it work?
 
+**EOmaps** is built on top of [matplotlib](https://matplotlib.org/) and [cartopy](https://scitools.org.uk/cartopy/docs/latest/) and integrates well with the scientific python infrastructure (e.g., `numpy`, `pandas`, `geopandas`, `xarray` etc.), allowing you to visualize point-, raster- or vector-datasets provided in almost any format you can imagine, no matter if you're dealing with just a few unsorted datapoints or multi-dimensional stacks of global high-resolution datasets.  
+
+Figures created with EOmaps are multi-layered, so you can (transparently) overlay and interactively compare your datasets with ease. With the accompanying GUI widget, you can quickly switch layers, change the layout, examine the large collection of features and web-map services, and explore the capabilities of EOmaps. 
+Once you're map is ready, you can export it as high-resolution image or vector-graphic for further editing.
+Leveraging the powers of matplotlib, you can also embed interactive maps in Jupyter Notebooks, GUI frameworks like `Qt`, `tkinter`, `wx` or on a webpage (html).
+
+
+> [!TIP]
 > For a quick hands-on introduction, checkout this article on dev.to:  
 > [Geographic data visualization and analysis with EOmaps: Interactive maps in python!](https://dev.to/raphaelquast/geographic-data-visualization-and-analysis-with-eomaps-interactive-maps-in-python-48e1)
->
-> ---
+
+---
 
 ### What can EOmaps do for you?
 
-- Create [▤ multi-layered maps](https://eomaps.readthedocs.io/en/latest/api_basics.html#layer-management) and interactively compare different layers with each other
-- [🔴 Visualize datasets](https://eomaps.readthedocs.io/en/latest/api_data_visualization.html) with  millions of datapoints and handle reprojections
+<!-- 
+<img src=https://eomaps.readthedocs.io/en/latest/_images/intro.png width=40%>
[…]
