### 2024-02-14 7b54a43
make blit-manager methods private
