+++ eomaps/draw.py
@@ -253,7 +253,7 @@ class ShapeDrawer:
         for cb in self._on_poly_remove:
             cb()
 
-        self._m.BM.on_draw(None)
+        self._m.BM._on_draw_cb(None)
 
     def _init_draw_line(self):
         if self._line is None:
@@ -291,7 +291,7 @@ class ShapeDrawer:
         """Trigger re-drawing shapes."""
         # NOTE: If a drawer is active, this function is also called on any ordinary
         # draw-event (e.g. zoom/pan/resize) to keep the indicators visible.
-        # see "m.BM.on_draw()"
+        # see "m.BM._on_draw_cb()"
 
         artists = self._indicator_artists
 
@@ -656,7 +656,7 @@ class ShapeDrawer:
                     self._m.BM.add_artist(ph, layer=self.layer)
                 else:
                     self._m.BM.add_bg_artist(ph, layer=self.layer)
-                    self._m.BM.on_draw(None)
+                    self._m.BM._on_draw_cb(None)
 
                 ID = max(self._artists) + 1 if self._artists else 0
                 self._artists[ID] = ph
@@ -748,7 +748,7 @@ class ShapeDrawer:
                     self._m.BM.add_artist(ph, layer=self.layer)
                 else:
                     self._m.BM.add_bg_artist(ph, layer=self.layer)
-                    self._m.BM.on_draw(None)
+                    self._m.BM._on_draw_cb(None)
 
                 ID = max(self._artists) + 1 if self._artists else 0
                 self._artists[ID] = ph
@@ -833,7 +833,7 @@ class ShapeDrawer:
                     self._m.BM.add_artist(ph, layer=self.layer)
                 else:
                     self._m.BM.add_bg_artist(ph, layer=self.layer)
-                    self._m.BM.on_draw(None)
+                    self._m.BM._on_draw_cb(None)
 
                 ID = max(self._artists) + 1 if self._artists else 0
                 self._artists[ID] = ph
