### 2024-02-14 0b30c7a
make "BM.cleanup_layer" method private
