### 2024-02-14 ccf4937
init blit-manager with layer of parent maps-object
