### 2024-02-14 0151563
make blit-manager, data_specs and classify_specs accessible for sphinx
