+++ docs/gen_autodoc_file.py
@@ -66,6 +66,9 @@ def make_feature_toctree_file():
         "read_file",
         "util",
         "add_wms",
+        "BM",
+        "data_specs",
+        "classify_specs",
     ):
         members.extend(get_members(Maps, key, False))
     for key in ("add_feature", "cb"):
