+++ docs/conf.py
@@ -68,6 +68,11 @@ def setup(app):
     Maps.cb.move.attach.__name__ = "attach"
     Maps.cb.move.get.__name__ = "get"
 
+    Maps.BM.__name__ = "BM"
+
+    Maps.data_specs.__name__ = "data_specs"
+    Maps.classify_specs.__name__ = "classify_specs"
+
 
 # -- Project information
 
