### 2024-02-14 8420944
add basic docstrings for data container properties
