### 2024-02-15 0184b7d
fix masked-array issues with datashader/numba when using "shade_points"
