### 2024-02-15 b455a95
properly handle pick-events that do not have an "idx" property
