### 2024-02-16 527d149
avoid issues if collection has no _coll_dynamic attribute (e.g. shade shapes)
