### 2024-02-16 8e1f80b
trigger keypress callbacks irrespective of the toolbar mode
