### 2024-02-16 1c40f50
handle raster-aggregation if block-size cannot be evaluated (e.g. None)
