+++ README.md
@@ -13,56 +13,57 @@
 <a href="https://app.gitter.im/#/room/#EOmaps:gitter.im" target="_blank"><img src="https://img.shields.io/gitter/room/raphaelquast/EOmaps?style=social" alt="chat on gitter" align="left" style="height: 20px !important;" ></a>
 
 ----
-  
 
 <h3 align="center">A python package to visualize and analyze geographical datasets.</h3>
 
-> **EOmaps** aims to provide a comprehensive, flexible, well-documented and easy-to-use API to create publication-ready maps that can directly be used for interactive data analysis.
+<table>
+<tr><td>  
+<i><b>EOmaps</b> aims to provide a comprehensive, flexible, well-documented and easy-to-use API to create publication-ready maps that can directly be used for interactive data analysis.</i>
+</td></tr>
+</table>
 
-## How does it work?
+## What can I do with EOmaps?
 
 **EOmaps** is built on top of [matplotlib](https://matplotlib.org/) and [cartopy](https://scitools.org.uk/cartopy/docs/latest/) and integrates well with the scientific python infrastructure (e.g., `numpy`, `pandas`, `geopandas`, `xarray` etc.), allowing you to visualize point-, raster- or vector-datasets provided in almost any format you can imagine, no matter if you're dealing with just a few unsorted datapoints or multi-dimensional stacks of global high-resolution datasets.  
 
-Figures created with EOmaps are multi-layered, so you can (transparently) overlay and interactively compare your datasets with ease. With the accompanying GUI widget, you can quickly switch layers, change the layout, examine the large collection of features and web-map services, and explore the capabilities of EOmaps. 
+Figures created with EOmaps are multi-layered, so you can (transparently) overlay and interactively compare your datasets with ease. With the accompanying GUI widget, you can quickly switch layers, change the layout, examine the large collection of features and web-map services, and explore the capabilities of EOmaps.
 Once you're map is ready, you can export it as high-resolution image or vector-graphic for further editing.
-Leveraging the powers of matplotlib, you can also embed interactive maps in Jupyter Notebooks, GUI frameworks like `Qt`, `tkinter`, `wx` or on a webpage (html).
+Leveraging the powers of matplotlib, you can also embed interactive maps in Jupyter Notebooks, on a webpage or in GUI frameworks like Qt, tkinter etc..
 
+> [!IMPORTANT]
+>
+> EOmaps is 100% free and open-source.  
+> As such, acknowledgement is extremely important to allow continued support and development of the package.
+>  
+> Did EOmaps help in your research? $\Rightarrow$ **Add a ✔️ Citation to your publication!**
+>
+> <a href="https://doi.org/10.5281/zenodo.6459598"><img src="https://zenodo.org/badge/410829039.svg" alt="https://zenodo.org/badge/latestdoi/410829039" align="left"></a>
+>
+> <details>
+> <summary>BibTeX</summary>
+> <br>
+> The following BibTeX entry uses a DOI that always points to the latest release of EOmaps!<br>
+> (You can get the DOI for a specific version form the <a href="https://doi.org/10.5281/zenodo.6459598">zenodo-page</a>)
+>
+> ```bibtex
+> @software{eomaps,
+>   author       = {Raphael Quast},
[…]
