### 2024-02-16 159b7c9
use parameter-name as colorbar label by default
