### 2024-02-16 #138 @raphaelquast
I finally found the time to draft an updated README to tackle the last open point from the review (PR: https://github.com/raphaelquast/EOmaps/pull/230).

If any of you find the time, feedback would be highly appreciated!
(You can look at it [here](https://github.com/raphaelquast/EOmaps/tree/update_readme))
