### 2024-02-17 a851526
fix jupyter notebook issues if blit-manager is not defined as property
