+++ docs/gen_autodoc_file.py
@@ -4,6 +4,14 @@ from itertools import chain
 
 from eomaps import Maps
 
+# TODO there must be a better way than this...
+# BM needs to be a property otherwise there are problems with jupyter notebooks
+# In order to make BM still accessible to sphinx, override it prior to generating
+# the autodoc-files
+from eomaps._blit_manager import BlitManager
+
+Maps.BM = BlitManager
+
 
 def get_autosummary(
     currentmodule="eomaps.eomaps",
