+++ docs/api_basics.rst
@@ -54,16 +54,19 @@ for a map.
   (e.g.: ``figsize``, ``frameon``, ``edgecolor`` etc).
 
 
-Possible ways for specifying the ``crs`` for plotting are:
+.. dropdown:: Possible ways for specifying the ``crs`` for plotting
+    :open:
+    :icon: info
+    :color: info
 
-- If you provide an integer, it is identified as an epsg-code (e.g. ``4326``, ``3035``, etc.)
+    - If you provide an integer, it is identified as an epsg-code (e.g. ``4326``, ``3035``, etc.)
 
-  - 4326 defaults to `PlateCarree` projection
+      - 4326 defaults to `PlateCarree` projection
 
-- All other CRS usable for plotting are accessible via ``Maps.CRS``, e.g.: ``crs=Maps.CRS.Orthographic()``, ``crs=Maps.CRS.Equi7_EU``...
+    - All other CRS usable for plotting are accessible via ``Maps.CRS``, e.g.: ``crs=Maps.CRS.Orthographic()``, ``crs=Maps.CRS.Equi7_EU``...
 
-  - ``Maps.CRS`` is just an accessor for ``cartopy.crs``
-  - For a full list of available projections see: `Cartopy projections <https://scitools.org.uk/cartopy/docs/v0.15/crs/projections.html>`_
+      - ``Maps.CRS`` is just an accessor for ``cartopy.crs``
+      - For a full list of available projections see: `Cartopy projections <https://scitools.org.uk/cartopy/docs/v0.15/crs/projections.html>`_
 
 .. autosummary::
     :nosignatures:
@@ -79,48 +82,70 @@ Possible ways for specifying the ``crs`` for plotting are:
 Layer management
 ~~~~~~~~~~~~~~~~
 
+A :py:class:`Maps` object represents one (or more) of the following things **on the assigned layer**:
+
+- a collection of features, callbacks,..
+- a single dataset  (and associated callbacks)
+
+
+You can create as many layers as you need! The following image explains how it works in general:
+
 .. image:: _static/intro.png
    :width: 70%
 
+.. dropdown:: Creating new layers
+    :icon: info
+    :color: info
 
-A :py:class:`Maps` object represents a collection of features, callbacks,.. **on the assigned layer**.
+    To create **a NEW layer**, use :py:meth:`m.new_layer("layer-name") <Maps.new_layer>`.
 
-Once you have created a map, you can create **additional** :py:class:`Maps` **objects for the same map** by using :py:meth:`Maps.new_layer`.
-
-🌱  If no explicit layer-name is provided, the returned :py:class:`Maps` object will use the same layer as the parent :py:class:`Maps` object.
-
-  - This is especially useful if you want to plot **multiple datasets on the same map and layer**.
-
-🌱  To create **a NEW layer** named ``"my_layer"``, use ``m2 = m.new_layer("my_layer")``
-
-  - Features, Colorbars etc. added to a :py:class:`Maps` object are only visible if the associated layer is visible.
-  - Callbacks are only executed if the associated layer is visible.
-  - See :ref:`combine_layers` on how to select the currently visible layer(s).
+    - Features, Colorbars etc. added to a :py:class:`Maps` object are only visible if the associated layer is visible.
+    - Callbacks are only executed if the associated layer is visible.
+    - See :ref:`combine_layers` on how to select the currently visible layer(s).
 
 
-.. code-block:: python
-    :name: test_layers_01
+    .. code-block:: python
+        :name: test_layers_create_new_layer
 
-    from eomaps import Maps
-    m = Maps()                           # same as `m = Maps(crs=4326, layer="base")`
-    m.add_feature.preset.coastline()     # add coastlines to the "base" layer
+        from eomaps import Maps
[…]
