### 2024-02-20 #138 @raphaelquast
@banesullivan @jhkennedy @yeelauren

Dear all, I wanted to ask if it is possible to get a quick update on the time-schedule for this review.

The reason I'm asking is because I am currently postponing the release for `EOmaps v8.0` in hope that it might be aligned with the outcome of this review. However, since time is flying, I am now facing several upcoming activities (most importantly [EGU General Assembly](https://www.egu24.eu/) in 14-19 April and [GeoPython 2024](https://2024.geopython.net/) conference on May 27) where I plan to participate with all the improvements introduced with `EOmaps v8.0`.

I would highly appreciate your opinion if it is feasible to arrive at a conclusion for this review by **end of March**.
If you think this is out of scope, I would release `EOmaps v8.0` in the upcoming weeks irrespective of the status of the PyOpenSci review (and we can continue here without a time-constraint).
