### 2024-02-20 #138 @jhkennedy
@banesullivan @raphaelquast Sorry I got slammed the last couple of weeks.

As far as I'm concerned, EOMaps is in fantastic shape, @raphaelquast did an amazing job responding to feedback and updating EOMaps, and I definitely would recommend accepting EOmaps v8.0.
