### 2024-02-27 7005836
fix identification of coordinate-names on read_file (and in widget)
