### 2024-02-28 9c35aa0
make sure to update after adding a text to a figure
