### 2024-02-28 a23eeda
fix issues with string-type dpi values passed to savefig (encountered with datashader in JupyterLab if show is called after plot)
