### 2024-02-28 4e90a41
explicitly include subpackes in pyproject.toml
