### 2024-03-03 raphaelquast/EOmaps#227 @yeelauren
Pasting the latest results from my test run in case it's helpful:
```
=========================================== short test summary info ============================================
FAILED tests/test_plot_shapes.py::TestPlotShapes::test_contour - AttributeError: 'NoneType' object has no attribute 'dpi'
=========================== 1 failed, 93 passed, 911 warnings in 2320.22s (0:38:40) ============================
```
