### 2024-03-03 raphaelquast/EOmaps#227 @raphaelquast
@yeelauren this seems to be a run from an outdated `dev` branch?
The most recent test-suite contains 100 tests and currently finishes with 22 warnings on python 3.11 (see [here](https://github.com/raphaelquast/EOmaps/actions/runs/8086470303/job/22096302599#step:4:266))
