+++ eomaps/widgets.py
@@ -1,0 +1,379 @@
+from abc import abstractmethod
+from functools import wraps
+
+import numpy as np
+import matplotlib.pyplot as plt
+
+from . import _log
+from ._blit_manager import LayerParser
+
+try:
+    import ipywidgets
+except ImportError:
+    _log.exception("EOmaps-widgets are missing the required dependency 'ipywidgets'!")
+
+
+def _check_backend():
+    backend = plt.get_backend()
+    if "ipympl" not in backend.lower():
+        raise AssertionError(
+            "EOmaps-widgets only work with the 'ipympl (widget)' backend! "
+            "Make sure you have 'ipympl' installed and use the magic-command "
+            "'%matplotlib widget' to switch to the interactive jupyter backend!"
+        )
+
+
+# %% Layer Selector Widgets
+
+
+class _LayerSelectionWidget:
+    def __init__(self, m, layers=None):
+        """
+        A widget to switch layers of a given Maps-object.
+
+        Parameters
+        ----------
+        m : eomaps.Maps
+            The Maps-object to use.
+        layers : list, optional
+            A list of layer-names to use.
+            If None, all available layers of the provided Maps-object are used.
+
+            The following options are possible:
+
+            A list of layer-names
+
+            >>> ["layer1", "layer2", ...]
+
+            A list of layer-names and/or layer-names with transparency-assignments
+
+            >>> ["layer1", "layer2", ("layer", 0.4)]
+
+            To provide explict display-names for a layer, pass a list of the form
+            `[display-name, <layer-assignment>]`
+
+            >>> [["My Layer", "layer1"],
+            >>>  ["My secondl lyer", "layer2"],
+            >>>  ["Multiple layers", ("layer1",
+            >>>                       ("layer", 0.4))]
+            >>> ]
+
+        """
+        _check_backend()
+
+        self._m = m
+        self._set_layers(layers)
+
+    def _set_layers(self, layers):
+        if layers is None:
+            self._layers = self._m._get_layers()
+        else:
+
+            self._layers = []
+            for l in layers:
+                if isinstance(l, str):
+                    self._layers.append((l, l))
+                elif isinstance(l, tuple):
+                    l = self._parse_layer(l)
+                    self._layers.append((l, l))
+                elif isinstance(l, list):
+                    self._layers.append((l[0], self._parse_layer(l[1])))
+
+    @staticmethod
+    def _parse_layer(l):
+        # check if a single transparent layer is provided
+        if isinstance(l, tuple):
+            if (
+                len(l) == 2
+                and isinstance(l[0], str)
+                and isinstance(l[1], (int, float, np.number))
+            ):
+                return LayerParser._get_combined_layer_name(l)
+            else:
+                return LayerParser._get_combined_layer_name(*l)
+        else:
+            return l
+
+    def handler(self, change):
+        try:
+            if self.value is not None:
+                self._m.show_layer(self.value)
+                self._m.BM.update()
+        except Exception:
+            _log.error("Problem in LayerSelectionWidget handler...", exc_info=True)
+
+
+class LayerDropdown(ipywidgets.Dropdown, _LayerSelectionWidget):
+    _description = "Layers"
+
+    @wraps(_LayerSelectionWidget.__init__)
+    def __init__(self, *args, description=None, **kwargs):
+
+        _LayerSelectionWidget.__init__(self, *args, **kwargs)
+
+        super().__init__(
+            options=self._layers,
+            description=self._description if description is None else description,
[…]
