### 2024-03-04 6dc34e4
don't call .show() automatically with ipympl backend
