### 2024-03-04 d52b078
add PeekLayerCheckbox widget
