### 2024-03-05 9c433c5
update widgets - improve inheritance concept - pass additional kwargs to ipython widgets - force full updates on layer-changes
