+++ eomaps/widgets.py
@@ -1,5 +1,6 @@
 from abc import abstractmethod
 from functools import wraps
+from contextlib import contextmanager
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -23,10 +24,24 @@ def _check_backend():
         )
 
 
+@contextmanager
+def _force_full(m):
+    """A contextmanager to force a full update of the figure (to avoid glitches)"""
+    force_full = getattr(m.BM, "_mpl_backend_force_full", False)
+
+    try:
+        m.BM._mpl_backend_force_full = True
+        yield
+    finally:
+        m.BM._mpl_backend_force_full = force_full
+
+
 # %% Layer Selector Widgets
 
 
 class _LayerSelectionWidget:
+    _description = "LayerSelectionWidget"
+
     def __init__(self, m, layers=None):
         """
         A widget to switch layers of a given Maps-object.
@@ -62,22 +77,28 @@ class _LayerSelectionWidget:
         _check_backend()
 
         self._m = m
-        self._set_layers(layers)
+        self._set_layers_options(layers)
 
-    def _set_layers(self, layers):
+    def _set_layers_options(self, layers):
+        # _layers is a list of the actual layer-names
+        # _options is a list of tuples (name, value) passed to the widget-init
+
         if layers is None:
             self._layers = self._m._get_layers()
+            self._options = [*self._layers]
         else:
-
-            self._layers = []
+            self._layers, self._options = [], []
             for l in layers:
                 if isinstance(l, str):
-                    self._layers.append((l, l))
+                    self._layers.append(l)
+                    self._options.append((l, l))
                 elif isinstance(l, tuple):
                     l = self._parse_layer(l)
-                    self._layers.append((l, l))
+                    self._layers.append(l)
+                    self._options.append((l, l))
                 elif isinstance(l, list):
-                    self._layers.append((l[0], self._parse_layer(l[1])))
+                    self._options.append((l[0], self._parse_layer(l[1])))
+                    self._layers.append(self._parse_layer(l[1]))
 
     @staticmethod
     def _parse_layer(l):
@@ -94,153 +115,94 @@ class _LayerSelectionWidget:
         else:
             return l
 
+
+class _SingleLayerSelectionWidget(_LayerSelectionWidget):
+    _description = "Layers"
+    _widget_cls = None
+
+    @wraps(_LayerSelectionWidget.__init__)
+    def __init__(self, m, layers=None, **kwargs):
+
+        _LayerSelectionWidget.__init__(self, m=m, layers=layers)
+
+        self._set_default_kwargs(kwargs)
+        self._widget_cls.__init__(self, options=self._options, **kwargs)
+
+        self.observe(self.handler)
+
+    def _set_default_kwargs(self, kwargs):
+        kwargs.setdefault("description", self._description)
+        if self._m.BM.bg_layer in self._layers:
+            kwargs.setdefault("value", self._m.BM.bg_layer)
+
     def handler(self, change):
         try:
             if self.value is not None:
-                self._m.show_layer(self.value)
-                self._m.BM.update()
+                with _force_full(self._m):
+                    self._m.show_layer(self.value)
+                    self._m.BM.update()
+
         except Exception:
             _log.error("Problem in LayerSelectionWidget handler...", exc_info=True)
 
 
-class LayerDropdown(ipywidgets.Dropdown, _LayerSelectionWidget):
-    _description = "Layers"
+class _MultiLayerSelectionWidget(_SingleLayerSelectionWidget):
+    def _set_default_kwargs(self, kwargs):
[…]
