### 2024-03-06 600aae8
add basic docstrings for widgets
