+++ eomaps/widgets.py
@@ -36,44 +36,71 @@ def _force_full(m):
         m.BM._mpl_backend_force_full = force_full
 
 
+from textwrap import dedent, indent
+
+
+def _add_docstring(prefix="", suffix="", replace_with=None):
+    def _add_docstring(cls):
+
+        if replace_with is None and cls.__doc__ is not None:
+            doc = f"{prefix}\n{dedent(cls.__doc__)}\n{suffix}"
+        elif replace_with is not None:
+            doc = f"{prefix}\n{dedent(replace_with.__doc__)}\n{suffix}"
+        else:
+            doc = f"{prefix}\n{dedent(suffix)}"
+
+        doc = indent(doc, "    ")
+        cls.__doc__ = doc
+        cls.__init__.__doc__ = doc
+
+        return cls
+
+    return _add_docstring
+
+
 # %% Layer Selector Widgets
 
 
 class _LayerSelectionWidget:
+    # A widget to switch layers of a given Maps-object.
+
+    """
+
+    For more information on how to customize the widgets, have a look at the
+    documentation for Jupyter Widgets (https://ipywidgets.readthedocs.io).
+
+    Parameters
+    ----------
+    m : eomaps.Maps
+        The Maps-object to use.
+    layers : list, optional
+        A list of layer-names to use.
+        If None, all available layers of the provided Maps-object are used.
+
+        The following options are possible:
+
+        A list of layer-names
+
+        >>> ["layer1", "layer2", ...]
+
+        A list of layer-names and/or layer-names with transparency-assignments
+
+        >>> ["layer1", "layer2", ("layer", 0.4)]
+
+        To provide explict display-names for a layer, pass a list of the form
+        `[display-name, <layer-assignment>]`
+
+        >>> [["My Layer", "layer1"],
+        >>>  ["My secondl lyer", "layer2"],
+        >>>  ["Multiple layers", ("layer1",
+        >>>                       ("layer", 0.4))]
+        >>> ]
+
+    """
+
     _description = "LayerSelectionWidget"
 
     def __init__(self, m, layers=None):
-        """
-        A widget to switch layers of a given Maps-object.
-
-        Parameters
-        ----------
-        m : eomaps.Maps
-            The Maps-object to use.
-        layers : list, optional
-            A list of layer-names to use.
-            If None, all available layers of the provided Maps-object are used.
-
-            The following options are possible:
-
-            A list of layer-names
-
-            >>> ["layer1", "layer2", ...]
-
-            A list of layer-names and/or layer-names with transparency-assignments
-
-            >>> ["layer1", "layer2", ("layer", 0.4)]
-
-            To provide explict display-names for a layer, pass a list of the form
-            `[display-name, <layer-assignment>]`
-
-            >>> [["My Layer", "layer1"],
-            >>>  ["My secondl lyer", "layer2"],
-            >>>  ["Multiple layers", ("layer1",
-            >>>                       ("layer", 0.4))]
-            >>> ]
-
-        """
         _check_backend()
 
         self._m = m
@@ -156,26 +183,62 @@ class _MultiLayerSelectionWidget(_SingleLayerSelectionWidget):
             kwargs.setdefault("value", (self._layers[0][1],))
 
 
+@_add_docstring(
+    "A Dropdown list to select the visible layer.", replace_with=_LayerSelectionWidget
+)
[…]
