### 2024-03-06 2f99964
add section on Jupyter Widgets to the docs
