### 2024-03-07 bcb6df3
add basic unittests for widgets
