### 2024-03-07 b180731
fix init of multiselector if unavailable layer is visible
