### 2024-03-07 b742ace
fix doc unittests (add option to ignore specific code-cells) - use the tag "ignore_in_unittest" to ignore a code-cell
