### 2024-03-07 f653e4d
update version to 8.0rc1
