### 2024-03-07 e14d993
update layout editor - allow "-1" as width/height - make sure x0 and y0 are fixed to the provided values - update docstrings accordingly
