### 2024-03-08 69e6527
synchronize layer-change widgets and avoid multiple handler calls
