### 2024-03-08 1a354c3
fix deactivating callback checkbox widgets
