### 2024-03-09 f0ab726
fix widget unittests
