### 2024-03-10 2a98e4c
improvements for data_identification + allow passing xarray.Dataset objects to m.set_data
