+++ eomaps/_data_manager.py
@@ -165,79 +165,150 @@ class DataManager:
         if self.on_fetch_bg in self.m.BM._before_update_actions:
             self.m.BM._before_update_actions.remove(self.on_fetch_bg)
 
-    def _identify_data(self, data=None, x=None, y=None, parameter=None):
-        # identify the way how the data has been provided and convert to the internal
-        # structure
+    def _identify_pandas(self, data=None, x=None, y=None, parameter=None):
+        (pd,) = register_modules("pandas", raise_exception=False)
 
-        if data is None:
-            data = self.m.data_specs.data
-        if x is None:
-            x = self.m.data_specs.x
-        if y is None:
-            y = self.m.data_specs.y
-        if parameter is None:
-            parameter = self.m.data_specs.parameter
+        if pd is None or not isinstance(data, pd.DataFrame):
+            return None
 
-        # check other types before pandas to avoid unnecessary import
-        if data is not None and not isinstance(data, (list, tuple, np.ndarray)):
-            (pd,) = register_modules("pandas", raise_exception=False)
+        if parameter is not None:
+            # get the data-values
+            z_data = data[parameter].values
+        else:
+            # use the first found numeric column of the DataFrame if possible.
+            numeric_columns = data.select_dtypes("number").columns
+            if len(numeric_columns) > 0:
+                z_data = data[numeric_columns[0]].values
+            else:
+                # TODO remove this and raise an error?
+                z_data = np.repeat(np.nan, len(data))
 
-            if pd is None:
+        # get the index-values
+        ids = data.index.values
+
+        if isinstance(x, str) and isinstance(y, str):
+            if (x in data) and (y in data):
+                # get the coordinates from columns
+                xorig = data[x].values
+                yorig = data[y].values
+            elif (x in data.index.names) and (y in data.index.names):
+                # get the coordinates from index-values
+                xorig = data.index.get_level_values(x)
+                yorig = data.index.get_level_values(y)
+            else:
                 raise TypeError(
-                    f"EOmaps: Unable to handle the input-data type: {type(data)}"
+                    "EOmaps: unable to identify coordinate values for " f"x={x}, y={y}"
                 )
+        elif (x is None) and (y is None):
+            if data.index.nlevels == 2:
+                # get the coordinates from index-values
+                xorig = data.index.get_level_values(0)
+                yorig = data.index.get_level_values(1)
+            else:
+                raise TypeError(
+                    "EOmaps: Either specify explicit column-names to use "
+                    "for `x` and `y` or pass a multi-index DataFrame "
+                    "with exactly 2 index-levels!"
+                )
+        else:
+            assert isinstance(x, (list, np.ndarray, pd.Series)), (
+                "'x' must be either a column-name, or explicit values "
+                " specified as a list, a numpy-array or a pandas.Series "
+                + f"if you provide the data as '{type(data)}'"
+            )
+            assert isinstance(y, (list, np.ndarray, pd.Series)), (
+                "'y' must be either a column-name, or explicit values "
+                " specified as a list, a numpy-array or a pandas.Series "
+                + f"object if you provide the data as '{type(data)}'"
+            )
 
-            if isinstance(data, pd.DataFrame):
-                if parameter is not None:
-                    # get the data-values
-                    z_data = data[parameter].values
-                else:
-                    z_data = np.repeat(np.nan, len(data))
+            xorig = np.asanyarray(x)
+            yorig = np.asanyarray(y)
 
-                # get the index-values
-                ids = data.index.values
+        return z_data, xorig, yorig, ids, parameter
 
-                if isinstance(x, str) and isinstance(y, str):
-                    # get the data-coordinates
[…]
