+++ eomaps/_containers.py
@@ -21,8 +21,8 @@ class DataSpecs(object):
         self,
         m,
         data=None,
-        x="lon",
-        y="lat",
+        x=None,
+        y=None,
         crs=4326,
         parameter=None,
         encoding=None,
