+++ docs/api_data_visualization.rst
@@ -82,6 +82,197 @@ A dataset is fully specified by setting the following properties:
 - ``cpos``, ``cpos_radius`` (optional): the pixel offset
 
 
+The following data-types are currently accepted as input:
+
+.. dropdown:: **pandas.DataFrames**
+    :icon: square-fill
+    :color: info
+
+    .. tab-set::
+
+        .. tab-item:: Assignments
+
+            - ``data``: pandas.DataFrame
+            - ``x``, ``y``: The column-names to use as coordinates (``string``)
+            - ``parameter``: The column-name to use as data-values (``string``)
+
+        .. tab-item:: basic DataFrame
+
+            .. code-block:: python
+                :name: test_assign_pandas_01
+
+                from eomaps import Maps
+                import pandas as pd
+
+                df = pd.DataFrame(dict(lon=[1,2,3], lat=[2,5,4], data=[12, 43, 2]))
+                m = Maps()
+                m.set_data(df, x="lon", y="lat", crs=4326, parameter="data")
+                m.plot_map()
+
+        .. tab-item:: multiindex DataFrame
+
+            .. code-block:: python
+                :name: test_assign_pandas_02
+
+                from eomaps import Maps
+                import pandas as pd
+
+                data = dict(param=[10, 29, 39])
+                index = pd.MultiIndex.from_arrays([[10,20,30], [10,20,30]], names=("lon", "lat"))
+                df = pd.DataFrame(data=data, index=index)
+
+                m = Maps()
+                m.set_data(df, x="lon", y="lat", crs=4326, parameter="param")
+                m.plot_map()
+
+
+
+.. dropdown:: **numpy.Array** | **pandas.Series** | **list**
+    :icon: square-fill
+    :color: info
+
+    .. tab-set::
+
+        .. tab-item:: Assignments
+
+            - ``data``, ``x``, ``y``: ``numpy.array``, ``pandas.Series`` or ``list``
+
+              - either data and coordinates have the same 1D/2D shape or
+              - 2D ``data=(m, n)`` and 1D coordinates ``x=(m,)``, ``y=(n,)``
+
+            - ``parameter``: (optional) parameter name (``string``)
+
+        .. tab-item:: 1D list
+
+            .. code-block:: python
+                :name: test_assign_list
+
+                from eomaps import Maps
+
+                x, y, data = [1,2,3], [2, 5, 4], [12, 43, 2]
+                m = Maps()
+                m.set_data(data, x=x, y=y, crs=4326, parameter="param_name")
+                m.plot_map()
+
+        .. tab-item:: 1D numpy
+
+            .. code-block:: python
+                :name: test_assign_numpy_1d
+
+                from eomaps import Maps
+                import numpy as np
+
+                x, y, data = np.array([1,2,3]), np.array([5, 7, 9]), np.array([1, 2, 3])
+                m = Maps()
+                m.set_data(data=data, x=x, y=y, crs=4326, parameter="param_name")
+                m.plot_map()
+
+
+        .. tab-item:: 2D numpy
+
+            .. code-block:: python
+                :name: test_assign_numpy_2d
+
+                from eomaps import Maps
+                import numpy as np
+
+                x, y = np.meshgrid(np.array([1,2,3]), np.array([5, 7, 9]))
+                data = np.random.randint(0, 10, x.shape)
+                m = Maps()
+                m.set_data(data=data, x=x, y=y, crs=4326, parameter="param_name")
+                m.plot_map()
+
+
+        .. tab-item:: 2D + 1D numpy
+
+            .. code-block:: python
+                :name: test_assign_numpy_1d2d
+
+                from eomaps import Maps
+                import numpy as np
+
[…]
