### 2024-03-10 raphaelquast/EOmaps#233 @raphaelquast
Update contour-plots to new syntax introduced in mpl v3.8
This needs to be addressed to ensure (future) compatibility with `matplotlib >= 3.10`

```
  /home/runner/work/EOmaps/EOmaps/eomaps/shapes.py:51: MatplotlibDeprecationWarning: The collections attribute was deprecated in Matplotlib 3.8 and will be removed two minor releases later.
    self.collections = self.contour_set.collections
```

- `ContourSet` is now a `Collection` by itself (see: https://discourse.matplotlib.org/t/collections-attribute-deprecation-in-version-3-8/24164/5)
   - the `_CollectionAccessor` class for plotting contours should no longer be needed.

- [ ] TODO: update syntax to remove `_CollectionAccessor`.
