### 2024-03-11 raphaelquast/EOmaps#227 @raphaelquast
OK, latest test-suite is now at 123 tests and down to 17 warnings of which I consider only 1 as critical.
Therefore I've decided to close this issue for now. Feel free to re-open if you think it's still relevant!

- dedicated issue for problematic warning is here: #233
  (to be done before `matplotlib 3.10` is released)
