### 2024-03-11 e359513
add ipywidgets and ipympl to top-level "environment.yml" (as they are used for development and required in the tests)
