### 2024-03-11 #138 @raphaelquast
@banesullivan Could you provide an update on how we proceed with this review?

Test-warnings (https://github.com/raphaelquast/EOmaps/issues/227) are now down to 17 of which I only 1 requires attention.
  - Dedicated issue here: https://github.com/raphaelquast/EOmaps/issues/233).
     However, since this is only critical for `matplotlib >= 3.10` and requires a few bigger changes in the way how contour-plots are implemented, I'll not include a fix in v8.0 already)

Aside of that, I believe all comments have been addressed as requested.

---

In addition, there are a few nice features that now made it into the v8.0 branch!
- EOmaps now provides a set of pre-configured [**Jupyter Widgets**](https://eomaps.readthedocs.io/en/dev/notebooks/widgets.html) to get nice control-elements in Jupyter Notebooks
- `xarray.Dataset` is now accepted as input for `m.set_data()` (see info [here](https://eomaps.readthedocs.io/en/dev/api_data_visualization.html#assign-the-data))

> The latest release-candidate can be installed with `pip install eomaps[all]==8.0rc2`
