### 2024-03-11 2793b98
make sure encoding fill-value mask is always applied if provided
