### 2024-03-15 #138 @banesullivan
@raphaelquast, thank your for your patience! The last few weeks got the best of me between moving and some sickness.

> Aside of that, I believe all comments have been addressed as requested.

I believe you're right! I think we're ready to proceed with the acceptance of EOmaps given @jhkennedy and @yeelauren have both recommended accepting! 🎉 🎉

I'm going to change the status of this review and will handle the next steps from here.

----

I want to close the loop on an earlier conversation about submitting to JOSS: We highly recommend you submit EOmaps to JOSS, but are going to encourage you to go through the standard JOSS submission process as opposed to starting it here. Given acceptance here, I believe there is precedent for fast-tracking the JOSS submission or using this review to augment the JOSS software review. I will happily chime in to help in any way I can during the JOSS submission if you chose to do so, please `@` mention me as needed.

------

Amazing work here, @raphaelquast!!
