### 2024-03-15 #138 @raphaelquast
@banesullivan Thank you very much for all your time and efforts!
I fully understand that sometimes things get in the way... I hope you're feeling well again and that you had a successful move!

@yeelauren @jhkennedy The same also counts for you of course! Your reviews have been highly valuable and helped me to improve many parts of EOmaps!

---

🚀 This is amazing news! I'm thrilled that this review has reached a positive outcome and I'm very happy that EOmaps will become a part of PyOpenSci!

> I'd like to tie this review with version 8.0, would you let me know when that has been fully released?

[EOmaps v8.0](https://github.com/raphaelquast/EOmaps/pull/205) is already waiting in the starting blocks! I'll do a final checkup and try to release it as soon as possible!

> I'm going to change the status of this review and will handle the next steps from here.

@banesullivan Please keep me informed on any actions required from my side!

---

>I want to close the loop on an earlier conversation about submitting to JOSS: We highly recommend you submit EOmaps to JOSS, but are going to encourage you to go through the standard JOSS submission process as opposed to starting it here. Given acceptance here, I believe there is precedent for fast-tracking the JOSS submission or using this review to augment the JOSS software review. I will happily chime in to help in any way I can during the JOSS submission if you chose to do so, please @ mention me as needed.

Thank you for following up on this. I agree that this is the best way forward. I'll try to compose a JOSS submission later this year once all the conference contributions etc. are done. I highly appreciate your support on this and I will reach out to you when the time comes!
