### 2024-03-15 a164125
use numpy implementation for 1D/2D conversion with raster shape
