+++ docs/FAQ.rst
@@ -1,3 +1,5 @@
+.. _faq:
+
 👀 FAQ
 =======
 
