### 2024-03-15 97a1155
fix missing companion-widget source files in tarball
