### 2024-03-15 7eef546
Fix issues with pip install and missing "dask-expr" package https://github.com/dask/dask/issues/10995
