### 2024-03-17 #138 @raphaelquast
> I'd like to tie this review with version 8.0, would you let me know when that has been fully released?

@banesullivan EOmaps v8 is now released!
