### 2024-03-18 #138 @raphaelquast
PyOpenSci review Badge has been added to the [README](https://github.com/raphaelquast/EOmaps)
